% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_ArcsineSymmetric.R
\name{cf_ArcsineSymmetric}
\alias{cf_ArcsineSymmetric}
\title{Characteristic function of a linear combination
of independent zero-mean symmetric ARCSINE random variables}
\usage{
cf_ArcsineSymmetric(t, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{coef}{vector of the coefficients of the linear combination of the
symmetric Arcsinne distributed random variables. If \code{coef} is scalar, it is
assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)} is independently
and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent zero-mean symmetric ARCSINE random variables.
}
\description{
\code{cfS_ArcsineSymmetric(t, coef, niid)} evaluates Characteristic function of a linear combination
(resp. convolution) of independent zero-mean symmetric ARCSINE random variables defined on the interval \eqn{(-1,1)}.

That is, \code{cfS_ArcsineSymmetric} evaluates the characteristic function
\eqn{cf(t) of  Y = sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ ArcsineSymmetric}
are independent RVs defined on \eqn{(-1,1)}, for all \eqn{i = 1,...,N}.

The characteristic function of \eqn{X ~ ArcsineSymmetric} is defined by
\deqn{cf(t) = cf_ArcsineSymmetric(t) = besselj(0,t).}
}
\examples{
## EXAMPLE 1
# CF of the symmetric Arcsine distribution on (-1,1)
t <- seq(from = -50,
         to = 50,
         length.out = 501)
plotReIm(function(t)
        cf_ArcsineSymmetric(t),
        t,
        title = "CF of the the Arcsine distribution on (-1,1)")

## EXAMPLE 2
# CF of a linear combination of independent Arcsine RVs
t <- seq(from = -1,
         to = 1,
         length.out = 501)
coef <- c(1, 2, 3, 4, 5)
plotReIm(function(t)
        cf_ArcsineSymmetric(t, coef),
        t,
        title = "CF of a linear combination of independent Arcsine RVs")

## EXAMPLE 3
## PDF/CDF of a linear combination of independent Arcsine RVs
coef <- c(1, 2, 3, 4, 5)
cf   <- function(t)
        cf_ArcsineSymmetric(t, coef)
x    <- seq(from = -20,
            to = 20,
            length.out = 201)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$N <- 2 ^ 12
result <- cf2DistGP(cf, x, prob, options)
}
\references{
WITKOVSKY V. (2016). Numerical inversion of a characteristic
function: An alternative tool to form the probability distribution
of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Arcsine_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PDF}}, \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}}, \code{\link{cf_BetaNC}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}

Other Symmetric Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_TrapezoidalSymmetric}}
}
