% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_Beta.R
\name{cf_Beta}
\alias{cf_Beta}
\title{Characteristic function of a linear combination of independent BETA random variables}
\usage{
cf_Beta(t, alpha, beta, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{alpha}{vector of the 'shape' parameters \code{alpha > 0}. If empty, default value is \code{alpha = 1}.}

\item{beta}{vector of the 'shape' parameters \code{beta > 0}. If empty, default value is \code{beta = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the Beta distributed random variables.
If coef is scalar, it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of independent BETA random variables.
}
\description{
\code{cf_Beta(t, alpha, beta, coef, niid)} evaluates the characteristic function of a linear combination
(resp. convolution) of independent BETA random variables defined on the interval \eqn{(0,1)}.

That is, \code{cf_Beta} evaluates the characteristic function \eqn{cf(t)}
of  \eqn{Y = sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ Beta(\alpha_i,\beta_i)}
are independent RVs, with the shape parameters \eqn{\alpha_i > 0} and \eqn{\beta_i >0},
and with the \eqn{mean = \alpha_i / (\alpha_i + \beta)_i} and the
\eqn{variance = (\alpha_i*\beta_i) / ((\alpha_i+\beta_i)^2*(\alpha_i+\beta_i+1))}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{X ~ Beta(\alpha,\beta)} is
\deqn{cf(t) = cf_Beta(t,\alpha,\beta) = 1F1(\alpha; \alpha + \beta; i*t),}
where \eqn{1F1(.;.;.)} is the Confluent hypergeometric function. Hence,
the characteristic function of \eqn{Y  = coef(1)*X_1 + ... + coef(N)*X_N}
is \eqn{cf(t) =  cf_X_1(coef(1)*t) * ... * cf_X_N(coef(N)*t)},
where \eqn{X_i ~ Beta(\alpha(i),\beta(i))} with \eqn{cf_X_i(t)}.
}
\examples{
## EXAMPLE 1
# CF of a Beta RV
alpha <- 1
beta  <- 3
t <- seq(from = -50,
         to = 50,
         length.out = 501)
plotReIm(function(t)
        cf_Beta(t, alpha, beta), t, title = "CF of a Beta RVs")

## EXAMPLE 2
# PDF/CDF of a Beta RV
alpha <- 1
beta  <- 3
cf <- function(t) {
        cf_Beta(t, alpha, beta)
}
options <- list()
options$N <- 2 ^ 12
options$xMin <- 0
options$xMax <- 1
x <- seq(from = 0,
         to = 1,
         length.out = 201)
prob <- c(0.9, 0.95, 0.99)
result <- cf2DistGP(cf, x, prob, options)

## EXAMPLE 3
# CF of a linear combination of independent Beta RVs
alpha <- 1
beta  <- 3
coef  <- 1 / (((1:50) - 0.5) * pi) ^ 2
weights <- coef / sum(coef)
t <- seq(from = -100,
         to = 100,
         length.out = 501)
plotReIm(function(t)
        cf_Beta(t, alpha, beta, weights),
        t,
        title = "CF of a weighted linear combination of independent Beta RVs")

## EXAMPLE 4
# PDF/CDF of a weighted linear combination of independent Beta RVs
alpha <- 1
beta  <- 3
coef  <- 1 / (((1:50) - 0.5) * pi) ^ 2
weights <- coef / sum(coef)
cf <- function(t) {
        cf_Beta(t, alpha, beta, weights)
}
options <- list()
options$xMin <- 0
options$xMax <- 1
x <- seq(from = 0,
         to = 1,
         length.out = 201)
prob <- c(0.9, 0.95, 0.99)
result <- cf2DistGP(cf, x, prob, options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Beta_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PDF}}, \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
