% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_BetaNC.R
\name{cf_BetaNC}
\alias{cf_BetaNC}
\title{Characteristic function of a linear combination
of independent non-central BETA random variables}
\usage{
cf_BetaNC(t, alpha, beta, delta, coef, niid, tol)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{alpha}{vector of the 'shape' parameters \code{alpha > 0}. If empty, default value is \code{alpha = 1}.}

\item{beta}{vector of the 'shape' parameters \code{beta > 0}. If empty, default value is \code{beta = 1}.}

\item{delta}{vector of the non-centrality parameters \code{delta > 0}.
If empty, default value is \code{delta = 0}.}

\item{coef}{vector of the coefficients of the linear combination of the Beta distributed random variables.
If coef is scalar, it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}

\item{tol}{tolerance factor for selecting the Poisson weights, i.e. such that \eqn{PoissProb > tol}.
If empty, default value is \code{tol = 1e-12}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent non-central BETA random variables.
}
\description{
\code{cf_BetaNC(t, alpha, beta, delta, coef, niid, tol)} evaluates characteristic function
of a linear combination (resp. convolution) of independent non-central BETA random variables,
with distributions \eqn{Beta(\alpha_i,\beta_i,\delta_i)} defined on the interval \eqn{(0,1)}.

That is, \code{cf_BetaNC} evaluates the characteristic function \eqn{cf(t)}
of  \eqn{Y = sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ Beta(\alpha_i,\beta_i,\delta_i)}
are independent RVs, with the shape parameters \eqn{\alpha_i > 0} and \eqn{\beta_i > 0},
and the noncentrality parameters \eqn{\delta_i > 0}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{X ~ Beta(\alpha,\beta,\delta)} is Poisson mixture
of the central Beta CFs of the form \deqn{cf(t) = cf_BetaNC(t,\alpha,\beta,\delta) = exp(-\delta/2) sum_{j=1}^Inf (\delta/2)^j/j! * cf_beta(\alpha+j,\beta)}
where \eqn{cf_beta(\alpha+j,\beta)} are the CFs of central Beta RVs with parameters \eqn{\alpha+j} and \eqn{\beta}.
Hence,the characteristic function of \eqn{Y = coef(1)*X_1 + ... + coef(N)*X_N} is
\deqn{cf(t) =  cf_X_1(coef(1)*t) * ... * cf_X_N(coef(N)*t),}
where \eqn{X_i ~ BetaNC(\alpha(i),\beta(i),\delta(i))} with \eqn{cf_X_i(t)}.
}
\examples{
## EXAMPLE 1
# CF of non-central Beta RV with delta = 1
alpha <- 1
beta  <- 3
delta <- 1
t <- seq(from = -50,
         to = 50,
         length.out = 201)
plotGraf(function(t)
        cf_BetaNC(t, alpha, beta, delta), t, title = 'CF of a Beta RVs')

## EXAMPLE 2
# CDF/PDF of non-central Beta RV with delta = 1
alpha <- 1
beta  <- 3
delta <- 1
cf <- function(t)
        cf_BetaNC(t, alpha, beta, delta)
options <- list()
options$xMin <- 0
options$xMax <- 1
result <- cf2DistGP(cf = cf, options = options)

## EXAMPLE 3
# CDF/PDF of the linear combination of non-central Beta RVs
alpha <- c(5, 4, 3)
beta  <- c(3, 4, 5)
delta <- c(0, 1, 2)
coef  <- 1 / 3
cf <- function(t)
        cf_BetaNC(t, alpha, beta, delta, coef)
options <- list()
options$xMin <- 0
result <- cf2DistGP(cf = cf, options = options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Noncentral_beta_distribution}.

Other Continuous Probability Distribution: \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}}
}
