% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_BetaSymmetric.R
\name{cf_BetaSymmetric}
\alias{cf_BetaSymmetric}
\title{Characteristic function of a linear combination
of independent zero-mean symmetric BETA random variables}
\usage{
cf_BetaSymmetric(t, theta, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{theta}{the 'shape' parameter \code{theta > 0}. If empty, default value is \code{theta = 1}.}

\item{coef}{vector of the coefficients of the linear combination of Beta distributed random variables. If \code{coef} is scalar, it is
assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)} is independently
and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent zero-mean symmetric BETA random variables.
}
\description{
\code{cf_BetaSymmetric(t, theta, coef, niid)} evaluates the characteristic function of a linear combination
(resp. convolution) of independent zero-mean symmetric BETA random variables defined on the interval  \eqn{(-1,1)}.

 That is, \code{cf_BetaSymmetric} evaluates the characteristic function
 \eqn{cf(t)} of  \eqn{Y = sum_{i=1}^N coef_i * X_i}, where \eqn{X_i~BetaSymmetric(\theta_i)}
 are independent RVs defined on \eqn{(-1,1)}, for all \eqn{i = 1,...,N}.

The characteristic function of \eqn{X ~ BetaSymmetric(\theta)} is defined by
\deqn{cf(t) = cf_BetaSymmetric(t,\theta) = gamma(1/2+\theta) * (t/2)^(1/2-\theta) * besselj(\theta-1/2,t).}

SPECIAL CASES: \cr
1) \eqn{\theta = 1/2}; Arcsine distribution on \eqn{(-1,1)}:     \eqn{cf(t) = besselj(0,t)}, \cr
2) \eqn{\theta = 1};   Rectangular distribution on \eqn{(-1,1)}: \eqn{cf(t) = sin(t)/t}.
}
\examples{
## EXAMPLE 1
# CF of the symmetric Beta distribution with theta = 3/2 on (-1,1)
theta <- 3/2
t <- seq(from = -50,to = 50,length.out = 201)
plotGraf(function(t)
  cf_BetaSymmetric(t, theta), t, title = "CF of the symmetric Beta distribution on (-1,1)")

## EXAMPLE 2
# CF of a linear combination of independent Beta RVs
t <- seq(from = -20,to = 20,length.out = 201)
theta <- c(3, 3, 4, 4, 5)/2
coef <- c(1, 2, 3, 4, 5)/15
plotGraf(function(t)
  cf_BetaSymmetric(t,theta,coef), t, title = "CF of a linear combination of independent Beta RVs")

## EXAMPLE 3
# PDF/CDF of a weighted linear combination of independent Beta RVs
theta <- c(3, 3, 4, 4, 5)/2
coef <- c(1, 2, 3, 4, 5)/15
cf   <- function(t) cf_BetaSymmetric(t,theta,coef)
x <- seq(from = -1,to = 1,length.out = 201)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$N <- 2^12
options$xMin <- -1
options$xMax <- 1
result <- cf2DistGP(cf,x,prob,options)
}
\references{
WITKOVSKY V. (2016). Numerical inversion of a characteristic
function: An alternative tool to form the probability distribution
of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Beta_distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_Gamma}}, \code{\link{cf_InverseGamma}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}
}
