% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_ChiSquare.R
\name{cf_ChiSquare}
\alias{cf_ChiSquare}
\title{Characteristic function  of a linear combination of independent CHI-SQUARE random variables}
\usage{
cf_ChiSquare(t, df, ncp, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{df}{the degrees of freedom parameter \code{df > 0}. If empty, default value is \code{df = 1}.}

\item{ncp}{the non-centrality parameter \code{ncp > 0}. If empty, default value is \code{ncp = 0}.}

\item{coef}{vector of the coefficients of the linear combination
of the chi-squared random variables. If coef is scalar,
it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of independent CHI-SQUARE random variables.
}
\description{
\code{cf_ChiSquare(t, df, ncp, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
 of a linear combination (resp.convolution)
 of independent (possibly non-central) CHI-SQUARE random variables.

That is, \code{cf_ChiSquare} evaluates the characteristic function \eqn{cf(t)}
of  \eqn{Y = sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ ChiSquare(df_i,ncp_i)}
are inedependent RVs, with \eqn{df_i > 0} degrees of freedom the 'non-centrality'
parameters \eqn{ncp_i > 0}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{Y} is defined by
\deqn{cf(t) = Prod ( (1-2*i*t*coef(i))^(-df(i)/2) * exp((i*t*ncp(i))/(1-2*i*t*coef(i))) ).}
}
\note{
Ver.: 16-Sep-2018 18:16:32 (consistent with Matlab CharFunTool v1.3.0, 10-May-2017 18:11:50).
}
\examples{
## EXAMPLE 1
# CF of the distribution of ChiSquare RV with DF = 1, NCP = 1
df <- 1
ncp <- 1
coef <- 1
t <- seq(from = -100,
         to = 100,
         length.out = 501)
plotReIm(
        function(t)
                cf_ChiSquare(t, df, ncp, coef),
        t,
        title = expression('Characteristic function of the' ~ chi ^ 2 ~ 'RV with DF=1 and NCP=1')
)

## EXAMPLE 2
# CF of a linear combination of K=100 independent ChiSquare RVs
coef <- 1 / (((1:50) - 0.5) * pi) ^ 2
plot(
        1:50,
        coef,
        xlab = "",
        ylab = "",
        type = "p",
        pch = 20,
        col = "blue",
        cex = 1,
        main = expression('Coefficients of the linear combination of' ~ chi ^ 2 ~ 'RVs with DF=1')
)
lines(1:50, coef, col = "blue")
df <- 1
t <- seq(from = -100,
         to = 100,
         length.out = 501)
plotReIm(
        function(t)
                cf_ChiSquare(t = t, df = df, coef = coef),
        t,
        title = expression('CF of the linear combination of' ~ chi ^ 2 ~ 'RVs with DF=1')
)

## EXAMPLE 3
# PDF/CDF from the CF by cf2DistGP
df <- 1
coef <-  1 / (((1:50) - 0.5) * pi) ^ 2
cf <- function(t)
        cf_ChiSquare(t = t, df = df, coef = coef)
options <- list()
options$N <- 2 ^ 12
options$xMin <- 0
x <- seq(from = 0,
         to = 3.5,
         length.out = 501)
prob <- c(0.9, 0.95, 0.975, 0.99)
result <- cf2DistGP(cf, x, prob, options)
result
}
\references{
IMHOF J. (1961): Computing the distribution of quadratic forms in normal variables. Biometrika 48, 419-426.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Chi-squared_distribution},\cr
\url{https://en.wikipedia.org/wiki/Noncentral_chi-squared_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}},
  \code{\link{cfX_FisherSnedecor}},
  \code{\link{cfX_Gamma}}, \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_LogRV_MeansRatioW}},
  \code{\link{cf_LogRV_MeansRatio}},
  \code{\link{cf_LogRV_WilksLambdaNC}},
  \code{\link{cf_LogRV_WilksLambda}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
