% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_Conv.R
\name{cf_Conv}
\alias{cf_Conv}
\title{Characteristic function of a linear combination (convolution) of iid random variables X}
\usage{
cf_Conv(t, cfX, coef, n)
}
\arguments{
\item{t}{vector (or array) of input values \code{t} where the \code{cf_Conv} is evaluated.}

\item{cfX}{function handle to the given chracteristic function \eqn{cfX(t)}.}

\item{coef}{vector of coeficients of the linear combination of the iid random variables,
such that \eqn{Y = sum(coef(k) * X_k), cf = Prod_{k=1}^N cfX(coef(k)*t)}.}

\item{n}{optional power coeficient of additional convolution of the combiend CF of \eqn{Y}.
With using \eqn{n} (if empty, default value is \code{n = 1}), \eqn{cf = (Prod_{k=1}^N cfX(coef(k)*t)^n}.}
}
\value{
The characteristic function of a linear combination of iid RVs with characteristic function \eqn{cf},
evaluated at \eqn{t}.
}
\description{
\code{cf_Conv(t, cfX, coef, n)} computes the characteristic function of a linear combination
(convolution) of iid random variables \eqn{X} with given (common) characteristic function \eqn{cfX(t)}
and the coefficients coef, and such that \eqn{Y = coef(1)*X + ... + coef(N)*X},
i.e. \eqn{cf = cfX(coef(1)*t) * ... * cfX(coef(N)*t)}.

If or all \code{coef == 1}, then \eqn{cf = cfX(t)^N}, . Moreover, with using the optional parameter \code{n},
the algorithm evaluates CF of the convolution of \eqn{n} independent copies of the random variable \eqn{Y},
i.e. \eqn{Z = Y + ... + Y}, where \eqn{Y = coef(1)*X + ... + coef(N)*X},
i.e. \eqn{cf = (cfX(coef(1)*t) * ... * cfX(coef(N)*t))^n}.
}
\note{
Ver.: 06-Oct-2018 18:14:22 (consistent with Matlab CharFunTool v1.3.0, 9-May-2017 10:22:48).
}
\examples{
## EXAMPLE 1
# CF of a linear combination of chi-square random variables
df <- 1
cfX <- function(t) cfX_ChiSquare(t, df)
coef <- 1 / (1:100)
cf <- function(t) cf_Conv(t, cfX, coef)
t <- seq(-10, 10, length.out = 501)
plotReIm(cf, t, title = 'CF of a Linear Combination of iid Chi-Square RVs')
options <- list()
options$xMin <- 0
result <- cf2DistGP(cf, options = options)

## EXAMPLE 2
# CF of a linear combination of iid RVs sampled from the empirical
# distribution function (atrtificially generated data)
n <- 30
p <- c(0.2, 0.7, 0.1)
data <- p[1] * rchisq(n, 5) + p[2] * rnorm(n, 10, 1) + p[3] * rt(n, 1)
cfE <- function(t) cfE_Empirical(t, data)
coef <- 1 / c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
cf <- function(t) cf_Conv(t, cfE, coef)
t <- seq(-10, 10, length.out = 501)
plotReIm(cf, t, title = 'CF of a Linear Combination of iid RVs')
result <- cf2DistGP(cf)
}
\seealso{
Other CF Tool: \code{\link{cfX_PDF}}
}
