% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_Exponential.R
\name{cf_Exponential}
\alias{cf_Exponential}
\title{Characteristic function of a linear combination of independent EXPONENTIAL random variables}
\usage{
cf_Exponential(t, lambda, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{lambda}{vector of the 'rate' parameters \code{lambda > 0}. If empty, default value is \code{lambda = 1}.}

\item{coef}{vector of the coefficients of the linear combination
of the GAMMA random variables. If \code{coef} is scalar,
it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y +...+ Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of independent EXPONENTIAL random variables.
}
\description{
\code{cf_Exponential(t, lambda)} evaluates characteristic function of a linear combination
(resp. convolution) of independent EXPONENTIAL random variables.

That is, \code{cf_Exponential} evaluates the characteristic function \eqn{cf(t)}
of \eqn{Y = sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ EXP(\lambda_i)} are inedependent RVs,
with the rate parameters \eqn{\lambda_i > 0}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{Y} is defined by
\deqn{cf(t) = Prod( \lambda_i / (\lambda_i - 1i*t) ).}
}
\examples{
## EXAMPLE 1
# CF of the Exponential distribution with lambda = 5
lambda <- 5
t <- seq(from = -50,to = 50,length.out = 501)
plotGraf(function(t)
  cfX_Exponential(t, lambda), t, title = expression('CF of the Exponential distribution with' ~ lambda ~ '= 5'))

## EXAMPLE 2
# CF of a linear combination of independent Exponential RVs
coef <- 1/(((1:50) - 0.5)*pi)^2
lambda <- 5
t <- seq(from = -100,to = 100,length.out = 201)
plotGraf(function(t)
  cfX_Exponential(t, lambda, coef), t, title = "CF of a linear combination of EXPONENTIAL RVs")

## EXAMPLE 3
# PDF/CDF of the compound Binomial-Exponential distribution
n <- 25
p <- 0.3
coef <- 1/(((1:50) - 0.5)*pi)^2
lambda <- 5
cfX <- function(t) cf_Exponential(t,lambda,coef)
cf <- function(t) cfN_Binomial(t,n,p,cfX)
x <- seq(from = 0,to = 5,length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$isCompound <- TRUE
options$N <- 2^12
options$SixSigmaRule <- 15
result <- cf2DistGP(cf,x,prob,options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Exponential_distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}}, \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}
}
