% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_Gamma.R
\name{cf_Gamma}
\alias{cf_Gamma}
\title{Characteristic function of a linear combination of independent GAMMA random variables}
\usage{
cf_Gamma(t, alpha, beta, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{alpha}{the shape parameter \code{alpha > 0}. If empty, default value is \code{alpha = 1}.}

\item{beta}{the rate (\eqn{1/scale}) parameter \code{beta > 0}. If empty, default value is \code{beta = 1}.}

\item{coef}{- vector of the coefficients of the linear combination
of the GAMMA random variables. If coef is scalar,
it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
 is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of independent GAMMA random variables.
}
\description{
\code{cf_Gamma(t, alpha, beta, coef, niid)} evaluates the characteristic function
of a linear combination (resp. convolution) of independent GAMMA random variables.

That is, \code{cf_Gamma} evaluates the characteristic function \eqn{cf(t)}
of  \eqn{Y =sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ Gamma(\alpha_i,\beta_i)}
are inedependent RVs, with the shape parameters \eqn{\alpha_i > 0}
and the rate parameters \eqn{\beta_i > 0}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{Y} is defined by
\deqn{cf(t) = Prod( (1 - i*t*coef(i)/\beta(i))^(-\alpha(i)) ).}
}
\details{
PARAMETRIZATION:\cr
Notice that there are three different parametrizations for GAMMA distribution in common use:\cr
i)   With a shape parameter \eqn{k} and a scale parameter \eqn{theta}.\cr
ii)  With a shape parameter \eqn{\alpha = k} and an inverse scale parameter \eqn{\beta = 1/\theta},
called a rate parameter.\cr
iii) With a shape parameter \eqn{k} and a mean parameter \eqn{\mu = k/\beta}.
In each of these three forms, both parameters are positive real numbers.\cr

Here, \code{cf_Gamma} implements the shape-rate parametrization with parameters \code{alpha}
and \code{beta}, respectively.

SPECIAL CASES:\cr
1) If \eqn{X ~ Gamma(1,\lambda)} (shape-rate parametrization), then \eqn{X} has an exponential distribution
with rate parameter lambda.
2) If \eqn{X ~ Gamma(df/2,1/2)}(shape-rate parametrization), then \eqn{X ~ ChiSquared(df)},
the chi-squared distribution with df degrees offreedom. Conversely,
if \eqn{Q ~ ChiSquared(df)} and c is a positive constant, then \eqn{cQ ~ Gamma(df/2,1/2c)}.
3) If \eqn{X ~ Gamma(\alpha,\theta)} and \eqn{Y ~ Gamma(\beta,\theta)} are independently distributed,
then \eqn{X/(X + Y)} has a beta distribution with parameters \eqn{\alpha} and \eqn{\beta}.
}
\examples{
## EXAMPLE 1
# CF of a linear combination of independent Gamma RVs
coef <- 1 / (((1:50) - 0.5) * pi) ^ 2
plot(
        1:50,
        coef,
        xlab = "",
        ylab = "",
        type = "p",
        pch = 20,
        col = "blue",
        cex = 1,
        main = expression('Coefficients of the linear combination of' ~ chi ^ 2 ~ 'RVs with DF=1')
)
lines(1:50, coef, col = "blue")
alpha <- 5 / 2
beta <- 1 / 2
t <- seq(from = -10,
         to = 10,
         length.out = 201)
plotGraf(function(t)
        cf_Gamma(t, alpha, beta, coef),
        t,
        title = "Characteristic function of the linear combination of GAMMA RVs")

## EXAMPLE 2
# PDF/CDF from the CF by cf2DistGP
alpha <- 5 / 2
beta <- 1 / 2
coef <- 1 / (((1:50) - 0.5) * pi) ^ 2
cf <- function(t)
        cf_Gamma(t, alpha, beta, coef)
options <- list()
options$N <- 2 ^ 10
options$xMin <- 0
result <- cf2DistGP(cf = cf, options = options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Gamma_distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_InverseGamma}}, \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}
}
