% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_InverseGamma.R
\name{cf_InverseGamma}
\alias{cf_InverseGamma}
\title{Characteristic function of a linear combination of independent INVERSE-GAMMA random variables}
\usage{
cf_InverseGamma(t, alpha, beta, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{alpha}{the shape parameter \code{alpha > 0}. If empty, default value is \code{alpha = 1}.}

\item{beta}{the rate (1/scale) parameter \code{beta > 0}. If empty, default value is \code{beta = 1}.}

\item{coef}{- vector of the coefficients of the linear combination
of the IGamma random variables. If coef is scalar,
it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y +...+ Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef_i * X_i}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of independent INVERSE-GAMMA random variables.
}
\description{
\code{cf_InverseGamma(t, alpha, beta, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
of a linear combination (resp. convolution) of independent INVERSE-GAMMA random variables.

That is, \code{cf_InvGamma} evaluates the characteristic function \eqn{cf(t)}
of  \eqn{Y =  sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ InvGamma(\alpha_i,\beta_i)}
are inedependent RVs, with the shape parameters \eqn{\alpha_i > 0} and
the  rate parameters \eqn{\beta_i > 0}, for \eqn{i = 1,...,N}.

The characteristic function of Y is defined by
\deqn{cf(t) = Prod( 2 / gamma(\alpha(i)) * (-1i*\beta(i)*t).^(\alpha(i)/2) * besselk(\alpha(i),sqrt(-4i*\beta(i)*t)) ).}
}
\details{
PARAMETRIZATION:\cr
As for the GAMMA distribution, also for the Inverse-Gamma distribution
there are three different possible parametrizations:\cr
i)   With a shape parameter \eqn{k} and a scale parameter \eqn{\theta}.\cr
ii)  With a shape parameter \eqn{\alpha = k} and an inverse scale parameter
\eqn{\beta = 1/\theta}, called a rate parameter.\cr
iii) With a shape parameter \eqn{k} and a mean parameter \eqn{\mu = k/\beta}.
In each of these three forms, both parameters are positive real numbers.

Here, \code{cf_InverseGamma} implements the shape-rate parametrization with
parameters \eqn{\alpha} and \eqn{\beta}, respectively.

If \eqn{X ~ IGamma(df/2,1/2)}(shape-rate parametrization), then \eqn{X} is identical
to \eqn{ICHI2(df)}, the Inverse-Chi-squared distribution with \eqn{df} degrees of freedom.
}
\note{
Ver.: 16-Sep-2018 18:28:11 (consistent with Matlab CharFunTool v1.3.0, 10-May-2017 18:11:50).
}
\examples{
## EXAMPLE 1
# CF of a linear combination of K=100 independent IGamma RVs
coef <- 1 / (((1:50) - 0.5) * pi) ^ 2
plot(
        1:50,
        coef,
        xlab = "",
        ylab = "",
        type = "p",
        pch = 20,
        col = "blue",
        cex = 1,
        main = expression('Coefficients of the linear combination of' ~ chi ^ 2 ~ 'RVs with DF=1')
)
lines(1:50, coef, col = "blue")
alpha <- 5 / 2
beta <- 2
t <- seq(from = -100,
         to = 100,
         length.out = 201)
plotReIm(function(t)
        cf_InverseGamma(t, alpha, beta, coef), t,
        title = "Characteristic function of the linear combination of IGamma RVs")

## EXAMPLE 2
# PDF/CDF from the CF by cf2DistGP
alpha <- 5 / 2
beta <- 1 / 2
coef <- 1 / (((1:50) - 0.5) * pi) ^ 2
cf <- function(t)
        cf_InverseGamma(t, alpha, beta, coef)
options <- list()
options$N <- 2 ^ 10
options$xMin = 0
x <- seq(from = 0,
         to = 4,
         length.out = 201)
prob <- c(0.9, 0.95, 0.99)
result <- cf2DistGP(cf, x, prob, options)
result
}
\references{
WITKOVSKY, V.: Computing the distribution of a linear combination
of inverted gamma variables, Kybernetika 37 (2001), 79-90.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Inverse-gamma_distribution}.

Other Continuous Probability Distribution: 
\code{\link{cfS_Arcsine}()},
\code{\link{cfS_Beta}()},
\code{\link{cfS_Gaussian}()},
\code{\link{cfS_Laplace}()},
\code{\link{cfS_Rectangular}()},
\code{\link{cfS_Student}()},
\code{\link{cfS_TSP}()},
\code{\link{cfS_Trapezoidal}()},
\code{\link{cfS_Triangular}()},
\code{\link{cfS_Wigner}()},
\code{\link{cfX_ChiSquare}()},
\code{\link{cfX_Exponential}()},
\code{\link{cfX_FisherSnedecor}()},
\code{\link{cfX_Gamma}()},
\code{\link{cfX_InverseGamma}()},
\code{\link{cfX_LogNormal}()},
\code{\link{cf_ArcsineSymmetric}()},
\code{\link{cf_BetaNC}()},
\code{\link{cf_BetaSymmetric}()},
\code{\link{cf_Beta}()},
\code{\link{cf_ChiSquare}()},
\code{\link{cf_Exponential}()},
\code{\link{cf_FisherSnedecorNC}()},
\code{\link{cf_FisherSnedecor}()},
\code{\link{cf_Gamma}()},
\code{\link{cf_Laplace}()},
\code{\link{cf_LogRV_BetaNC}()},
\code{\link{cf_LogRV_Beta}()},
\code{\link{cf_LogRV_ChiSquareNC}()},
\code{\link{cf_LogRV_ChiSquare}()},
\code{\link{cf_LogRV_FisherSnedecorNC}()},
\code{\link{cf_LogRV_FisherSnedecor}()},
\code{\link{cf_LogRV_MeansRatioW}()},
\code{\link{cf_LogRV_MeansRatio}()},
\code{\link{cf_LogRV_WilksLambdaNC}()},
\code{\link{cf_LogRV_WilksLambda}()},
\code{\link{cf_Normal}()},
\code{\link{cf_RectangularSymmetric}()},
\code{\link{cf_Student}()},
\code{\link{cf_TSPSymmetric}()},
\code{\link{cf_TrapezoidalSymmetric}()},
\code{\link{cf_TriangularSymmetric}()},
\code{\link{cf_vonMises}()}
}
\concept{Continuous Probability Distribution}
