% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_Laplace.R
\name{cf_Laplace}
\alias{cf_Laplace}
\title{Characteristic function of a linear combination (resp. convolution) of
independent LAPLACE random variables with location parameter \eqn{mu} (real)
and scale parameter \code{beta > 0}.}
\usage{
cf_Laplace(t, mu, beta, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{mu}{vector of real location parameters. If empty, default value is \code{mu = 0}.}

\item{beta}{vector of the scale parameters \code{beta > 0}. If empty, default value is \code{beta = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the LAPLACE random variables.
If \code{coef} is scalar, it is assumed that all coefficients are equal. If empty, default value is coef = 1.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y} is sum of \eqn{niid} iid random variables \eqn{Y},
where each \eqn{Y = sum_{i=1}^N coef(i) * X_i} is independently and identically distributed random variable.
If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent LAPLACE random variables.
}
\description{
\code{cf_Laplace(t,mu,beta,coef,niid)} evaluates the characteristic function
\eqn{cf(t)} of \eqn{Y = \sum_{i=1}^N coef_i * X_i}, where  \eqn{X_i ~ Laplace (\mu_i,\beta_{i})}
are inedependent RVs, with real location parameters \eqn{\mu_{i}} and the scale parameters \eqn{\beta_{i} > 0}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{Y} is defined by
\deqn{ cf(t)=Prod(exp(li* t * coef(i) \eqn {mu(i)} ) / (1+(t*coef(i)*\eqn{beta(i)})^2) )}
}
\note{
Ver.: 08-Aug-2021 16:19:30 (consistent with Matlab CharFunTool v1.5.1, 16-Aug-2018 16:00:43).
}
\examples{
## EXAMPLE1
# CF of the Laplace RV
mu <- 0
beta <- 1
t  <- seq(from = -10,
              to = 10,
              length.out =201)
plotReIm(function(t)
        cf_Laplace(t, mu, beta),
        t,
        title = "Characteristic function of the Laplace RVs")



##EXAMPLE2
# PDF/CDF of the Laplace RV
mu <- 0
beta <- 1
x <- seq(-5, 5, length.out = 101)

prop <- c(0.80, 0.85, 0.90, 0.925, 0.95, 0.975, 0.99, 0.995, 0.999)
cf <- function(t)
        cf_Laplace(t, mu, beta)
result <- cf2DistGP(cf, x, prob)

##EXAMPLE3
# PDF/CDF of the linear combination of Laplace RVs
mu <- c(-4, -1, 2, 3)
beta <- c(0.1, 0.2, 0.3, 0.4)
coef <- c(1, 2, 3, 4)
prob <- c(0.80, 0.85, 0.90, 0.925, 0.95, 0.975, 0.99, 0.995, 0.999)
cf <- function(t)
        cf_Laplace(t, mu, beta, coef)
options <- list()
options$N <- 2^12
result <- cf2DistGP(cf,prob=prob,options=options)

## EXAMPLE 4
# PDF/CDF of the linear combination of the Laplace RVs
mu <- c(-10, 10, 20, 30, 40)
beta <- c(1, 2, 3, 4, 5)
coef <- c(1/2, 1, 3/4, 5, 1)
prop <- c(0.80, 0.85, 0.90, 0.925, 0.95, 0.975, 0.99, 0.995, 0.999)
cf <- function(t)
        cf_Laplace(t, mu, beta, coef)
options <- list()
options$N <- 2^12
result <- cf2DistGP(cf,c(),prob,options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Laplace_distribution}.

Other Continuous Probability Distribution: 
\code{\link{cfS_Arcsine}()},
\code{\link{cfS_Beta}()},
\code{\link{cfS_Gaussian}()},
\code{\link{cfS_Laplace}()},
\code{\link{cfS_Rectangular}()},
\code{\link{cfS_Student}()},
\code{\link{cfS_TSP}()},
\code{\link{cfS_Trapezoidal}()},
\code{\link{cfS_Triangular}()},
\code{\link{cfS_Wigner}()},
\code{\link{cfX_ChiSquare}()},
\code{\link{cfX_Exponential}()},
\code{\link{cfX_FisherSnedecor}()},
\code{\link{cfX_Gamma}()},
\code{\link{cfX_InverseGamma}()},
\code{\link{cfX_LogNormal}()},
\code{\link{cf_ArcsineSymmetric}()},
\code{\link{cf_BetaNC}()},
\code{\link{cf_BetaSymmetric}()},
\code{\link{cf_Beta}()},
\code{\link{cf_ChiSquare}()},
\code{\link{cf_Exponential}()},
\code{\link{cf_FisherSnedecorNC}()},
\code{\link{cf_FisherSnedecor}()},
\code{\link{cf_Gamma}()},
\code{\link{cf_InverseGamma}()},
\code{\link{cf_LogRV_BetaNC}()},
\code{\link{cf_LogRV_Beta}()},
\code{\link{cf_LogRV_ChiSquareNC}()},
\code{\link{cf_LogRV_ChiSquare}()},
\code{\link{cf_LogRV_FisherSnedecorNC}()},
\code{\link{cf_LogRV_FisherSnedecor}()},
\code{\link{cf_LogRV_MeansRatioW}()},
\code{\link{cf_LogRV_MeansRatio}()},
\code{\link{cf_LogRV_WilksLambdaNC}()},
\code{\link{cf_LogRV_WilksLambda}()},
\code{\link{cf_Normal}()},
\code{\link{cf_RectangularSymmetric}()},
\code{\link{cf_Student}()},
\code{\link{cf_TSPSymmetric}()},
\code{\link{cf_TrapezoidalSymmetric}()},
\code{\link{cf_TriangularSymmetric}()},
\code{\link{cf_vonMises}()}
}
\concept{Continuous Probability Distribution}
