% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_LogRV_Beta.R
\name{cf_LogRV_Beta}
\alias{cf_LogRV_Beta}
\title{Characteristic function of a linear combination of independent
 LOG-TRANSFORMED BETA random variables}
\usage{
cf_LogRV_Beta(t, alpha, beta, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{alpha}{vector of the 'shape' parameters \code{alpha > 0}. If empty, default value is \code{alpha = 1}.}

\item{beta}{vector of the 'shape' parameters \code{beta > 0}. If empty, default value is \code{beta = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the
log-transformed random variables. If \code{coef} is scalar, it is
assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y} is
sum of \code{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N}
\eqn{coef(i) * log(X_i)} is independently and identically distributed
random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of independent
 LOG-TRANSFORMED BETA random variables.
}
\description{
\code{cf_LogRV_Beta(t, alpha, beta, coef, niid)} evaluates characteristic function of a linear combination (resp.
convolution) of independent LOG-TRANSFORMED BETA random variables (RVs)
\eqn{log(X)}, where \eqn{X ~ BETA(\alpha,\beta)}, and \eqn{\alpha > 0} and \eqn{\beta > 0}
represent the shape parameters of the BETA distribution.

That is, \code{cf_LogRV_Beta} evaluates the characteristic function \eqn{cf(t)} of  \eqn{Y =}
\eqn{coef_i*log(X_1) +...+ coef_N*log(X_N)}, where \eqn{X_i ~ Beta(\alpha_i,\beta_i)}
are inedependent RVs, with the shape parameters \eqn{\alpha_i > 0} and \eqn{\beta_i > 0}, for \eqn{i = 1,...,N}.
}
\details{
The characteristic function of \eqn{Y = log(X)}, with \eqn{X ~ Beta(\alpha,\beta)} is
defined by \eqn{cf_Y(t) = E(exp(1i*t*Y)) = E(exp(1i*t*log(X))) = E(X^(1i*t))}.
That is, the characteristic function can be derived from expression for
the \eqn{r}-th moment of \eqn{X}, E(X^r) by using \eqn{(1i*t)} instead of \eqn{r}. In
particular, the characteristic function of \eqn{Y = log(X)}, with \eqn{X ~}
\eqn{Beta(\alpha,\beta)} is defined by
\eqn{cf_Y(t) = gamma(\alpha + 1i*t) / gamma(\alpha) .* ...}
\eqn{gamma(\alpha + \beta) / gamma(\alpha + \beta + 1i*t)}.
Hence,the characteristic function of \eqn{Y  = coef(1)*Y1 + ... + coef(N)*YN}
is \eqn{cf_Y(t) = cf_Y1(coef(1)*t) * ... * cf_YN(coef(N)*t)}, where \eqn{cf_Yi(t)}
is evaluated with the parameters \code{alpha(i)} and \code{beta(i)}.
}
\examples{
## EXAMPLE 1
# CF of a linear combination of K=50 independent log-Beta RVs
coef <- 1 / ((((1:50) - 0.5) * pi) ^ 2)
plot(
        1:50,
        coef,
        xlab = "",
        ylab = "",
        type = "p",
        pch = 20,
        col = "blue",
        cex = 1,
        main = expression('Coefficients of the linear combination of' ~ chi ^ 2 ~ 'RVs with DF=1')
)
lines(1:50, coef, col = "blue")
alpha <- 5 / 2
beta <- 3 / 2
t <- seq(-100, 100, (100 - (-100)) / 200)
plotReIm(function(t)
        cf_LogRV_Beta(t, alpha, beta, coef),
        t,
        title = 'Characteristic function of a linear combination of log-Beta RVs')

#library(ggplot2)
#df <- data.frame(x_axis = t, y1_axis = Re(cf), y2_axis = Im(cf))
#ggplot(df, aes(x = x_axis)) +
#  geom_line(aes(y = y1_axis), colour="blue") +
#  geom_line(aes(y = y2_axis), colour = "red") +
#  ylab(label="") +
#  xlab("") +
#  ggtitle("Characteristic function of a linear combination of log-Beta RVs") +
#  theme(plot.title = element_text(hjust = 0.5))
#   figure; plot(t,real(cf),t,imag(cf)); grid on;
#   title('Characteristic function of a linear combination of log-Beta RVs')
#

##  EXAMPLE 2:
# PDF/CDF from the CF by cf2DistGP
alpha <- 5 / 2
beta  <- 3 / 2
coef <- 1 / ((((1:50) - 0.5) * pi) ^ 2)
cf <- function(t) {
        cf_LogRV_Beta(t, alpha, beta, coef)
}
options <- list()
options$xMax <- 0
result <- cf2DistGP(cf = cf, options = options)

## EXAMPLE 3:
# Distribution of log(R), where R=geometric/arithmetic mean of Gamma RVs
# Let X_1,...,X_n are iid RVs, X_j ~ Gamma(A,B), where A > 0 is the known
# shape parameter and B > 0 is the (unknown, common) rate parameter.
# Let R = geometricmean(X)/mean(X). According to Glaser (JASA 1976)
# log(R) ~ (1/n) * sum_{j=1}^{n-1} log(Y_j), Y_j ~ Beta(alpha,beta_j),
# where alpha = A and beta_j = j/n for j = 1,...,n-1. That is, log(R) is
# distributed as linear combination of independent logBeta random
# variables log(Y_j).
# Here we evaluate the PDF/CDF of W = -log(R) (i.e. minus of log(R))
n <- 10
alpha <- 1 #A = 1, i.e. X_j are from exponential distribution
beta <- (1:n - 1) / n
coef <- -1 / n
t <- seq(from = -25,
         to = 25,
         length.out = 201)
plotReIm(function(t)
        cf_LogRV_Beta(t, alpha, beta, coef),
        t,
        title = 'Characteristic function of a linear combination of log-Beta RVs')
prob <- c(0.9, 0.95, 0.99)
options <- list()
options.xMin = 0
result <- cf2DistGP(cf = cf, prob = prob, options = options)
str(result)

## EXAMPLE 4 (Distribution of Wilks Lambda statistic)
# If E ~ Wp(m,Sigma) and H ~ Wp (n,Sigma) with m >= p, then the
# distribution of Lambda = L = det(E)/det(E + H) is Wilks Lambda
# distribution, denoted by L ~ Lambda(p,m,n), with L in (0,1).
# It holds that L ~ Prod_{i=1}^p B_i, where B_i follow independent
# Beta distributions with Bi ~ B{(m + 1 - i)/2, n/2)}, for i = 1,...,p.
# Let W = -log(L) and cdf_W(x) = Prob(W <= x) = Prob(-sum(log(Bi)) <= x).
# Then, cdf_L(u) = Prob(L <= u) = Prob(W > -log(u)) =
# 1 - Prob(W <= -log(u)) = 1 - cdf_W(x), where x = -log(u) and u =
# exp(-x). Moreover, pdf_Lambda(u) = pdf_W(x)/exp(-x) = pdf_W(-log(u))/u.
# The Lambda statistic is used to test null (significance) hypothesis
# expressed by the matrix H. The null hypothesis is rejected for small
# values of the observed statistic Lambda, or large value -log(Lambda).
p <- 10
m <- 20
n <- 7
i  <- 1:p
alpha <- (m + 1 - i) / 2
beta  <- n / 2
coef  <- -1
t <- seq(from = -5,
         to = 5,
         length.out = 201)
plotReIm(function(t)
        cf_LogRV_Beta(t, alpha, beta, coef),
        t,
        title = 'Characteristic function of a linear combination of log-Beta RVs')
prob <- c(0.99, 0.95, 0.9)
x <- seq(from = 2,
         to = 8,
         length.out = 100)
options <- list()
options.xMin = 0

# Distribution of -log(Lambda)
result <- cf2DistGP(
        cf = cf,
        x = x,
        prob = prob,
        options = options
)
str(result)

# PDF of Lambda
# If W = -log(Lambda), then pdf_Lambda(u) = pdf_W(x)/exp(-x),
# where x = -log(u) and u = exp(-x)
plot.new()
plot.window(xlim = c(0, 0.15), ylim = c(0, 40))
axis(1)
axis(2)
#title(main=paste("PDF of Wilks", expression(Lambda), "distribution with p=10, m = 20, n=7"))
title(main = expression("PDF of Wilks" ~ Lambda ~ "distribution with p=10, m = 20, n=7"))
title(xlab = expression(Lambda))
title(ylab = "PDF")
grid(lty = "solid")
lines(exp(-result$x),
      result$pdf / exp(-result$x),
      col = "blue",
      lwd = 2)

# CDF of Lambda
# If W = -log(Lambda), then cdf_Lambda(u) = 1-cdf_W(x),
# where x = -log(u) and u = exp(-x)
plot.new()
plot.window(xlim = c(0, 0.15), ylim = c(0, 1))
axis(1)
axis(2)
#title(main=paste("PDF of Wilks", expression(Lambda), "distribution with p=10, m = 20, n=7"))
title(main = expression("CDF of Wilks" ~ Lambda ~ "distribution with p=10, m = 20, n=7"))
title(xlab = expression(Lambda))
title(ylab = "PDF")
grid(lty = "solid")
lines(exp(-result$x), 1 - result$cdf, col = "blue", lwd = 2)
print("Quantiles of Wilks Lambda distribution")
print(paste("alpha =", prob))
print(paste("quantile =", exp(-result$qf)))
}
\references{
GLASER, R.E. (1976). The ratio of the geometric mean to the arithmetic
mean for a random sample from a gamma distribution. \emph{Journal of the American Statistical Association}, 71(354), 480-487.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Beta_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
