% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_LogRV_BetaNC.R
\name{cf_LogRV_BetaNC}
\alias{cf_LogRV_BetaNC}
\title{Characteristic function of a linear combination
of independent LOG-TRANSFORMED non-central BETA random variables}
\usage{
cf_LogRV_BetaNC(t, alpha, beta, delta, coef, niid, tol)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{alpha}{vector of the 'shape' parameters \code{alpha > 0}. If empty, default value is \code{alpha = 1}.}

\item{beta}{vector of the 'shape' parameters \code{beta > 0}. If empty, default value is \code{beta = 1}.}

\item{delta}{vector of the non-centrality parameters \code{delta > 0}. If empty, default value is \code{delta = 0}.}

\item{coef}{vector of the coefficients of the linear combination of the Beta distributed random variables.
If coef is scalar, it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}

\item{tol}{tolerance factor for selecting the Poisson weights, i.e. such that \eqn{PoissProb > tol}.
If empty, default value is \code{tol = 1e-12}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent LOG-TRANSFORMED non-central BETA random variables.
}
\description{
\code{cf_LogRV_BetaNC(t, alpha, beta, delta, coef, niid, tol)} evaluates characteristic function
of a linear combination (resp. convolution) of independent LOG-TRANSFORMED non-central BETA random variables,
with distributions \eqn{Beta(\alpha_i,\beta_i,\delta_i)} defined on the interval \eqn{(0,1)}.

That is, \code{cf_LogRV_BetaNC} evaluates the characteristic function \eqn{cf(t)}
of \eqn{Y = coef_i*log(X_1) +...+ coef_N*log(X_N)}, where \eqn{X_i ~ Beta(\alpha_i,\beta_i,\delta_i)}
are inedependent RVs, with the shape parameters \eqn{\alpha_i > 0}, \eqn{\beta_i > 0},
and the noncentrality parameters \eqn{\delta_i > 0}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{Y = log(X)} with \eqn{X ~ Beta(\alpha,\beta,\delta)}
is Poisson mixture of CFs of the central log-transformed Beta RVs of the form
\deqn{cf(t) = cf_LogRV_BetaNC(t,\alpha,\beta,\delta) = exp(-\delta/2) sum_{j=1}^Inf (\delta/2)^j/j! * cf_LogRV_Beta(\alpha+j,\beta),}
where \eqn{cf_LogRV_Beta(\alpha+j,\beta)} are the CFs of central log-transformed
Beta RVs with parameters \eqn{\alpha+j} and \eqn{\beta}.
Hence, the characteristic function of \eqn{Y  = coef(1)*Y1 + ... + coef(N)*YN} is
\deqn{cf_Y(t) =  cf_Y1(coef(1)*t) * ... * cf_YN(coef(N)*t),}
where \eqn{cf_Yi(t)} is evaluated with the parameters \eqn{\alpha(i)} and \eqn{beta(i)}.
}
\examples{
## EXAMPLE 1
# CF of the log-transformed non-central Beta RV with delta = 1, coef = -1
alpha <- 1
beta  <- 3
delta <- 1
coef  <- -1
t <- seq(from = -10,
         to = 10,
         length.out = 201)
plotReIm(function(t)
        cf_LogRV_BetaNC(t, alpha, beta, delta, coef),
        t,
        title = 'CF of minus log-transformed Beta RV')

## EXAMPLE 2
# CDF/PDF of the minus log-transformed non-central Beta RV with delta = 1
alpha <- 1
beta  <- 3
delta <- 1
coef  <- -1
cf <- function(t)
        cf_LogRV_BetaNC(t, alpha, beta, delta, coef)
options <- list()
options$xMin <- 0
result <- cf2DistGP(cf = cf, options = options)

## EXAMPLE 3
# CDF/PDF of the linear combination of minus log-transformed non-central Beta RVs
alpha <- c(1, 2, 3)
beta  <- c(3, 4, 5)
delta <- c(0, 1, 2)
coef  <- c(-1,-1,-1) / 3
cf <- function(t)
        cf_LogRV_BetaNC(t, alpha, beta, delta, coef)
options <- list()
options$xMin <- 0
result <- cf2DistGP(cf = cf, options = options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Noncentral_beta_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PDF}}, \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}

Other Non-central Probability Distribution: \code{\link{cf_BetaNC}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_FisherSnedecorNC}}
}
