% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_LogRV_ChiSquare.R
\name{cf_LogRV_ChiSquare}
\alias{cf_LogRV_ChiSquare}
\title{Characteristic function of a linear combination
of independent LOG-TRANSFORMED CHI-SQUARE random variables}
\usage{
cf_LogRV_ChiSquare(t, df, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{df}{vector of the  degrees of freedom \code{df > 0}. If empty, default value is \code{df = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the logGamma random variables.
If coef is scalar, it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent LOG-TRANSFORMED CHI-SQUARE random variables.
}
\description{
\code{cf_LogRV_ChiSquare(t, df, coef, niid)} evaluates characteristic function of a linear combination
(resp. convolution) of independent LOG-TRANSFORMED CHI-SQUARE random variables (RVs)
\eqn{log(X)}, where \eqn{X ~ ChiSquare(df)} is central CHI-SQUARE distributed RV with df degrees of freedom.

That is, \code{cf_LogRV_ChiSquare} evaluates the characteristic function \eqn{cf(t)}
of \eqn{Y = coef_1*log(X_1) +...+ coef_N*log(X_N)}, where \eqn{X_i ~ ChiSquare(df_i)}, with \eqn{df_i > 0}
degrees of freedom, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{Y = log(X)}, with \eqn{X ~ ChiSquare(df)} is defined by
\deqn{cf_Y(t) = E(exp(1i*t*Y)) = E(exp(1i*t*log(X))) = E(X^(1i*t)).}
That is, the characteristic function can be derived from expression for the r-th moment of \eqn{X},
\eqn{E(X^r)} by using \eqn{(1i*t)} instead of \eqn{r}.
In particular, the characteristic function of \eqn{Y = log(X)} is
\deqn{cf_Y(t) = 2^(1i*t) * gamma(df/2 + 1i*t) / gamma(df/2).}

Hence,the characteristic function of \eqn{Y  = coef_1*X_1 +...+ coef_N*X_N}
is  \deqn{cf_Y(t) =  cf_1(coef_1*t) *...* cf_N(coef_N*t),}
where \eqn{cf_i(t)} is the characteristic function of the ChiSquare distribution with \eqn{df_i} degrees of freedom.
}
\note{
Ver.: 16-Sep-2018 18:34:40 (consistent with Matlab CharFunTool v1.3.0, 02-Jun-2017 12:08:24).
}
\examples{
## EXAMPLE 1
# CF of a weighted linear combination of independent log-ChiSquare RVs
coef <- c(1, 2, 3, 4, 5)
weight <- coef / sum(coef)
df <- c(1, 2, 3, 4, 5)
t <- seq(from = -20,
         to = 20,
         length.out = 1001)
plotReIm(function(t)
        cf_LogRV_ChiSquare(t, df, weight),
        t,
        title = 'CF of a linear combination of minus log-ChiSquare RVs')

## EXAMPLE 2
# PDF/CDF of a linear combination of independent log-ChiSquare RVs
coef <- c(1, 2, 3, 4, 5)
weight <- coef / sum(coef)
df <- c(1, 2, 3, 4, 5)
cf <- function(t)
        cf_LogRV_ChiSquare(t, df, weight)
options <- list()
options$N <- 2 ^ 12
prob <- c(0.9, 0.95, 0.99)
result <- cf2DistGP(cf = cf, prob = prob, options = options)
str(result)
}
\references{
[1] PHILLIPS, P.C.B. The true characteristic function of the F distribution. Biometrika (1982), 261-264.

[2] WITKOVSKY, V.: On the exact computation of the density and of the quantiles of linear combinations
of t and F random variables. Journal of Statistical Planning and Inference 94 (2001), 1–13.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Chi-squared_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}},
  \code{\link{cfX_FisherSnedecor}},
  \code{\link{cfX_Gamma}}, \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_LogRV_MeansRatioW}},
  \code{\link{cf_LogRV_MeansRatio}},
  \code{\link{cf_LogRV_WilksLambdaNC}},
  \code{\link{cf_LogRV_WilksLambda}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
