% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_LogRV_ChiSquareNC.R
\name{cf_LogRV_ChiSquareNC}
\alias{cf_LogRV_ChiSquareNC}
\title{Characteristic function of a linear combination of independent LOG-TRANSFORMED non-central ChiSquare random variables}
\usage{
cf_LogRV_ChiSquareNC(t, df, delta, coef, niid, tol)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{df}{vector of the  degrees of freedom \code{df > 0}. If empty, default value is \code{df = 1}.}

\item{delta}{vector of the non-centrality parameters \code{delta > 0}. If empty, default value is \code{delta = 0}.}

\item{coef}{vector of the coefficients of the linear combination of the Beta distributed random variables.
If coef is scalar, it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}

\item{tol}{tolerance factor for selecting the Poisson weights, i.e. such that \eqn{PoissProb > tol}.
If empty, default value is \code{tol = 1e-12}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of independent
LOG-TRANSFORMED non-central ChiSquare random variables.
}
\description{
\code{cf_LogRV_ChiSquareNC(t, df, delta, coef, niid, tol)} evaluates characteristic function
of a linear combination (resp. convolution) of independent LOG-TRANSFORMED non-central ChiSquare random variables,
with distributions \eqn{ChiSquare(df_i,\delta_i)}.

That is, \code{cf_LogRV_ChiSquareNC} evaluates the characteristic function \eqn{cf(t)}
of \eqn{Y = coef_i*log(X_1) +...+ coef_N*log(X_N)}, where \eqn{X_i ~ ChiSquare(df_i,\delta_i)}
are inedependent RVs, with \eqn{df_i} degrees of freedom and the noncentrality parameters \eqn{\delta_i >0}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{Y = log(X)} with \eqn{X ~ ChiSquare(df,\delta)} is Poisson mixture
of CFs of the central log-transformed ChiSquare RVs of the form
\deqn{cf(t) = cf_LogRV_ChiSquareNC(t,df,\delta) = exp(-\delta/2) sum_{j=1}^Inf (\delta/2)^j/j! * cf_LogRV_ChiSquare(df+2*j),}
where \eqn{cf_LogRV_ChiSquare(df+j)} are the CFs of central log-transformed ChiSquare RVs with \eqn{df+j} degrees of freedom.
Hence, the characteristic function of \eqn{Y  = coef(1)*Y1 + ... + coef(N)*YN}
\deqn{cf_Y(t) =  cf_Y1(coef(1)*t) * ... * cf_YN(coef(N)*t),}
where \eqn{cf_Yi(t)} is evaluated with the parameters \eqn{df_i} and \eqn{delta_i}.
}
\examples{
## EXAMPLE 1
# CF of the minus log-transformed non-central ChiSquare RV with delta = 1
df <- 3
delta <- 1
coef  <- -1
t <- seq(from = -10,
         to = 10,
         length.out = 201)
plotReIm(function(t)
        cf_LogRV_ChiSquareNC(t, df, delta, coef),
        t,
        title = 'CF of minus log-transformed ChiSquare RV')

## EXAMPLE 2
# CDF/PDF of the minus log-transformed non-central ChiSquare RV
df <- 3
delta <- 1
coef <- -1
cf <- function(t)
        cf_LogRV_ChiSquareNC(t, df, delta, coef)
options <- list()
options$N <- 2 ^ 10
result <- cf2DistGP(cf = cf, options = options)

## EXAMPLE 3
# CDF/PDF of the linear combination of the minus log-transformed non-central ChiSquare RVs
df <- c(3, 4, 5)
delta <- c(0, 1, 2)
coef  <- c(-1,-1,-1) / 3
cf <- function(t)
        cf_LogRV_ChiSquareNC(t, df, delta, coef)
options <- list()
options$N <- 2 ^ 10
result <- cf2DistGP(cf = cf, options = options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Noncentral_chi-squared_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}

Other Non-central Probability Distribution: \code{\link{cf_BetaNC}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_FisherSnedecorNC}}
}
