% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_LogRV_MeansRatio.R
\name{cf_LogRV_MeansRatio}
\alias{cf_LogRV_MeansRatio}
\title{Characteristic function of a linear combination of \eqn{N} independent
LOG-TRANSFORMED MEANS-RATIO random variables}
\usage{
cf_LogRV_MeansRatio(t, n, alpha, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{n}{vector of sample size parameters \eqn{n = (n_1,...,n_N)}.
If empty, default value is \code{n = (1,...,1)}.}

\item{alpha}{vector of the 'shape' parameters \eqn{\alpha = (\alpha_1,...,\alpha_N)}.
If empty, default value is \code{alpha = (1,...,1)}.}

\item{coef}{vector of the coefficients of the linear combination of the
log-transformed random variables. If \code{coef} is scalar, it is
assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y} is
sum of \code{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N}
\eqn{coef(i) * log(X_i)} is independently and identically distributed
random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of N independent
LOG-TRANSFORMED MEANS-RATIO random variables.
}
\description{
\code{cf_LogRV_MeansRatio(t, n, alpha, coef, niid)} evaluates characteristic function
of a linear combination (resp. convolution) of \eqn{N} independent LOG-TRANSFORMED MEANS-RATIO
random variables (RVs), \eqn{W_i = log(R_i)} for \eqn{i = 1,...,N} , where each \eqn{R_i = G_i/A_i} is a ratio
of \eqn{G_i} (the geometric mean) and \eqn{A_i} (the arithmetic mean) of a random sample
\eqn{X_{i,1},...,X_{i,n_i}} from a GAMMA distribution.

That is, cf_LogRV_MeansRatio evaluates the characteristic function
of a random variable \eqn{Y  = coef_1*W_1 +...+ coef_N*W_N}, such that
\eqn{cf_Y(t) =  cf_W_1(coef_1*t) *...* cf_W_N(coef_N*t)}, where \eqn{cf_W_i(t)}
is CF of \eqn{W_i = log(R_i)}, and  \eqn{R_i ~ MeansRatio(n_i,\alpha_i)},
where \eqn{MeansRatio(n_i,\alpha_i)} denotes the distribution of \eqn{R_i}, for \eqn{i = 1,...,N}.

Here, we define \eqn{G_i = (X_{i,1} *...* X_{i,n_i})^(1/n_i)} and
\eqn{A_i = (X_{i,1} +...+ X_{i,n_i})/n_i}, with \eqn{X_{i,j} ~ Gamma(\alpha_i,\beta_i)}
for all \eqn{j = 1,...,n_i}, where \eqn{\alpha_i} is the shape parameter and \eqn{\beta_i}
is the rate parameter of the GAMMA distribution in the i-th random sample.

Note that the \eqn{R_i} random variables are scale invariant,
so the distribution does not depend on the rate (scale) parameters \eqn{\beta_i}.

The MeansRatio distribution of \eqn{R_i ~ MeansRatio(k_i,\alpha_i)},
with \eqn{R_i} in \eqn{(0,1)}, is defined by \eqn{R_i ~ (Prod_{j=1}^{k_i-1} B_{i,j})^(1/k_i)},
where \eqn{B_{i,j} ~ Beta{\alpha_i, j/k_i)}} for \eqn{j = 1,...,k_{i-1}},
i.e. \eqn{B_{i,j}} follow independent Beta distributions for all \eqn{i = 1,...,N} and
\eqn{j = 1,...,k_{i-1}}. Alternatively, \eqn{log(R_i) ~ (log(B_{i,1}) +...+ log(B_{i,k_{i-1}}))/k_i}.
}
\details{
The MeansRatio distribution is the distribution of the Bartlett's test
 statistitic for testing homogeneity of variances of \eqn{k} populations,
 based on equal sample sizes of size \eqn{m}, i.e. with \eqn{df = m-1}
 for all \eqn{j = 1,...,k}. Let \eqn{DF = k*df}. The Bartlett test statistic defined
 in Glaser (1976b) is
 \deqn{L = Prod((df*S^2_j/sigma^2)^(df/DF)) / (df/DF)*Sum(df*S^2_j/sigma^2) = Prod((S^2_j)^(1/k))/Sum(S^2_j)/k,}
 where \eqn{S^2_j = (1/(m-1))*Sum(X_jk - mean(X_jk))^2}.
 Notice that \eqn{df*S^2_j/sigma^2 ~ Chi^2_df = Gamma(df/2,1/2)} for all \eqn{j = 1,...,k}.
 The exact critical values can be calculated from the distribution of \eqn{W = -log(L)}
 by inverting the characteristic function \code{cf_LogRV_MeansRatio}.
}
\note{
Ver.: 05-Oct-2018 13:54:18 (consistent with Matlab CharFunTool v1.3.0, 17-Jun-2017 17:18:39).
}
\examples{
## EXAMPLE 1
# CF of log MeansRatio RV with n = 5 and alpha = 7/2
n <- 5
alpha <- 7/2
t <- seq(-100, 100, length.out = 201)
plotReIm(function(t) cf_LogRV_MeansRatio(t, n, alpha), t,
         title = 'CF of log MeansRatio RV with n = 5 and alpha = 7/2')

## EXAMPLE 2
# CF of a weighted linear combination of minus log MeansRatio RVs
n <- c(5, 7, 10)
alpha <- c(7, 10, 15) / 2
coef <- -c(5, 7, 10) / 22
t <- seq(-100, 100, length.out = 201)
plotReIm(function(t) cf_LogRV_MeansRatio(t, n, alpha, coef), t,
         title = 'CF of a weighted linear combination of -log MeansRatio RVs')

## EXAMPLE 3
# PDF/CDF of minus log MeansRatio RV, n = 5 and alpha = 7/2, from its CF
n <- 5
alpha <- 7/2
coef <- -1
cf <- function(t) cf_LogRV_MeansRatio(t, n, alpha, coef)
x <- seq(0, 0.6, length.out = 100)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$xMin <- 0
result <- cf2DistGP(cf, x, prob, options)
str(result)

## EXAMPLE 4
# Compare the exact distribution with the Bartlett's approximation
k <- 25 # number of normal populations
df <- 3  # degrees of freedom used in each of n populations
DF <- k * df
alpha <- df / 2
C_B <- (1 + 1 / (3 * (k - 1)) * (k / df - 1 / DF))
coef <- -DF / C_B
cf <- function(t) cf_LogRV_MeansRatio(t, k, alpha, coef)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$xMin <- 0
result <- cf2DistGP(cf, prob = prob, options = options)
str(result)
x <- result$x
matplot(cbind(x, x), cbind(result$cdf, pchisq(x, k - 1)),
        xlab = 'corrected test statistic', ylab = 'CDF',
        main = 'Exact CDF vs. the Bartlett approximation')
matplot(cbind(x, x), cbind(result$cdf, pchisq(x, k - 1)),
        xlab = 'corrected test statistic', ylab = 'CDF',
        main = 'Exact CDF vs. the Bartlett approximation', type = "l", lwd = 2)
prob
result$qf
qchisq(prob, k-1)

##  EXAMPLE 5
# Exact Critical Values for Bartlett's Test for Homogeneity of Variances
# See and compare the selected results in Glaser (1976b, Table 1)
k <- 3
df <- c(4, 5, 6, 7, 8, 9, 10, 11, 14, 19, 24, 29, 49, 99)
alpha <- df / 2
coef <- -1
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$N <- 2^12
options$SixSigmaRule <- 15
options$xMin <- 0
options$isPlot <- FALSE
critW = matrix(0, length(df), length(prob))
for(i in 1:length(df)) {
        cf <- function(t) cf_LogRV_MeansRatio(t, k, alpha[i], coef)
        result <- cf2DistGP(cf, prob = prob, options = options)
        critW[i,] <- result$qf
}
critR <- exp(-critW)
cat('k = ', k, '\n')
cat('alpha = ', 1 - prob, '\n')
cat(critR)
}
\references{
Glaser, R. E. (1976a). The ratio of the geometric mean to the arithmetic
mean for a random sample from a gamma distribution.
\emph{Journal of the American Statistical Association}, 71(354), 480-487.

Glaser, R. E. (1976b). Exact critical values for Bartlett's test
for homogeneity of variances. \emph{Journal of the American Statistical Association}, 71(354), 488-490.

Chao, M. T., & Glaser, R. E. (1978). The exact distribution
of Bartlett's test statistic for homogeneity of variances with unequal sample sizes.
\emph{Journal of the American Statistical Association}, 73(362), 422-426.
}
\seealso{
For more details see WIKIPEDIA:
https://en.wikipedia.org/wiki/Beta_distribution, \cr
https://en.wikipedia.org/wiki/Bartlett%27s_test.

Other Continuous Probability Distribution: 
\code{\link{cfS_Arcsine}()},
\code{\link{cfS_Beta}()},
\code{\link{cfS_Gaussian}()},
\code{\link{cfS_Laplace}()},
\code{\link{cfS_Rectangular}()},
\code{\link{cfS_Student}()},
\code{\link{cfS_TSP}()},
\code{\link{cfS_Trapezoidal}()},
\code{\link{cfS_Triangular}()},
\code{\link{cfS_Wigner}()},
\code{\link{cfX_ChiSquare}()},
\code{\link{cfX_Exponential}()},
\code{\link{cfX_FisherSnedecor}()},
\code{\link{cfX_Gamma}()},
\code{\link{cfX_InverseGamma}()},
\code{\link{cfX_LogNormal}()},
\code{\link{cf_ArcsineSymmetric}()},
\code{\link{cf_BetaNC}()},
\code{\link{cf_BetaSymmetric}()},
\code{\link{cf_Beta}()},
\code{\link{cf_ChiSquare}()},
\code{\link{cf_Exponential}()},
\code{\link{cf_FisherSnedecorNC}()},
\code{\link{cf_FisherSnedecor}()},
\code{\link{cf_Gamma}()},
\code{\link{cf_InverseGamma}()},
\code{\link{cf_Laplace}()},
\code{\link{cf_LogRV_BetaNC}()},
\code{\link{cf_LogRV_Beta}()},
\code{\link{cf_LogRV_ChiSquareNC}()},
\code{\link{cf_LogRV_ChiSquare}()},
\code{\link{cf_LogRV_FisherSnedecorNC}()},
\code{\link{cf_LogRV_FisherSnedecor}()},
\code{\link{cf_LogRV_MeansRatioW}()},
\code{\link{cf_LogRV_WilksLambdaNC}()},
\code{\link{cf_LogRV_WilksLambda}()},
\code{\link{cf_Normal}()},
\code{\link{cf_RectangularSymmetric}()},
\code{\link{cf_Student}()},
\code{\link{cf_TSPSymmetric}()},
\code{\link{cf_TrapezoidalSymmetric}()},
\code{\link{cf_TriangularSymmetric}()},
\code{\link{cf_vonMises}()}
}
\concept{Continuous Probability Distribution}
