% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_LogRV_MeansRatioW.R
\name{cf_LogRV_MeansRatioW}
\alias{cf_LogRV_MeansRatioW}
\title{Characteristic function of a linear combination of \eqn{N} independent
LOG-TRANSFORMED WEIGHTED MEANS-RATIO random variables}
\usage{
cf_LogRV_MeansRatioW(t, n, alpha, weight, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{n}{vector of sample size parameters \eqn{n = (n_1,...,n_N)}.
If empty, default value is \code{n = (1,...,1)}.}

\item{alpha}{list of weights vectors with shape parameters \eqn{alpha[[i]] = (\alpha_{i,1},...,alpha_{i,n_i})}.
If empty, default value is \code{alpha[[i]] = c(1,...,1)}.}

\item{weight}{list of weights vectors with \eqn{weights[[i]] = (w_{i,1},...,w_{i,n_i})}, for \eqn{i = 1,...,N}.
If empty, default value is \code{weights[[i]]  = c(1/k_i,...,1/k_i)}.}

\item{coef}{vector of the coefficients of the linear combination of the
log-transformed random variables. If \code{coef} is scalar, it is
assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y} is
sum of \code{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N}
\eqn{coef(i) * log(X_i)} is independently and identically distributed
random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of N independent
LOG-TRANSFORMED WEIGHTED MEANS-RATIO random variables.
}
\description{
\code{cf_LogRV_MeansRatioW(t, n, alpha, weight, coef, niid)} evaluates characteristic function
of a linear combination (resp. convolution) of independent LOG-TRANSFORMED
WEIGHTED MEANS-RATIO random variables (RVs) \eqn{W_i = log(R_i)}, for \eqn{i = 1,...,N},
where each \eqn{R_i = G_i/A_i} is a ratio of the weighted geometric mean  \eqn{G_i}
and the (unweighted) arithmetic mean \eqn{A_i} of independent RVs \eqn{X_{i,1},...,X_{i,n_i}}
with GAMMA distributions, with the shape parameters \eqn{\alpha_{i,j}} and the (common) rate parameter
\eqn{\beta_i} for \eqn{i = 1,...,N} and \eqn{j = 1,...,n_i}.

That is, \code{cf_LogRV_MeansRatioW} evaluates the characteristic function of a random
variable \eqn{Y = coef_1*W_1 +...+ coef_N*W_N}, such that \eqn{cf_Y(t) = cf_W_1(coef_1*t) *...* cf_W_N(coef_N*t)},
where \eqn{cf_W_i(t)} is CF of \eqn{W_i = log(R_i)}, where \eqn{R_i} is the ratio statistic of the weighted geometric
mean and the arithmetic mean of independent gamma distributed RVs,
which distribution depends on the shape parameter \eqn{\alpha_{i,j}} and the vector
of weights \eqn{w_{i,j}}, for \eqn{i = 1,...,N} and \eqn{j = 1,...,n_i}.

Here, for each fixed \eqn{i = 1,...,N}, the weighted geometric mean is defined
by \eqn{G_i = (X_{i,1}^w_{i,1} *...* X_{i,n_i}^w_{i,n_i})}, where the weights \eqn{w_{i,j}},
\eqn{j = 1,...,n_i}, are such that \eqn{w_{i,1} +...+ w_{i,n_i} = 1}, and the arithmetic mean is defined
by \eqn{A_i = (X_{i,1} +...+ X_{i,n_i})/n_i}.
The random variables  \eqn{X_{i,j}} are mutually independent with \eqn{X_{i,j} ~ \Gamma(\alpha_{i,j},\beta_i)}
for all \eqn{i = 1,...,N} and \eqn{j = 1,...,n_i}, where \eqn{\alpha_{i,j}} are the shape parameters and
\eqn{\beta_i} is the (common) rate parameter of the gamma distributions.

Note that the weighted ratio random variables \eqn{R_i} are scale invariant,
so their distribution does not depend on the common rate (or scale) parameter \eqn{\beta_i}
for each \eqn{i = 1,...,N}.

The distribution of the logarithm of the means ratio \eqn{log(R_i)} is defined
by its characteristic function, see e.g. Chao and Glaser (JASA 1978), which is
\deqn{cf_{log(R_i)}(t) = (n_i)^(1i*t) * ... \Gamma(sum(\alpha_{i,j}))/\Gamma(sum(\alpha_{i,j})+1i*t)) * ... Prod_{j=1}^n_i \Gamma(\alpha_{i,j}+1i*w_{i,j}*t)/\Gamma(\alpha_{i,j})},
for each \eqn{i = 1,...,N}.
}
\note{
Ver.: 05-Oct-2018 17:54:23 (consistent with Matlab CharFunTool v1.3.0, 17-Jun-2017 17:18:39).
}
\examples{
## EXAMPLE 1
# CF of log MeansRatio RV with n = 5 and alpha = 7/2
n <- 5
alpha <- 7/2
t <- seq(-100, 100, length.out = 201)
plotReIm(function(t) cf_LogRV_MeansRatioW(t, n, alpha), t,
         title = 'CF of log MeansRatio RV with n = 5 and alpha = 7/2')

## EXAMPLE 2
# CF of a weighted linear combination of minus log MeansRatio RVs
rm(list=ls())
n <- c(5, 7, 10)
alpha <- list(c(7/2), c(10/2), c(3/2))
weight = list()
coef <- -1/3
t <- seq(-100, 100, length.out = 201)
plotReIm(function(t) cf_LogRV_MeansRatioW(t, n, alpha, weight, coef), t,
         title = 'CF of a weighted linear combination of minus log MeansRatio RVs')

## EXAMPLE 3
# PDF/CDF of minus log MeansRatio RV, n = 5 and alpha = 7/2, from its CF
rm(list=ls())
n <- 5
alpha <- 7/2
coef <- -1
cf <- function(t) cf_LogRV_MeansRatioW(t, n, alpha, coef = coef)
x <- seq(0, 0.6, length.out = 100)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$xMin <- 0
result <- cf2DistGP(cf, x, prob, options)
str(result)

## EXAMPLE 4
# PDF/CDF of minus log of the (weighted) means ratio RV, from its CF
rm(list=ls())
n <- 5
alpha <- list(c(3, 5, 7, 10, 3) / 2)
weight <- list(alpha[[1]] / sum(alpha[[1]]))
coef <- -1
cf <- function(t) cf_LogRV_MeansRatioW(t, n, alpha, weight, coef)
x <- seq(-0.15, 1.15, length.out = 200)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$N <- 2^12
result <- cf2DistGP(cf, prob = prob, options = options)
str(result)

## EXAMPLE 5
# Compare the exact distribution with the Bartlett's approximation
rm(list=ls())
k <- 15 # k normal populations with unequal sample sizes
df <- c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3) # degrees of freedom
DF <- sum(df)
alpha <- list(df / 2)
weight <- list(alpha[[1]] / sum(alpha[[1]]))
C <- DF * log(k * prod(weight[[1]] ^ weight[[1]]))
C_B <- 1 + 1 / (3 * (k - 1)) * (sum(1 / df) - 1 / DF)
shift <- C / C_B
coef <- -DF / C_B
cf_R <- function(t) cf_LogRV_MeansRatioW(t, k, alpha, weight, coef)
cf <- function(t) exp(1i * t * shift) * cf_R(t)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$xMin <- 0
options$N <- 2^12
result <- cf2DistGP(cf, prob = prob, options = options)
str(result)
x <- result$x
matplot(cbind(x, x), cbind(result$cdf, pchisq(x, k - 1)),
        xlab = 'corrected test statistic', ylab = 'CDF',
        main = 'Exact CDF vs. the Bartlett approximation')
matplot(cbind(x, x), cbind(result$cdf, pchisq(x, k - 1)),
        xlab = 'corrected test statistic', ylab = 'CDF',
        main = 'Exact CDF vs. the Bartlett approximation',
        type = "l", lwd = 2)
print(prob)
print(result$qf)
print(qchisq(prob, k - 1))
}
\references{
Glaser, R. E. (1976a). The ratio of the geometric mean to the arithmetic
mean for a random sample from a gamma distribution.
\emph{Journal of the American Statistical Association}, 71(354), 480-487.

Glaser, R. E. (1976b). Exact critical values for Bartlett's test
for homogeneity of variances. \emph{Journal of the American Statistical Association}, 71(354), 488-490.

Chao, M. T., & Glaser, R. E. (1978). The exact distribution
of Bartlett's test statistic for homogeneity of variances with unequal sample sizes.
\emph{Journal of the American Statistical Association}, 73(362), 422-426.
}
\seealso{
For more details see WIKIPEDIA:
https://en.wikipedia.org/wiki/Bartlett%27s_test.

Other Continuous Probability Distribution: 
\code{\link{cfS_Arcsine}()},
\code{\link{cfS_Beta}()},
\code{\link{cfS_Gaussian}()},
\code{\link{cfS_Laplace}()},
\code{\link{cfS_Rectangular}()},
\code{\link{cfS_Student}()},
\code{\link{cfS_TSP}()},
\code{\link{cfS_Trapezoidal}()},
\code{\link{cfS_Triangular}()},
\code{\link{cfS_Wigner}()},
\code{\link{cfX_ChiSquare}()},
\code{\link{cfX_Exponential}()},
\code{\link{cfX_FisherSnedecor}()},
\code{\link{cfX_Gamma}()},
\code{\link{cfX_InverseGamma}()},
\code{\link{cfX_LogNormal}()},
\code{\link{cf_ArcsineSymmetric}()},
\code{\link{cf_BetaNC}()},
\code{\link{cf_BetaSymmetric}()},
\code{\link{cf_Beta}()},
\code{\link{cf_ChiSquare}()},
\code{\link{cf_Exponential}()},
\code{\link{cf_FisherSnedecorNC}()},
\code{\link{cf_FisherSnedecor}()},
\code{\link{cf_Gamma}()},
\code{\link{cf_InverseGamma}()},
\code{\link{cf_Laplace}()},
\code{\link{cf_LogRV_BetaNC}()},
\code{\link{cf_LogRV_Beta}()},
\code{\link{cf_LogRV_ChiSquareNC}()},
\code{\link{cf_LogRV_ChiSquare}()},
\code{\link{cf_LogRV_FisherSnedecorNC}()},
\code{\link{cf_LogRV_FisherSnedecor}()},
\code{\link{cf_LogRV_MeansRatio}()},
\code{\link{cf_LogRV_WilksLambdaNC}()},
\code{\link{cf_LogRV_WilksLambda}()},
\code{\link{cf_Normal}()},
\code{\link{cf_RectangularSymmetric}()},
\code{\link{cf_Student}()},
\code{\link{cf_TSPSymmetric}()},
\code{\link{cf_TrapezoidalSymmetric}()},
\code{\link{cf_TriangularSymmetric}()},
\code{\link{cf_vonMises}()}
}
\concept{Continuous Probability Distribution}
