% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_LogRV_WilksLambda.R
\name{cf_LogRV_WilksLambda}
\alias{cf_LogRV_WilksLambda}
\title{Characteristic function of a linear combination of independent
LOG-TRANSFORMED WILK's LAMBDA distributed random variables}
\usage{
cf_LogRV_WilksLambda(t, p, n, q, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{p}{vector of the dimension parameters \eqn{p = (p_1,...,p_N)}.
If empty, default value is \code{p = (1,...,1)}.}

\item{n}{vector of degrees of freedom of the Wishart matrices \eqn{E_i},
\eqn{n = (n_1,...,n_N)}. If empty, default value is \code{n = (1,...,1)}.}

\item{q}{vector of degrees of freedom of the Wishart matrices \eqn{H_i},
\eqn{q = (q_1,...,q_N)}. If empty, default value is \code{q = (1,...,1)}.}

\item{coef}{vector of the coefficients of the linear combination of the log-transformed random variables.
If coef is scalar, it is assumed that all coefficients are equal.
If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient niid, such that \eqn{Z = Y + ... + Y}
is sum of niid random variables Y, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination of independent
LOG-TRANSFORMED WILK's LAMBDA distributed random variables.
}
\description{
\code{cf_LogRV_WilksLambda(t, p, n, q, coef, niid)} evaluates characteristic function
of a linear combination (resp. convolution) of independent
LOG-TRANSFORMED WILK's LAMBDA distributed random variables.
}
\details{
\code{cf_LogRV_WilksLambda} evaluates the characteristic function
of a random variable \eqn{Y  = coef_1*W_1 +...+ coef_N*W_N}, such that
\eqn{cf_Y(t) = cf_W_1(coef_1*t) *...* cf_W_N(coef_N*t)}, where \eqn{cf_W_i(t)} is CF of \eqn{W_i = log(\Lambda_i)},
and each \eqn{\Lambda_i} has central WILK's LAMBDA distribution, \eqn{\Lambda_i ~ Lambda(p_i,n_i,q_i)},
for \eqn{i = 1,...,N}.
In particular \eqn{\Lambda_i = det(E_i)/det(E_i + H_i)}, with  \eqn{E_i} and \eqn{H _i}
being independent random matrices with central Wishart distributions,
\eqn{E_i ~  Wp_i(n_i,\Sigma_i)} and \eqn{H_i ~ Wp_i(q_i,\Sigma_i)} with \eqn{n_i >= p_i} and \eqn{\Sigma_i > 0}
(an unknown positive definite symmetric covariance matrix), for all \eqn{i = 1,...,N}.

Each particular Lambda_i statistic can be used to test specific null hypothesis
(by measuring its effect expressed by the matrix \eqn{H_i}). In this case,
the null hypothesis is rejected for small values of the observed statistic \eqn{\Lambda_i},
or large value of \eqn{-log(\Lambda_i)}.

 The central Wilks' distribution of \eqn{\Lambda_i ~ Lambda(p_i,n_i,q_i)},
 with \eqn{Lambda_i} in \eqn{(0,1)}, is defined by \eqn{\Lambda_i ~ Prod_{j=1}^p_i B_{i,j}},
 where \eqn{B_{i,j} ~ Beta{(n_i+1-j)/2, q_i/2)}}, i.e. \eqn{B_{i,j}} follow independent Beta distributions
 for all \eqn{i = 1,...,N} and \eqn{j = 1,...,p_i}.
}
\note{
Ver.: 04-Oct-2018 13:30:38 (consistent with Matlab CharFunTool v1.3.0, 19-Jul-2018 17:23:23).
}
\examples{
## EXAMPLE 1
# CF of log Wilks Lambda RV distribution Lambda(p,n,q)
p <- 5
n <- 10 # d.f. of within SS&P
q <- 3 # d.f. of between SS&P
t <- seq(-10, 10, length.out = 201)
plotReIm(function(t)
        cf_LogRV_WilksLambda(t, p, n, q),
        t,
        title = 'CF of log Wilks Lambda RV with Lambda(p,n,q), p=5, n=10, q=3')

## EXAMPLE 2
# CF of a weighted linear combination of minus log Wilks Lambda RVs
p <- c(5, 5, 5)
n <- c(10, 15, 20)
q <- c(3, 2, 1)
coef <- -c(10, 15, 20) / 45
t <- seq(-20, 20, length.out = 201)
plotReIm(function(t)
        cf_LogRV_WilksLambda(t, p, n, q, coef),
        t,
        title = 'CF of a weighted linear combination of -log Wilks Lambda RVs')

## EXAMPLE 3
# PDF/CDF of minus log Wilks Lambda RV (p=5, n=10, q=3) from its CF
p <- 5
n <- 10
q <- 3
coef <- -1
cf <- function(t) {cf_LogRV_WilksLambda(t, p, n, q, coef)}
x <- seq(0, 5, length.out = 100)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$xMin <- 0
result <- cf2DistGP(cf, x, prob, options)
str(result)

## EXAMPLE 4
# Compare the exact distribution with the Bartlett's approximation
# The Bartlett's approximation (see e.g. Wikipedia) is given by:
# ((p-q+1)/2 - n)*log(Lambda(p,n,q)) ~ chi^2_{q*p}
p <- 15
n <- 30
q <- 3
coef <- (p - q + 1) / 2 - n
cf <- function(t) {cf_LogRV_WilksLambda(t, p, n, q, coef)}
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$xMin <- 0
result <- cf2DistGP(cf, prob = prob, options = options)
str(result)
x <- result$x
matplot(cbind(x, x), cbind(result$cdf, pchisq(x, q * p)),
        main = 'Exact CDF vs. the Bartlett approximation',
        xlab = expression(paste("transformed ", Lambda, "(p,n,q)")),
        ylab = 'CDF')
matplot(cbind(x, x), cbind(result$cdf, pchisq(x, q * p)),
        main = 'Exact CDF vs. the Bartlett approximation',
        xlab = expression(paste("transformed ", Lambda, "(p,n,q)")),
        ylab = 'CDF', lwd = 2, type = "l")
prob
result$qf
qchisq(prob, q * p)
}
\references{
[1] Witkovsky, V., 2018. Exact distribution of selected multivariate test
criteria by numerical inversion of their characteristic functions.
arXiv preprint arXiv:1801.02248.
}
\seealso{
For more details see WIKIPEDIA:
https://en.wikipedia.org/wiki/Wilks%27s_lambda_distribution.

Other Continuous Probability Distribution: 
\code{\link{cfS_Arcsine}()},
\code{\link{cfS_Beta}()},
\code{\link{cfS_Gaussian}()},
\code{\link{cfS_Laplace}()},
\code{\link{cfS_Rectangular}()},
\code{\link{cfS_Student}()},
\code{\link{cfS_TSP}()},
\code{\link{cfS_Trapezoidal}()},
\code{\link{cfS_Triangular}()},
\code{\link{cfS_Wigner}()},
\code{\link{cfX_ChiSquare}()},
\code{\link{cfX_Exponential}()},
\code{\link{cfX_FisherSnedecor}()},
\code{\link{cfX_Gamma}()},
\code{\link{cfX_InverseGamma}()},
\code{\link{cfX_LogNormal}()},
\code{\link{cf_ArcsineSymmetric}()},
\code{\link{cf_BetaNC}()},
\code{\link{cf_BetaSymmetric}()},
\code{\link{cf_Beta}()},
\code{\link{cf_ChiSquare}()},
\code{\link{cf_Exponential}()},
\code{\link{cf_FisherSnedecorNC}()},
\code{\link{cf_FisherSnedecor}()},
\code{\link{cf_Gamma}()},
\code{\link{cf_InverseGamma}()},
\code{\link{cf_Laplace}()},
\code{\link{cf_LogRV_BetaNC}()},
\code{\link{cf_LogRV_Beta}()},
\code{\link{cf_LogRV_ChiSquareNC}()},
\code{\link{cf_LogRV_ChiSquare}()},
\code{\link{cf_LogRV_FisherSnedecorNC}()},
\code{\link{cf_LogRV_FisherSnedecor}()},
\code{\link{cf_LogRV_MeansRatioW}()},
\code{\link{cf_LogRV_MeansRatio}()},
\code{\link{cf_LogRV_WilksLambdaNC}()},
\code{\link{cf_Normal}()},
\code{\link{cf_RectangularSymmetric}()},
\code{\link{cf_Student}()},
\code{\link{cf_TSPSymmetric}()},
\code{\link{cf_TrapezoidalSymmetric}()},
\code{\link{cf_TriangularSymmetric}()},
\code{\link{cf_vonMises}()}
}
\concept{Continuous Probability Distribution}
