% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_Normal.R
\name{cf_Normal}
\alias{cf_Normal}
\title{Characteristic function of a linear combination of independent NORMAL random variables.}
\usage{
cf_Normal(t, mu, sigma, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{mu}{vector of the 'location' parameters \code{mu} in \eqn{R}. If empty, default value is \code{mu = 0}.}

\item{sigma}{vector of the 'scale' parameters \code{sigma > 0}. If empty, default value is \code{sigma = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the Normal random variables.
If \code{coef} is scalar, it is
assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)} is independently
and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent NORMAL random variables.
}
\description{
\code{cf_Normal(t, mu, sigma, coef, niid)} evaluates the characteristic function
\eqn{cf(t)} of a linear combination (resp. convolution) of independent NORMAL random variables.

That is, \code{cf_Normal} evaluates the characteristic function \eqn{cf(t)}
of  \eqn{Y = sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ N(\mu_i,\sigma_i)} are inedependent RVs,
with means \eqn{\mu_i} and standard deviations \eqn{\sigma_i > 0}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{Y} is defined by
\deqn{cf(t) = exp(1i*(coef'*\mu)*t - (1/2)*(coef^2'*\sigma^2)*t^2).}
}
\note{
The characteristic function of a lienar combination of independent
Normal random variables has well known exact form.
}
\examples{
## EXAMPLE 1
# CF of a linear combination of K=100 independent Norma RVs
coef <- 1 / (((1:50) - 0.5) * pi) ^ 2
plot(
        1:50,
        coef,
        xlab = "",
        ylab = "",
        type = "p",
        pch = 20,
        col = "blue",
        cex = 1,
        main = expression('Coefficients of the linear combination of' ~ chi ^ 2 ~ 'RVs with DF=1')
)
lines(1:50, coef, col = "blue")
mu <- seq(from = -3,
          to = 3,
          length.out = 50)
sigma <- seq(from = 0.1,
             to = 1.5,
             length.out = 50)
t <- seq(from = -100,
         to = 100,
         length.out = 2001)
plotReIm(function(t)
        cf_Normal(t, mu, sigma, coef),
        t,
        title = "Characteristic function of the linear combination of Normal RVs")

## EXAMPLE 2
# PDF/CDF from the CF by cf2DistGP
mu <- seq(from = -3,
          to = 3,
          length.out = 50)
sigma <- seq(from = 0.1,
             to = 1.5,
             length.out = 50)
coef <- 1 / (((1:50) - 0.5) * pi) ^ 2
cf <- function(t)
        cf_Normal(t, mu, sigma, coef)
options <- list()
options$N <- 2 ^ 10
options$SixSigmaRule <- 8
prob <- c(0.01, 0.05, 0.1, 0.5, 0.9, 0.950, 0.99)
result <- cf2DistGP(cf = cf, prob = prob, options = options)
result
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Normal_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
