% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_Student.R
\name{cf_Student}
\alias{cf_Student}
\title{Characteristic function of a linear combination
of independent STUDENT's t random variables}
\usage{
cf_Student(t, df, mu, sigma, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{df}{the degrees of freedom, \code{df > 0}. If empty, the default value is \code{df = 1}.}

\item{mu}{vector of location parameters, \code{mu} in Real. If empty, default value is \code{mu = 0}.}

\item{sigma}{vector of scale parameters, \code{sigma_i > 0}. If empty, default value is \code{sigma = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the STUDENT's t random variables.
If \code{coef} is scalar, it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient, such that \eqn{Z = Y + ... + Y} is
sum of \eqn{niid} random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef_i * X_i}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent STUDENT's t random variables.
}
\description{
\code{cf_Student(t, df, mu, sigma, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
of a linear combination (resp. convolution) of independent (location and scale shifted) STUDENT's t random variables.

That is, \code{cf_Student} evaluates the characteristic function \eqn{cf(t)}
of \eqn{Y = sum_{i=1}^N coef_i * (mu_i + sigma_i * X_i)}, where \eqn{X_i ~ t(df_i)}
are inedependent (symmetric) t-distributed RVs, with \eqn{df_i > 0} degrees of freedom, for \eqn{i = 1,...,N}.

The characteristic function of the random variable \eqn{\mu + \sigma*X}, where \eqn{X ~ t(df)} is given by
\deqn{cf(t) = exp(1i*t*mu) * besselk(df/2,abs(sigma*t)*sqrt(df),1) * exp(-abs(sigma*t)*sqrt(df)) * (sqrt(df)*abs(aigma*t))^(df/2) / 2^(df/2-1)/gamma(df/2).}

Hence, the characteristic function of \eqn{Y  = coef_1*(\mu_1+\sigma_1*X_1) +...+ coef_N*(\mu_N+\sigma_N*X_N)}
is \eqn{cf_Y(t) = exp(1i*\mu*t) * (cf_1(coef_1*\sigma_1*t) *...* cf_N(coef_N*\sigma_N*t))}, where \eqn{cf_i(t)}
is the characteristic function of \eqn{X_i ~ t(df_i)}.
}
\examples{
## EXAMPLE 1
# CF of a linear combination of independent Student's t RVs
coef <- 1/(1:50)
df   <- seq(50,1,-1)
t <- seq(from = -1,to = 1,length.out = 201)
plotGraf(function(t)
  cf_Student(t,df,coef), t, title = "Characteristic function of the linear combination of t RVs")

## EXAMPLE 2
# CDF/PDF of a linear combination of independent Student's t RVs
coef <- 1/(1:50)
df <- seq(50,1,-1)
cf <- function(t) cf_Student(t,df,coef)
x <- seq(from = -50,to = 50,length.out = 100)
prob <- c(0.9, 0.95, 0.975, 0.99)
options <- list()
options$N <- 2^12
result <- cf2DistGP(cf,x,prob,options)
result
}
\references{
WITKOVSKY V. (2016). Numerical inversion of a characteristic
function: An alternative tool to form the probability distribution
of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Student's_t-distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_Gamma}}, \code{\link{cf_InverseGamma}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}

Other Symetric Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_TrapezoidalSymmetric}}
}
