% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_TriangularSymmetric.R
\name{cf_TriangularSymmetric}
\alias{cf_TriangularSymmetric}
\title{Characteristic function of of a linear combination
of independent zero-mean symmetric TRIANGULAR random variables}
\usage{
cf_TriangularSymmetric(t, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{coef}{vector of the coefficients of the linear combination
of the zero-mean symmetric TRIANGULAR random variables. If coef is scalar,
it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent zero-mean symmetric TRIANGULAR random variables.
}
\description{
\code{cf_TriangularSymmetric(t, coef, niid)} evaluates the characteristic function
of a linear combination (resp. convolution) of independent zero-mean symmetric TRIANGULAR
random variables defined on the interval \eqn{(-1,1)}.

That is, \code{cf_TriangularSymmetric} evaluates the characteristic function
\eqn{cf(t)} of  \eqn{Y = sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ TriangularSymmetric}
are independent uniformly distributed RVs defined on \eqn{(-1,1)}, for all \eqn{i = 1,...,N}.

The characteristic function of \eqn{X ~ TriangularSymmetric} is defined
by \deqn{cf(t) = cf_TriangularSymmetric(t) = (2-2*cos(t))/t^2.}
}
\examples{
## EXAMPLE 1
# CF of the Triangular distribution on (-1,1)
t <- seq(from = -50,to = 50,length.out = 201)
plotGraf(function(t)
  cf_TriangularSymmetric(t), t, title = "CF of the Triangular distribution on (-1,1)")

## EXAMPLE 2
# CF of a weighted linear combination of independent Triangular RVs
t <- seq(from = -20,to = 20,length.out = 201)
coef <- c(1, 2, 3, 4, 5)/15
plotGraf(function(t)
  cf_TriangularSymmetric(t,coef), t, title = "CF of a weighted linear combination of Triangular RVs")

## EXAMPLE 3
# PDF/CDF of a weighted linear combination of independent Triangular RVs
coef <- c(1, 2, 3, 4, 5)/15
cf <- function(t) cf_TriangularSymmetric(t,coef)
x <- seq(from = -1,to = 1,length.out = 201)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$N <- 2^12
options$xMin <- -1
options$xMax <- 1
result <- cf2DistGP(cf,x,prob,options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Triangular_distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_Gamma}}, \code{\link{cf_InverseGamma}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}}
}
