% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_WignerSemicircle.R
\name{cf_WignerSemicircle}
\alias{cf_WignerSemicircle}
\title{Characteristic function of a linear combination (resp. convolution) of
independent WIGNER SEMICIRCLE random variables  defined on the interval
\eqn{(mu-R,mu+R)}.}
\usage{
cf_WignerSemicircle(t, mu, R, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{mu}{vector of the 'location' parameters mu in R. If empty, default value is \code{mu = 0}.}

\item{R}{vector of the 'radius' prameters \eqn{R > 0}. If empty, default value is \code{R =1}.}

\item{coef}{vector of the coefficients of the linear combination of the Beta distributed random variables.
If \code{coef} is scalar, it is assumed that all coefficients are equal. If empty, default value is coef = 1.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y} is sum of \eqn{niid} iid random variables \eqn{Y},
where each \eqn{Y = sum_{i=1}^N coef(i) * X_i} is independently and identically distributed random variable.
If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear
combination (res. convolution) of independent WIGNER SEMICIRCLE random variables.
}
\description{
That is, \code{cf_WignerSemicircle(t,mu,R,coef,niid)} evaluates the characteristic function
 \eqn{cf(t)} of  \eqn{Y = sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ WignerSemicircle}
 are independent RVs defined on \eqn{(mu_i-R_i,mu_i+R_i)}, for all \eqn{i = 1,...,N}.

 The characteristic function of \eqn{X ~ WignerSemicircle(mu,R)} is defined by
 \eqn{cf(t) = 2*exp(1i*t*mu).*besselj(1,R*t)/(R*t)};

 @family Continuous Probability Distribution

 @seealso For more details see WIKIPEDIA:
 \url{https://en.wikipedia.org/wiki/Wigner_semicircle_distribution}.
}
\note{
Ver.: 10-Aug-2021 17:44:17 (consistent with Matlab CharFunTool v1.5.1, 18-Sep-2018 00:32:58).
}
\examples{
## EXAMPLE 1
# CF of the symmetric Wigner Semicircle distribution on (-1,1)
t <- seq(from = -30,
         to = 30,
         length.out =501)
plotReIm(function(t)
        cf_WignerSemicircle(t),
        t,
        title = "CF of the Wigner Semicircle distribution on (-1,1)")



##EXAMPLE2
# PDF/CDF of Wigner Semicircle  RVs
cf <- function(t)
        cf_WignerSemicircle(t)
x <- seq(-1,1,length.out = 201)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$xMin <- -1
options$xMax <- 1
result <- cf2DistGP(cf, x, prob, options)

##EXAMPLE3
# CF of a linear combination of independent Wigner Semicircle RVs
t <- seq(-1,1, length.out=  501)
mu <- c(0, 1, 2, 1, 0)
R <- c(1, 1, 2, 2, 3)
coef <- c(1, 2, 3, 4, 5 )
plotReIm(function(t)
        cf_WignerSemicircle(t, mu, R, coef),
        t,
        title = "CF of a linear combination of independent Wigner Semicircle RVs")

## EXAMPLE 4
# PDF/CDF of a linear combination of independent Wigner Semicircle RVs
mu <- c(0, 1, 2, 1, 0)
R <- c(1, 1, 2, 2, 3)
coef <- c(1, 2, 3, 4, 5)
cf <- function(t)
        cf_WignerSemicircle(t, mu, R, coef)
x <- seq(-20, 40, length.out = 201)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$N = 2^12
result <- cf2DistGP(cf, x, prob, options)
}
