% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfC_vonMises.R
\name{cfC_vonMises}
\alias{cfC_vonMises}
\title{Characteristic function of a linear combination
of independent VON MISES random variables}
\usage{
cfC_vonMises(t, mu = 0, kappa = 1, coef, niid)
}
\arguments{
\item{t}{numerical values (number, vector...).}

\item{mu}{in \eqn{(-\pi, \pi)}.}

\item{kappa}{\eqn{> 0}.}

\item{coef}{vector of the coefficients of the linear combination
of the IGamma random variables. If coef is scalar, it is assumed
that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y +...+ Y} is
sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef_i * X_i}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent VON MISES random variables.
}
\description{
\code{cfC_vonMises(t, mu, kappa, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
of \eqn{Y = sum_{i=1}^N coef_i * X_i} where \eqn{X_i ~ vonMises(\mu_i,\kappa_i)}
inedependent RVs, with the locarion parameters \eqn{\mu_i} in Real and the rate
parameters \eqn{\kappa_i > 0}, for \eqn{i = 1,...,N}.

The characteristic function of the \eqn{vonMises(\mu,\kappa)} distribution is
\deqn{cf(t) = cf_vonMises(t,\mu,\kappa) = besseli(t,\kappa)/besseli(0,\kappa) * exp(1i*t*\mu).}
}
\details{
The VON MISES distribution is circular distribution on the interval
of length \eqn{2*\pi}, here we consider \eqn{(-\pi,\pi)}, equivalent of the normal
distribution with the real parameter \eqn{\mu} and rate parameter \eqn{\kappa > 0}
(\eqn{\mu} and \eqn{1/\kappa} are analogous to \eqn{\mu} and \eqn{\sigma^2}, the mean and variance
in the normal distribution), on a whole circle, i.e. the interval of angles \eqn{(-\pi,\pi)}.
}
\examples{
## EXAMPLE1
# CF of the uniform von Mises distribution on (-pi,pi)
t <- seq(-10, 10, length.out = 501)
plotGraf(function(t)
        cfC_vonMises(t), t, title = "CF of the uniform von Mises distribution on (-pi,pi)")

## EXAMPLE2
# CF of the mixture of the von Mises distribution on (-pi,pi)
mu1 <- 0
kappa1 <- 5
mu2 <- 1
kappa2 <- 15
cf <-
        function(t)
                0.25 * cfC_vonMises(t, mu1, kappa1) + 0.75 * cfC_vonMises(t, mu2, kappa2)
t <- seq(-10, 10, length.out = 501)
plotGraf(cf, t, title = "CF of the mixture of the von Mises distribution")

## EXAMPLE3
# PDF/CDF of the von Mises distribution on (-pi,pi)
mu <- 0
kappa <- 5
cf  <- function(t)
        cfC_vonMises(t, mu, kappa)
options <- list()
options$xMin <- -pi
options$xMax <- pi
result <- cf2DistGP(cf = cf, options = options)
angle <- result$x
radius <- result$pdf
plotPolar(angle, radius)

## EXAMPLE4
# PDF/CDF of the linear combinantion of 2 von Mises distribution on (-pi,pi)
mu1 <- 0
kappa1 <- 5
mu2 <- 1
kappa2 <- 15
cf <-
        function(t)
                cfC_vonMises(1 * t, mu1, kappa1) * cfC_vonMises(0.25 * t, mu2, kappa2)
options <- list()
options$xMin <- -pi
options$xMax <- pi
options$isCircular <- TRUE
result <- cf2DistGP(cf = cf, options = options)
angle <- result$x
radius <- result$pdf
plotPolar(angle, radius)

## EXAMPLE5
# PDF/CDF of the mixture of the von Mises distribution on (0,2*pi)
mu1 <- 0
kappa1 <- 5
mu2 <- 1
kappa2 <- 15
mu3 <- pi
kappa3 <- 10
cf <-
        function(t)
                0.25 * cfC_vonMises(t, mu1, kappa1) + 0.25 * cfC_vonMises(t, mu2, kappa2) + 0.5 *
        cfC_vonMises(t, mu3, kappa3)
options <- list()
options$xMin <- 0
options$xMax <- 2 * pi
options$isCircular <- TRUE
result <- cf2DistGP(cf, options)
angle <- result$x
radius <- result$pdf
plotPolar(angle, radius)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Von_Mises_distribution}.

Other Circular Probability Distribution: \code{\link{cf_vonMises}}
}
