% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfE_Empirical.R
\name{cfE_Empirical}
\alias{cfE_Empirical}
\title{Characteristic function of the EMPIRICAL distribution, based on the observed data}
\usage{
cfE_Empirical(t, data, cfX)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{data}{vector of data, i.e. constants where the DIRAC RVs
are concentrated. If empty, default value is \code{data = 1}.}

\item{cfX}{function handle of the characteristic function of a random
variable \eqn{X}. If \code{cfX} is non-empty, a compound CF is evaluated as \eqn{cf(t) = cf(-1i*log(cfX(t)),data)}.}
}
\value{
Characteristic function \eqn{cf(t)} of the EMPIRICAL distribution, based on the observed data.
}
\description{
That is, \eqn{cf(t) = (1/N)*(cfD(data_1*t) +...+ cfD(data_N*t))}, where
\eqn{cfD(t)} represents the characteristic function of the DIRAC RV
concentrated at the constant \eqn{d=1}, i.e. \eqn{cfD(t) = exp(1i*t)}.

\code{cfE_Empirical(t, data, cfX)} evaluates the compound characteristic function
\eqn{cf(t) = cfE_Empirical(-1i*log(cfX(t)),data) = (1/N) * sum_{j=1}^N cfX(t)^data_j};
where \code{cfX} is function handle of the characteristic function \eqn{cfX(t)} of the random variable \eqn{X}.
}
\note{
Ver.: 23-Sep-2018 14:40:43 (consistent with Matlab CharFunTool v1.3.0, 15-Sep-2018 12:56:00).
}
\examples{
## EXAMPLE1
# Empirical CF - a weighted mixture of independent Dirac variables
set.seed(101)
n <- 1000
data <- c(rnorm(3 * n, 5, 0.2), rt(n, 3), rchisq(n, 1))
t <- seq(-50, 50, length.out = 2 ^ 10)
plotReIm(function(t)
        cfE_Empirical(t, data),
        t,
        title = "Empirical CF - CF of the mixture of Dirac random variables")

## EXAMPLE2
# Convolution of the ECF and the Gaussian kernel)
set.seed(101)
n <- 1000
data <- c(rnorm(3 * n, 5, 0.2), rt(n, 3), rchisq(n, 1))
bandwidth <- 0.25
cf_DATA   <- function(t)
        cfE_Empirical(t, data)
cf_KERNEL <- function(t)
        exp(-(bandwidth * t) ^ 2 / 2)
cf <- function(t)
        cf_DATA(t) * cf_KERNEL(t)
t <- seq(-50, 50, length.out = 2 ^ 10)
plotReIm(cf, t, title = "Smoothed Empirical CF")
result <- cf2DistGP(cf)

## EXAMPLE3
# (PDF/CDF of the compound Empirical-Empirical distribution)
set.seed(101)
lambda <- 25
nN <- 10
Ndata <- rpois(nN, lambda)

mu <- 0.1
sigma <- 2
nX <- 1500
Xdata <- rlnorm(nX, mu, sigma)
cfX <- function(t)
        cfE_Empirical(t, Xdata)
cf  <- function(t)
        cfE_Empirical(t, Ndata, cfX)
t <- seq(-0.2, 0.2, length.out = 2 ^ 10)
plotReIm(cf, t, title = "Compound Empirical CF")

x <- seq(0, 1000, length.out = 501)
prob <- c(0.9, 0.95)
options <- list()
options$N <- 2 ^ 10
options$SixSigmaRule <- 10
result <- cf2DistGP(cf, x, prob, options)
}
\references{
WITKOVSKY V., WIMMER G., DUBY T. (2017). Computing the aggregate
loss distribution based on numerical inversion of the compound empirical
characteristic function of frequency and severity. arXiv preprint arXiv:1701.08299.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Empirical_characteristic_function}.

Other Empirical Probability Distribution: \code{\link{cfE_DiracMixture}},
  \code{\link{cfE_EmpiricalOgive}}
}
