% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfE_EmpiricalBootstrapped.R
\name{cfE_EmpiricalBootstrapped}
\alias{cfE_EmpiricalBootstrapped}
\title{Characteristic function of the EMPIRICAL BOOTSTRAPPED distribution, based on the bootstrapped data}
\usage{
cfE_EmpiricalBootstrapped(t, data, randID, cfX)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{data}{vector of original data. If empty, default value is \code{data = 1}.}

\item{randID}{vector of pseudorandom  integers between 1 and \code{lemngth(data)}. If
empty, randID is randomly generated by \code{randID = randi(length(data),size(data))}.}

\item{cfX}{function handle of the characteristic function of a random
variable \eqn{X}. If \code{cfX} is non-empty, a compound CF is evaluated as \eqn{cf(t) = cf(-1i*log(cfX(t)),data)}.}
}
\value{
Characteristic function \eqn{cf(t)} of the EMPIRICAL BOOTSTRAPPED distribution, based on the bootstrapped data.
}
\description{
That is, \eqn{cf(t) = (1/N)*(cfD(dataBoot_1*t) +...+ cfD(dataBoot_N*t))}, where
\eqn{cfD(t)} represents the characteristic function of the DIRAC RV
concentrated at the constant \eqn{d=1}, i.e. \eqn{cfD(t) = exp(1i*t)}.

\code{cfE_Empirical(t, data, randID, cfX)} evaluates the compound characteristic function
\eqn{cf(t) = cfE_EmpiricalBootstrapped(-1i*log(cfX(t)),data,randID) = (1/N) * sum_{j=1}^N cfX(t)^dataBoot_j};
where \code{cfX} is function handle of the characteristic function \eqn{cfX(t)} of the random variable \eqn{X}.
}
\note{
Ver.: 02-Sep-2022 17:21:10 (consistent with Matlab CharFunTool v1.5.1, 15-Sep-2018 11:47:29).
}
\examples{
#EXAMPLE 1 (Bootstrapped empirical CF based on given data)
set.seed(101)
n <- 1000
Data <- c(rnorm(3 * n, 5, 0.2), rt(n, 3),rchisq(n, df=1))

  dataN <- length(Data)
  randID <-pracma:: randi(dataN,dataN,1);
  t <- seq(-50, 50, length.out = 2^10)
  plotReIm(function(t) cfE_EmpiricalBootstrapped(t, Data,t(randID)),
           t, title = 'Bootstrapped empirical CF')


# EXAMPLE 2 (PDF/CDF of the compound bootstrapped empirical distributions)
  set.seed(101)
  lambda <- 25
  nN <- 100
  Ndata <- rpois(nN,lambda)
  mu <- 0.1
  sigma <- 2
  nX <- 15000
  Xdata <- rlnorm(nX,mu,sigma)
  randID <- vector()
  t   <-seq(-0.2,0.2,length.out = 2^10)
  cfX <-function(t) cfE_EmpiricalBootstrapped(t,Xdata,randID)
cf<-function(t) cfE_EmpiricalBootstrapped(t,Ndata,randID,cfX)

  plotReIm(function(t) cfE_EmpiricalBootstrapped(t,Ndata,randID,cfX),
           t, title = 'Bootstrapped empirical CF')
  x <- seq(0,1000,length.out = 501)
  prob <- c(0.9, 0.95)
  options <- list()
  options$N <- 2^12
  options$xMin <- 0
  options$SixSigmaRule <- 10
  result <- cf2DistGP(cf,x,prob,options)

# EXAMPLE 3 (PDF/CDF of the Stress-Strength reliability R = Pr(X<Y))
  set.seed(101)
  mu <- 0
  sigma <- 1
  n <- 50
  X <- rlnorm(n,mu,sigma)
  mu <- 2
  sigma <- 2
  n <- 20
  Y <- rlnorm(n,mu,sigma)
  nBoot <- 1000
  xCrit <- 0
  R     <- rep(0,nBoot)

  for (i in 1:nBoot){
      cfX <-function(t) cfE_EmpiricalBootstrapped(t,X)
      cfY <- function(t) cfE_EmpiricalBootstrapped(t,Y)
      cf  <- function(t) cfX(t) * cfY(-t)
      M <- cf2DistGP(cf,xCrit)
      R[i]<-M$cdf
  }
  bandwidth <- 0.02
  cf_KERNEL <- function(t) exp(-(bandwidth*t)^2/2)
  cfR <- function(t) cfE_Empirical(t,R) * cf_KERNEL(t)
  x <- seq(0,1,length.out=100)
  prob <-c(0.025, 0.5, 0.95, 0.975)
  options<-list()
  options$xMin <- 0
  options$xMax <- 1
  options$SixSigmaRule <- 10
  result <- cf2DistGP(cfR,x,prob,options)
}
\references{
WITKOVSKY V., WIMMER G., DUBY T. (2017). Computing the aggregate
loss distribution based on numerical inversion of the compound empirical
characteristic function of frequency and severity. arXiv preprint arXiv:1701.08299.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Empirical_distribution_function.}
\url{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}

Other Empirical Probability Distribution: 
\code{\link{cfE_DiracMixture}()},
\code{\link{cfE_EmpiricalOgive}()},
\code{\link{cfE_Empirical}()},
\code{\link{cfE_SampleMean}()}
}
\concept{Empirical Probability Distribution}
