% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfE_EmpiricalOgive.R
\name{cfE_EmpiricalOgive}
\alias{cfE_EmpiricalOgive}
\title{Characteristic function of the Ogive empirical distribution based on the observed histogram}
\usage{
cfE_EmpiricalOgive(t, data, counts)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{data}{vector of data (if the size is equal to the vector counts) or a vector of edges (if)the size is \eqn{n+1}, where \eqn{n} is the length of counts). If empty, default value is \deqn{data=1}.}

\item{counts}{vector of counts (integers) related to the values given by data. If empty, default value is \deqn{couns=1}.}
}
\value{
Characteristic function \eqn{cf(t)} of the Ogive empirical distribution, based on the observed data.
}
\description{
Characteristic function of the OGIVE EMPIRICAL distribution based on
the observed histogram given by data and counts (the vector lengths of
data and counts are equal) or by edges and counts (the length of the
edges is equal to the length of counts + 1).

That is, \eqn{cf(t)} is given as a weighted mixture of the RECTANGULAR
(continuous uniform) RVs,
\code{cf(t) = cfE_EmpiricalOgive(t,data,counts) = sum_{j=1}^n freq_j * cf_Rectangular(t,edges_{j},edges_{j+1})},
where \code{cf_Rectangular(t,edges_{j},edges_{j+1})} is CF of the RECTANGULAR
distribution on the interval \eqn{[edges_{j},edges_{j+1}]}. Here, the
frequencies are derived from the counts, \eqn{FREQ = COUNTS/SUM(COUNTS)}, and
the vector EDGES is derived automatically from the DATA (or is
specified by the user as an input vector EDGES of size n+1, instead of
DATA). In particular, for \deqn{j = 2,...,n} we set \eqn{EDGES(j) = (DATA(j-1) +
DATA(j))/2} and further,  \eqn{EDGES(1) = DATA(1) - (DATA(2)-DATA(1))/2} and
\code{EDGES(n+1) = DATA(n) + (DATA(n)-DATA(n-1))/2}.

In particular, \eqn{[COUNTS,EDGES] = HISTCOUNTS(DATA)} partitions the values
in DATA into bins, and returns the count in each bin, as well as the
bin edges. HISTCOUNTS uses an automatic binning algorithm that returns
bins with a uniform width, chosen to cover the range of elements in
DATA and reveal the underlying shape of the distribution. \eqn{COUNTS(k)}
will count the value \eqn{DATA(i)} if \deqn{EDGES(k) <= DATA(i) < EDGES(k+1)}. The
last bin will also include the right edge such that COUNTS(end) will
count \eqn{DATA(i)} if \deqn{EDGES(end-1) <= DATA(i) <= EDGES(end)}.
}
\note{
ver.:18-aug-2022 17:50:50  (consistent with Matlab CharFunTool, 8-May-2022 15:45:27).
}
\examples{
## EXAMPLE1 (Ogive ECF - a weighted mixture of independent Uniform RVs)
set.seed(101)
n <- 1000
Data <- c(rnorm(3 * n, 5, 0.2), rt(n, 10))
Edges <- -6:6
Counts <- hist(x = Data, breaks = Edges)$counts

t <- seq(-10, 10, length.out = 501)

plotReIm(function(t) cfE_EmpiricalOgive(t, Edges, Counts),
         t, title = 'Ogive ECF of the grouped (histogram) data')

## EXAMPLE2 (PDF/CDF of the Ogive distribution)
set.seed(101)
n <- 1000
Data <- c(rnorm(3 * n, 5, 0.2), rt(n, 10))
Edges <- -6:6
Counts <- hist(x = Data, breaks = Edges)$counts
cf <- function(t) cfE_EmpiricalOgive(t, Edges, Counts)
x <- seq(-30, 40, length.out = 101)
prob <- c(0.9, 0.95)
options <- list()
options$N <- 2^12
options$SixSigmaRule <- 8
result <- cf2DistGP(cf,x,prob,options)

## EXAMPLE3 (PDF/CDF of the compound Poisson-Ogive distribution)
set.seed(101)
n <- 1000
Data <- c(rnorm(3 * n, 5, 0.2), rt(n, 10))
Edges <- -6:6
Counts <- hist(x = Data, breaks = Edges)$counts
cfX <- function(t) cfE_EmpiricalOgive(t, Edges, Counts)
lambda <- 15
cf <- function(t) cfN_Poisson(t, lambda, cfX)
x <- seq(0, 100, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$isCompound <- TRUE
result <- cf2DistGP(cf, x, prob, options)
}
\references{
WITKOVSKY V., WIMMER G., DUBY T. (2017). Computing the aggregate
loss distribution based on numerical inversion of the compound empirical
characteristic function of frequency and severity. arXiv preprint arXiv:1701.08299.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Empirical_distribution_function}.
\url{https://en.wikipedia.org/wiki/Ogive_(statistics)}

Other Empirical Probability Distribution: 
\code{\link{cfE_DiracMixture}()},
\code{\link{cfE_EmpiricalBootstrapped}()},
\code{\link{cfE_Empirical}()},
\code{\link{cfE_SampleMean}()}
}
\concept{Empirical Probability Distribution}
