% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfE_SampleMean.R
\name{cfE_SampleMean}
\alias{cfE_SampleMean}
\title{Empirical characteristic function of the sample mean based on the observed random sample}
\usage{
cfE_SampleMean(t, data, counts, options)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{data}{vector of original data. If empty, default value is \code{data = 0}.}

\item{counts}{vector of the same size as data that represents the counts
of the repeated values. This is useful if the data are
discrete with many repeted values or if the data are to be
specified from a histogram. If counts is a scalar integer
value, each value specified by data is repeated equally
counts times. If empty, default value is \code{counts = 1}.}

\item{options}{structure with further optional specificaions:- \code{options.isOgive = false}.}
}
\value{
Characteristic function \eqn{cf(t)} of the empirical distribution of the
sample mean evalueted at the specified vector argument t.

structure with further useful details.
}
\description{
The empirical distribution of the sample mean is estimated by the
bootstrap distribution of the sample mean (i.e. the bootstrap mean
distribution).

  The characteristic function of the empirical distribution of the sample
  mean is derived as the characteristic function of an equally weighted
  linear combination (arithmetic mean) of n independent random variables
  each having distribution defined by the observed empirical cumulative
  distribution function \eqn{(ECDF)}, defined from the observed data. This can
  be expressed as a weighted mixture of the DIRAC distributions whose
  support is concentrated on the observed unique x-values and the weights
  are equal to the empirical probability mass function \eqn{(EPMF)} derived
  from the \eqn{ECDF}.

In fact, we define the characteristic function of the sample mean as \code{cf =cfE_DiracMixture(t/n,X,EPMF)^n)}, where
\eqn{X} is a vector of the observed unique x-values and \eqn{EPMF} is the vector of the weights.

The characteristic function of the sample mean can be used further,
e.g. to test hypotheses about the equality of the means (see the
Example section).
}
\note{
Ver.: 02-Sep-2022 17:53:40 (consistent with Matlab CharFunTool, 08-May-2022 17:20:15).
}
\examples{
##   EXAMPLE 1:
##Empirical CF of the sample mean
    data <- c(1, 1, 2, 2, 2, 3, 3, 9, 105, 105, 106, 106, 106, 107, 107)
    result <- cfE_SampleMean(c(),data)
    x  <- unique(result$X)
    cf  <- result$cf
    t   <- seq(-0.5,0.5,length.out=1001)
    plotReIm(function(t) cf(t),ylab = "cf(t)",
             t, title = 'Empirical CF of the Sample Mean')

##   EXAMPLE 2:
## Empirical CF and PMF/CDF of the sample mean
    data   <- c(-2, -1, 0, 1, 2, 3, 4)
    counts <- c( 2,  2, 3, 1, 2, 3, 2)
    resultCF <- cfE_SampleMean(c(),data,counts)
    x   <- unique(resultCF$X)
    cf  <- resultCF$cf
    t   <- seq(-100,100,length.out=1001)

    plotReIm(function(t) cf(t),
             t, title = 'Empirical CF of the Sample Mean')
    n <- resultCF$n
    delta <-1/n
    resultFFT <- cf2PMF_FFT(cf,-2,4,delta)
 ## Numerical inversion of the smoothed CF by cf2DistGP
    cfSmooth <- function(t) cf(t) * cf_Normal(t*delta)
    resultGP  <- cf2DistGP(cfSmooth)

##   EXAMPLE 3:
   ## Empirical distribution of the difference of the sample means
## Test of the hypothesis that the population means are equal
    dataA   <- c(1, 1, 2, 2, 2, 3, 3, 9, 105, 105, 106, 106, 106, 107, 107)
    dataB   <- c(5, 5, 6, 6, 6, 7, 7, 99, 101, 101, 102, 102, 102, 103, 103)
    cfA     <- function(t) cfE_SampleMean(t,dataA)
    cfB     <- function(t) cfE_SampleMean(t,dataB)
    cfDiff  <- function(t) cfA(t)$cf* cfB(-t)$cf
    minDiff <- min(dataA) - max(dataB)
    maxDiff <- max(dataA) - min(dataB)
    delta   <- 1/15
    result <- cf2PMF_FFT(cfDiff,minDiff,maxDiff,delta)

##   EXAMPLE 4:
## Empirical distribution of the difference of the sample means
# Test of the hypothesis that the population means are equal
    dataA   <- c(1, 1, 2, 2, 2, 3, 3, 9, 105, 105, 106, 106, 106, 107, 107)
    dataB  <- c(5, 5, 6, 6, 6, 7, 7, 99, 101, 101, 102, 102, 102, 103, 103)
    options<-list()
    options$isOgive <- TRUE
    cfA     <- function(t) cfE_SampleMean(t,dataA,c(),options)
    cfB     <- function(t) cfE_SampleMean(t,dataB,c(),options)
    cfDiff  = function(t) cfA(t)$cf * cfB(-t)$cf
    result <- cf2DistGP(cfDiff,c(),c(0.25,0.975))
}
\seealso{
Other Empirical Probability Distribution: 
\code{\link{cfE_DiracMixture}()},
\code{\link{cfE_EmpiricalBootstrapped}()},
\code{\link{cfE_EmpiricalOgive}()},
\code{\link{cfE_Empirical}()}
}
\concept{Empirical Probability Distribution}
