% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfN_Delaporte.R
\name{cfN_Delaporte}
\alias{cfN_Delaporte}
\title{Characteristic function of Delaporte distribution}
\usage{
cfN_Delaporte(t, a, b, c, cfX)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{a}{variable mean, \code{a > 0}.}

\item{b}{variable mean, \code{b > 0}.}

\item{c}{fixed mean, \code{c > 0}.}

\item{cfX}{function.}
}
\value{
Characteristic function \eqn{cf(t)} of the Delaporte distribution.
}
\description{
\code{cfN_Delaporte(t, a, b, c)} evaluates the characteristic function \eqn{cf(t)} of the
Delaporte distribution with the parameters \code{a} (parameter of variable mean, \code{a > 0})
\code{b} (parameter of variable mean, \code{b > 0} ), and \code{c} (fixed mean, \code{c > 0}), i.e.
\deqn{cfN_Delaporte(t, a, b, c) = (b/(1+b))^a * (1-e^(1i*t)/(b+1))^(-a) * exp(-c*(1-e^(1i*t))).}
For more details see [4].

\code{cfN_Delaporte(t, a, b, c, cfX)} evaluates the compound characteristic function
\deqn{cf(t) = cfN_Delaporte(-1i*log(cfX(t)), a, b, c),} where \code{cfX} is function
handle of the characteristic function \eqn{cfX(t)} of a continuous distribution
and/or random variable \eqn{X}.

Note that such CF is characteristic function of the compound distribution,
i.e. distribution of the random variable \eqn{Y = X_1 + ... + X_N}, where \eqn{X_i ~ F_X}
are i.i.d. random variables with common CF \eqn{cfX(t)}, and \eqn{N ~ F_N} is
independent RV with its CF given by \eqn{cfN(t)}.
}
\note{
Ver.: 16-Sep-2018 18:58:36 (consistent with Matlab CharFunTool v1.3.0, 15-Nov-2016 13:36:26).
}
\examples{
## EXAMPLE1
# CF of the Delaporte distribution with a = 2.2, b = 3.3, c = 4
a <- 2.2
b <- 3.3
c <- 4
t <- seq(-15, 15, length.out = 1001)
plotReIm(function(t)
        cfN_Delaporte(t, a, b, c),
        t,
        title = "CF of the Delaporte distribution with a=2.2, b=3.3, c=4")

## EXAMPLE2
# CF of the compound Delaport-Exponential distribution)
a <- 2.2
b <- 3.3
c <- 4
lambda <- 10
cfX <- function(t)
        cfX_Exponential(t, lambda)
t <- seq(-10, 10, length.out = 501)
plotReIm(function(t)
        cfN_Delaporte(t, a, b, c, cfX),
        t,
        title = "CF of the compound Delaport-Exponential distribution")

## EXAMPLE3
# PDF/CDF of the compound Delaport-Exponential distribution
a <- 2.2
b <- 3.3
c <- 4
lambda <- 5
cfX <- function(t)
        cfX_Exponential(t, lambda)
cf <- function(t)
        cfN_Delaporte(t, a, b, c, cfX)
x <- seq(0, 4, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$isCompound <- TRUE
result <- cf2DistGP(cf, x, prob, options)
}
\references{
[1] WITKOVSKY V., WIMMER G., DUBY T. (2016). Computing the aggregate loss
    distribution based on numerical inversion of the compound empirical
    characteristic function of frequency and severity. Preprint submitted
    to Insurance: Mathematics and Economics.

[2] DUBY T., WIMMER G., WITKOVSKY V.(2016). MATLAB toolbox CRM
    for computing distributions of collective risk models. Preprint submitted
    to Journal of Statistical Software.

[3] WITKOVSKY V. (2016). Numerical inversion of a characteristic function:
    An alternative tool to form the probability distribution
    of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.

[4] WIMMER G., ALTMANN G. (1999). Thesaurus of univariate discrete
    probability distributions. STAMM Verlag GmbH, Essen, Germany. ISBN 3-87773-025-6.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Delaporte_distribution}.

Other Discrete Probability Distribution: 
\code{\link{cfN_Binomial}()},
\code{\link{cfN_GeneralizedPoisson}()},
\code{\link{cfN_Geometric}()},
\code{\link{cfN_Logarithmic}()},
\code{\link{cfN_NegativeBinomial}()},
\code{\link{cfN_Poisson}()},
\code{\link{cfN_PolyaEggenberger}()},
\code{\link{cfN_Quinkert}()},
\code{\link{cfN_Waring}()}
}
\concept{Discrete Probability Distribution}
