% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfN_Geometric.R
\name{cfN_Geometric}
\alias{cfN_Geometric}
\title{Characteristic function of the Geometric distribution}
\usage{
cfN_Geometric(t, p = 1, type = "standard", cfX)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{p}{success probability, \eqn{0 \le} \code{p} \eqn{\le 1}, default value \code{p = 1}.}

\item{type}{\eqn{standard = 1}, \eqn{shifted = 2}, default \code{type = standard}.}

\item{cfX}{function.}
}
\value{
Characteristic function \eqn{cf(t)} of the Geometric distribution.
}
\description{
\code{cfN_Geometric(t, p, type, cfX)} evaluates the characteristic function \eqn{cf(t)} of the
Geometric distribution.

 The standard Geometric distribution (type = "standard" or "zero") is
 defined on non-negative integers \eqn{k = 0,1, \ldots} .

 The shifted Geometric distribution (type = "shifted")
 is defined on positive integers \eqn{k = 1,2, \ldots} .

Both types are parametrized by the success probability parameter \code{p} in \eqn{[0,1]}), i.e.
\eqn{cfN_Geometric(t, p, "standard") = p / (1 - (1-p) * exp(1i*t))},
\eqn{cfN_Geometric(t, p, "shifted")  = exp(1i*t) * (p / (1 - (1-p) * exp(1i*t)))}.
For more details see [4].

\code{cfN_Geometric(t, p, type, cfX)} evaluates the compound characteristic function
\deqn{cf(t) = Geometric(-1i*log(cfX(t)), p),}
where \code{cfX} is function
handle of the characteristic function cfX(t) of a continuous distribution
and/or random variable \eqn{X}.

Note that such CF is characteristic function of the compound distribution,
i.e. distribution of the random variable \eqn{Y = X_1 + ... + X_N}, where \eqn{X_i ~ F_X}
are i.i.d. random variables with common CF \eqn{cfX(t)}, and \eqn{N ~ F_N} is
independent RV with its CF given by \eqn{cfN(t)}.
}
\examples{
## EXAMPLE1
# CF of the Geometric distribution with the parameter p = 0.5
p <- 0.5
t <- seq(-10, 10, length.out = 501)
plotGraf(function(t)
        cfN_Geometric(t, p), t,
        title = "CF of the Geometric distribution with the parameter p = 0.5")

## EXAMPLE2
# CF of the Geometric distribution with the parameter p = 0.5, type = "shifted"
p <- 0.5
t <- seq(-10, 10, length.out = 501)
plotGraf(function(t)
        cfN_Geometric(t, p, "shifted"), t,
        title = "CF of the Geometric distribution with the parameter p = 0.5")

## EXAMPLE3
# CF of the compound Geometric-Exponential distribution
p <- 0.5
lambda <- 5
cfX <- function(t)
        cfX_Exponential(t, lambda)
t <- seq(-10, 10, length.out = 501)
plotGraf(function(t)
        cfN_Geometric(t, p, 1, cfX), t,
        title = "CF of the compound Geometric-Exponential distribution")

## EXAMPLE4
# PDF/CDF of the compound Geometric-Exponential distribution
p <- 0.5
lambda <- 5
cfX <- function(t)
        cfX_Exponential(t, lambda)
cf <- function(t)
        cfN_Geometric(t, p, cfX = cfX)
x <- seq(0, 4, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$isCompound <- TRUE
result <- cf2DistGP(cf, x, prob, options)
}
\references{
[1] WITKOVSKY V., WIMMER G., DUBY T. (2016). Computing the aggregate loss
    distribution based on numerical inversion of the compound empirical
    characteristic function of frequency and severity. Preprint submitted
    to Insurance: Mathematics and Economics.

[2] DUBY T., WIMMER G., WITKOVSKY V.(2016). MATLAB toolbox CRM
    for computing distributions of collective risk models. Preprint submitted
    to Journal of Statistical Software.

[3] WITKOVSKY V. (2016). Numerical inversion of a characteristic function:
    An alternative tool to form the probability distribution
    of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.

[4] WIMMER G., ALTMANN G. (1999). Thesaurus of univariate discrete
    probability distributions. STAMM Verlag GmbH, Essen, Germany. ISBN 3-87773-025-6.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Geometric_distribution}.

Other Discrete Probability Distribution: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_Poisson}},
  \code{\link{cfN_PolyaEggenberger}}
}
