% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfN_Poisson.R
\name{cfN_Poisson}
\alias{cfN_Poisson}
\title{Characteristic function of the Poisson distribution}
\usage{
cfN_Poisson(t, lambda = 1, cfX)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{lambda}{rate, \code{lambda > 0}, default value \code{lambda = 1}.}

\item{cfX}{function.}
}
\value{
Characteristic function \eqn{cf(t)} of the Poisson distribution.
}
\description{
\code{cfN_Poisson(t, lambda, cfX)} evaluates the characteristic function \eqn{cf(t)} of the
Poisson distribution with the rate parameter \code{lambda > 0}, i.e.
\deqn{cfN_Poisson(t, \lambda) = exp(\lambda*(exp(1i*t)-1))}.
For more details see [4].

\code{cfN_Poisson(t, lambda, cfX)} evaluates the compound characteristic function
\deqn{cf(t) = cfN_Poisson(-1i*log(cfX(t)), \lambda),} where \code{cfX} is function
handle of the characteristic function \eqn{cfX(t)} of a continuous distribution
and/or random variable \eqn{X}.

Note that such CF is characteristic function of the compound distribution,
i.e. distribution of the random variable \eqn{Y = X_1 + ... + X_N}, where \eqn{X_i ~ F_X}
are i.i.d. random variables with common CF \eqn{cfX(t)}, and \eqn{N ~ F_N} is
independent RV with its CF given by \eqn{cfN(t)}.
}
\examples{
## EXAMPLE1
# CF of the Poisson distribution with the parameter lambda = 10
lambda <- 10
t <- seq(-10, 10, length.out = 501)
plotReIm(function(t)
        cfN_Poisson(t, lambda), t,
        title = "CF of the Poisson distribution with the parameter lambda = 10")

## EXAMPLE2
# CF of the compound Poisson-Exponential distribution
lambda1 <- 10
lambda2 <- 5
cfX <- function(t)
        cfX_Exponential(t, lambda2)
t <- seq(-10, 10, length.out = 501)
plotReIm(function(t)
        cfN_Poisson(t, lambda1, cfX), t,
        title = "CF of the compound Poisson-Exponential distribution")

## EXAMPLE3
# PDF/CDF of the compound Poisson-Exponential distribution
lambda1 <- 10
lambda2 <- 5
cfX <- function(t)
        cfX_Exponential(t, lambda2)
cf <- function(t)
        cfN_Poisson(t, lambda1, cfX)
x <- seq(0, 8, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$isCompound <- TRUE
result <- cf2DistGP(cf, x, prob, options)
}
\references{
[1] WITKOVSKY V., WIMMER G., DUBY T. (2016). Computing the aggregate loss
    distribution based on numerical inversion of the compound empirical
    characteristic function of frequency and severity. Preprint submitted
    to Insurance: Mathematics and Economics.

[2] DUBY T., WIMMER G., WITKOVSKY V.(2016). MATLAB toolbox CRM
    for computing distributions of collective risk models. Preprint submitted
    to Journal of Statistical Software.

[3] WITKOVSKY V. (2016). Numerical inversion of a characteristic function:
    An alternative tool to form the probability distribution
    of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.

[4] WIMMER G., ALTMANN G. (1999). Thesaurus of univariate discrete
    probability distributions. STAMM Verlag GmbH, Essen, Germany. ISBN 3-87773-025-6.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Poisson_distribution}, \cr
\url{https://en.wikipedia.org/wiki/Compound_Poisson_distribution}.

Other Discrete Probability Distribution: \code{\link{cfN_Binomial}},
  \code{\link{cfN_Delaporte}},
  \code{\link{cfN_GeneralizedPoisson}},
  \code{\link{cfN_Geometric}},
  \code{\link{cfN_Logarithmic}},
  \code{\link{cfN_NegativeBinomial}},
  \code{\link{cfN_PolyaEggenberger}}
}
