% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfN_Quinkert.R
\name{cfN_Quinkert}
\alias{cfN_Quinkert}
\title{Characteristic function of the Quinkert distribution}
\usage{
cfN_Quinkert(t, a = 1, b = 1, cfX)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{a}{vector of the 'shape' parameters \code{a > 0}. If empty, default value is \code{a = 1}.}

\item{b}{vector of the 'shape' parameters \code{b > 0}. If empty, default value is \code{b = 1}.}

\item{cfX}{function.}
}
\value{
Characteristic function \eqn{cf(t)} of the Quinkert distribution.
}
\description{
\code{cfN_Quinkert(t,a,b)} evaluates the characteristic function \eqn{cf(t)}
of the Quinkert distribution, with the parameters \code{a} (a > 0) and \code{b} (b > 0), i.e.

 \eqn{cf(t) = cfN_Quinkert(t,a,b) = 1F1(a,a+b,e^(1i*t)-1)}, where \eqn{1F1} denotes the confluent hypergeometric (Kummer's) function.
  For more details see [4], p. 564.
}
\note{
ver.: 31-Jul-2021 12:36:48 (consistent with Matlab CharFunTool v1.5.1, 15-Nov-2016 13:36:26).
}
\examples{
## EXAMPLE 1
# CF of the Quinkert distribution with the parameter a=3, b=5
  a <- 3;
  b <- 5;
  t <- seq(from = -15,
           to = 15,
           length.out =501)
  plotReIm(function(t)
          cfN_Quinkert(t, a, b),
          t,
          title = "CF of the Quinkert distribution with a = 3, b = 5")

  ## EXAMPLE 2
  # CF of the compound Quinkert-Exponential distribution
  a <- 3
  b <- 5
  lambda <- 5
  cfX <- function(t)
          cfX_Exponential(t, lambda)
  t <- seq(from = -15,
           to = 15,
           length.out = 501)
  plotReIm(function(t)
          cfN_Quinkert(t, a, b, cfX),
          t,
          title = "CF of the compound Quinkert-Exponential distribution")

 ## EXAMPLE3
 # PDF/CDF of the compound Quinkert-Exponential distribution
    a <- 3;
    b <- 5;
    lambda <- 5;
    cfX <- function(t)
            cfX_Exponential(t, lambda)
    cf <- function(t)
            cfN_Quinkert(t,a,b,cfX)
    x <- seq(0,1.5, length.out = 101)
    prob <- c(0.9, 0.95, 0.99)
    options <- list()
    options$isCompound = TRUE
    result <- cf2DistGP(cf,x,prob,options)




}
\references{
[1] WITKOVSKY V., WIMMER G., DUBY T. (2016). Computing the aggregate loss
  distribution based on numerical inversion of the compound empirical
   characteristic function of frequency and severity. Preprint submitted
   to Insurance: Mathematics and Economics.

[2] DUBY T., WIMMER G., WITKOVSKY V.(2016). MATLAB toolbox CRM for
   computing distributions of collective risk models. Preprint submitted
   to Journal of Statistical Software.

[3] WITKOVSKY V. (2016). Numerical inversion of a characteristic
   function: An alternative tool to form the probability distribution of
   output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.

[4] WIMMER G., ALTMANN G. (1999). Thesaurus of univariate discrete
   probability distributions. STAMM Verlag GmbH, Essen, Germany. ISBN 3-87773-025-6.
}
\seealso{
Other Discrete Probability Distribution: 
\code{\link{cfN_Binomial}()},
\code{\link{cfN_Delaporte}()},
\code{\link{cfN_GeneralizedPoisson}()},
\code{\link{cfN_Geometric}()},
\code{\link{cfN_Logarithmic}()},
\code{\link{cfN_NegativeBinomial}()},
\code{\link{cfN_Poisson}()},
\code{\link{cfN_PolyaEggenberger}()},
\code{\link{cfN_Waring}()}
}
\concept{Discrete Probability Distribution}
