% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfS_Arcsine.R
\name{cfS_Arcsine}
\alias{cfS_Arcsine}
\title{Characteristic function of the symmetric zero-mean Arcsine distribution}
\usage{
cfS_Arcsine(t, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}
}
\value{
Characteristic function \eqn{cf(t)} of the Arcsine distribution.
}
\description{
\code{cfS_Arcsine(t, coef, niid)} evaluates the characteristic function \eqn{cf(t)} of
the zero-mean symmetric Arcsine distribution on the interval
\eqn{(-1,1)}.

\code{cfS_Arcsine} is an ALIAS of the more general function
\code{cf_ArcsineSymmetric}, used to evaluate the characteristic function of a
linear combination of independent ARCSINE distributed random variables.

The characteristic function of the symmetric ARCSINE distribution is \eqn{cf(t) = besselj(0,t)}.
}
\examples{
## EXAMPLE1 (
# CF of the symmetric Arcsine distribution on (-1,1)
t <- seq(-50, 50, length.out = 501)
plotGraf(function(t)
  cfS_Arcsine(t), t, title = "CF of the the Arcsine distribution on (-1,1)")

## EXAMPLE2
# PDF/CDF of the symmetric Arcsine distribution on (-1,1)
cf <- function(t)
  cfS_Arcsine(t)
x <- seq(-1, 1, length.out = 501)
xRange <- 2
options <- list()
options$N <- 2 ^ 12
options$dt <- 2 * pi / xRange
result <- cf2DistGP(cf = cf, x = x, options = options)
}
\references{
WITKOVSKY V. (2016). Numerical inversion of a characteristic
function: An alternative tool to form the probability distribution
of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Arcsine_distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Beta}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_Gamma}}, \code{\link{cf_InverseGamma}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}

Other Symetric Probability distribution: \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}}
}
