% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfS_Laplace.R
\name{cfS_Laplace}
\alias{cfS_Laplace}
\title{Characteristic function of a symmetric Laplace distribution
with scale parameter \eqn{beta>0}.}
\usage{
cfS_Laplace(t, beta, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{beta}{vector of the scale parameters \code{beta > 0}. If empty, default value is \code{beta = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the LAPLACE random variables.
If \code{coef} is scalar, it is assumed that all coefficients are equal. If empty, default value is coef = 1.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y} is sum of \eqn{niid} iid random variables \eqn{Y},
where each \eqn{Y = sum_{i=1}^N coef(i) * X_i} is independently and identically distributed random variable.
If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of of a linear
combination of independent (symmetric) LAPLACE sistributed random variables.
}
\description{
\code{cfS_Laplace(t, beta, coef, niid)} evaluates the characteristic function of a linear
combination of independent (symmetric) Laplace distributed random variables.

That is, \code{cfS_Laplace} evaluates the characteristic function
\eqn{cf(t)} of \eqn{Y=sum_{i=1}^N coef_i * X_i}, where \eqn{X_i ~ Laplace (0,beta_i)}
are independent zero-mean RVs with the scale parameters
\eqn{beta_i > 0}, for \eqn{i = 1,...,N}
}
\note{
Ver.: 10-Aug-2021 17:21:39 (consistent with Matlab CharFunTool v1.5.1, 16-Aug-2018 16:00:43).
}
\examples{
## EXAMPLE 1
# CF of the linear combination of symmetric Laplace RV
beta <- 1
t <- seq(from = -10,
         to = 10,
         length.out =201)
plotReIm(function(t)
        cfS_Laplace(t, beta),
        t,
        title = "Characteristic function of the symmetric Laplace RV")



##EXAMPLE2
# PDF/CDF of the symmetric Laplace RVs
beta <- 1
x <- seq(-5, 5, length.out = 101)
prop <- c(0.80, 0.85, 0.90, 0.925, 0.95, 0.975, 0.99, 0.995, 0.999)
cf <- function(t)
        cfS_Laplace(t, beta)
result <- cf2DistGP(cf, x, prob)

##EXAMPLE3
# PDF/CDF of the linear combination of symmetric Laplace RVs
beta <- c(0.1, 0.2, 0.3, 0.4)
coef <- c(1, 2, 3, 4)
prop <- c(0.80, 0.85, 0.90, 0.925, 0.95, 0.975, 0.99, 0.995, 0.999)
cf <- function(t)
        cfS_Laplace(t, beta, coef)
options <- list()
options$N <- 2^12
result <- cf2DistGP(cf, c(), prob, options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Normal_distribution}.

Other Continuous Probability Distribution: 
\code{\link{cfS_Arcsine}()},
\code{\link{cfS_Beta}()},
\code{\link{cfS_Gaussian}()},
\code{\link{cfS_Rectangular}()},
\code{\link{cfS_Student}()},
\code{\link{cfS_TSP}()},
\code{\link{cfS_Trapezoidal}()},
\code{\link{cfS_Triangular}()},
\code{\link{cfS_Wigner}()},
\code{\link{cfX_ChiSquare}()},
\code{\link{cfX_Exponential}()},
\code{\link{cfX_FisherSnedecor}()},
\code{\link{cfX_Gamma}()},
\code{\link{cfX_InverseGamma}()},
\code{\link{cfX_LogNormal}()},
\code{\link{cf_ArcsineSymmetric}()},
\code{\link{cf_BetaNC}()},
\code{\link{cf_BetaSymmetric}()},
\code{\link{cf_Beta}()},
\code{\link{cf_ChiSquare}()},
\code{\link{cf_Exponential}()},
\code{\link{cf_FisherSnedecorNC}()},
\code{\link{cf_FisherSnedecor}()},
\code{\link{cf_Gamma}()},
\code{\link{cf_InverseGamma}()},
\code{\link{cf_Laplace}()},
\code{\link{cf_LogRV_BetaNC}()},
\code{\link{cf_LogRV_Beta}()},
\code{\link{cf_LogRV_ChiSquareNC}()},
\code{\link{cf_LogRV_ChiSquare}()},
\code{\link{cf_LogRV_FisherSnedecorNC}()},
\code{\link{cf_LogRV_FisherSnedecor}()},
\code{\link{cf_LogRV_MeansRatioW}()},
\code{\link{cf_LogRV_MeansRatio}()},
\code{\link{cf_LogRV_WilksLambdaNC}()},
\code{\link{cf_LogRV_WilksLambda}()},
\code{\link{cf_Normal}()},
\code{\link{cf_RectangularSymmetric}()},
\code{\link{cf_Student}()},
\code{\link{cf_TSPSymmetric}()},
\code{\link{cf_TrapezoidalSymmetric}()},
\code{\link{cf_TriangularSymmetric}()},
\code{\link{cf_vonMises}()}

Other Symmetric Probability Distribution: 
\code{\link{cfS_Arcsine}()},
\code{\link{cfS_Beta}()},
\code{\link{cfS_Gaussian}()},
\code{\link{cfS_Rectangular}()},
\code{\link{cfS_Student}()},
\code{\link{cfS_Trapezoidal}()},
\code{\link{cf_ArcsineSymmetric}()},
\code{\link{cf_BetaSymmetric}()},
\code{\link{cf_RectangularSymmetric}()},
\code{\link{cf_TSPSymmetric}()},
\code{\link{cf_TrapezoidalSymmetric}()}
}
\concept{Continuous Probability Distribution}
\concept{Symmetric Probability Distribution}
