% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfS_Student.R
\name{cfS_Student}
\alias{cfS_Student}
\title{Characteristic function of the STUDENT's t-distribution}
\usage{
cfS_Student(t, df = 1, mu, sigma, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{df}{the degrees of freedom, \code{df > 0}. If empty, the default value is \code{df = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of the STUDENT's t-distribution.
}
\description{
\code{cfS_Student(t, df, mu, sigma, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
of the STUDENT's t-distribution with \code{df > 0} degrees of freedom.

\code{cfS_Student} is an ALIAS of the more general function \code{cf_Student}, used
to evaluate the characteristic function of a linear combination
of independent (location-scale) STUDENT's t-distributed random variables.

The characteristic function of the STUDENT's t-distribution with \eqn{df} degrees of freedom is defined by
\deqn{cf(t) = cfS_Student(t,df) = besselk(df/2,abs(t)*sqrt(df),1) * exp(-abs(t)*sqrt(df)) * (sqrt(df)*abs(t))^(df/2) / 2^(df/2-1)/gamma(df/2).}
}
\note{
bessel function nefunguje dobre
}
\examples{
## EXAMPLE1
# CF of the Student t-distribution with df = 3
df <- 3
t <- seq(-5, 5, length.out = 501)
plotGraf(function(t)
  cfS_Student(t, df), t, title = "CF of the Student t-distribution with df = 3")

## EXAMPLE2
# PDF/CDF of the Student t-distribution with df = 3
df <- 3
cf <- function(t)
  cfS_Student(t, df)
x <- seq(-8, 8, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$N <- 2^12
options$SixSigmaRule <- 30
result <- cf2DistGP(cf, x, prob, options)
}
\references{
WITKOVSKY V. (2016). Numerical inversion of a characteristic
function: An alternative tool to form the probability distribution
of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Student's_t-distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PearsonV}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_Gamma}}, \code{\link{cf_InverseGamma}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}

Other Symetric Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}}
}
