% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfS_TSP.R
\name{cfS_TSP}
\alias{cfS_TSP}
\title{Characteristic function of the TWO-SIDED-POWER (TSP) distribution with shape parameter \code{theta > 0}.}
\usage{
cfS_TSP(t, theta, mu, sigma, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{theta}{the shape parameter, \code{theta > 0}. If empty, default value is \code{theta = 1}.}

\item{mu}{vector of location parameters, mu in Real. If empty, default value is \code{mu = 0}.}

\item{sigma}{vector of scale parameters, \code{sigma_i > 0}. If empty, default value is \code{sigma = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the log-transformed random variables. If \eqn{coef} is scalar, it is
assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient n, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)} is independently
and identically distributed random variable. If empty, default value is \code{niid = 1}.}
}
\value{
Characteristic function of a linear combination  of independent (location scale TSP distributed random variables.
}
\description{
\code{cfS_TSP(t, theta, mu, sigma, coef, niid)} is an ALIAS of the more general function \code{cf_TSPSymmetric}, used to evaluate the characteristic function of a linear combination of
independent (location-scale) TSP distributed random variables.

 The characteristic function of the random variable \eqn{X ~ TSP(theta)} is \eqn{cf(t) = 1/2 * (hypergeom1F1(1,1+theta,1i*t) + hypergeom1F1(1,1+theta,-1i*t))}.

#' SPECIAL CASES: \cr
1) \eqn{\theta = 1/2}; Arcsine distribution on \eqn{(-1,1)}:     \eqn{cf(t) = besselj(0,t)}, \cr
2) \eqn{\theta = 1};   Rectangular distribution on \eqn{(-1,1)}: \eqn{cf(t) = sin(t)/t}.
}
\note{
Ver.: 11-Aug-2021 16:37:22 (consistent with Matlab CharFunTool v1.5.1, 24-Jun-2017 18:25:56).
}
\examples{
## EXAMPLE 1
# CF of the symmetric TSP distribution with theta =3/2 on (-1,1)
theta <- 3/2
t <- seq(from = -50,
         to = 50,
         length.out =501)
plotReIm(function(t)
        cfS_TSP(t, theta),
        t,
        title = "CF of the symmetric TSP distribution  on (-1,1)")



##EXAMPLE2
# PDF/CDF of the symmetric TSP distribution on (-1,1)
thet <- 3/2
cf <- function(t)
        cfS_TSP(t, theta)
x <- seq(-1,1,length.out = 101)
xRange <- 2
options <- list()
options.N <-2^8
options.dt <- 2*pi/xRange
result <- cf2DistGP(cf, x, c(), options)
}
\references{
[1] WITKOVSKY V. (2016). Numerical inversion of a characteristic
function: An alternative tool to form the probability distribution of
output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.
[2] VAN DORP, R.J., KOTZ, S. (2003). Generalizations of two-sided power
distributions and their convolution. Communications in
Statistics-Theory and Methods, 32(9), 1703-1723.
}
\seealso{
Other Continuous Probability Distribution: 
\code{\link{cfS_Arcsine}()},
\code{\link{cfS_Beta}()},
\code{\link{cfS_Gaussian}()},
\code{\link{cfS_Laplace}()},
\code{\link{cfS_Rectangular}()},
\code{\link{cfS_Student}()},
\code{\link{cfS_Trapezoidal}()},
\code{\link{cfS_Triangular}()},
\code{\link{cfS_Wigner}()},
\code{\link{cfX_ChiSquare}()},
\code{\link{cfX_Exponential}()},
\code{\link{cfX_FisherSnedecor}()},
\code{\link{cfX_Gamma}()},
\code{\link{cfX_InverseGamma}()},
\code{\link{cfX_LogNormal}()},
\code{\link{cf_ArcsineSymmetric}()},
\code{\link{cf_BetaNC}()},
\code{\link{cf_BetaSymmetric}()},
\code{\link{cf_Beta}()},
\code{\link{cf_ChiSquare}()},
\code{\link{cf_Exponential}()},
\code{\link{cf_FisherSnedecorNC}()},
\code{\link{cf_FisherSnedecor}()},
\code{\link{cf_Gamma}()},
\code{\link{cf_InverseGamma}()},
\code{\link{cf_Laplace}()},
\code{\link{cf_LogRV_BetaNC}()},
\code{\link{cf_LogRV_Beta}()},
\code{\link{cf_LogRV_ChiSquareNC}()},
\code{\link{cf_LogRV_ChiSquare}()},
\code{\link{cf_LogRV_FisherSnedecorNC}()},
\code{\link{cf_LogRV_FisherSnedecor}()},
\code{\link{cf_LogRV_MeansRatioW}()},
\code{\link{cf_LogRV_MeansRatio}()},
\code{\link{cf_LogRV_WilksLambdaNC}()},
\code{\link{cf_LogRV_WilksLambda}()},
\code{\link{cf_Normal}()},
\code{\link{cf_RectangularSymmetric}()},
\code{\link{cf_Student}()},
\code{\link{cf_TSPSymmetric}()},
\code{\link{cf_TrapezoidalSymmetric}()},
\code{\link{cf_TriangularSymmetric}()},
\code{\link{cf_vonMises}()}
}
\concept{Continuous Probability Distribution}
