% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Beta.R
\name{cfX_Beta}
\alias{cfX_Beta}
\title{Characteristic function of the BETA distribution}
\usage{
cfX_Beta(t, alpha = 1, beta = 1)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{alpha}{shape, \code{alpha > 0}, default value \code{alpha = 1}.}

\item{beta}{shape, \code{beta > 0}, default value \code{beta = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of the BETA distribution.
}
\description{
\code{cfX_Beta(t, alpha, beta)} evaluates the characteristic function \eqn{cf(t)} of
the Beta distribution with the parameter \code{alpha} (shape, \eqn{\alpha > 0}) and \code{beta} (shape, \eqn{\beta > 0})
defined on the interval \eqn{(0,1)}, i.e. beta distribution with the \eqn{Mean = \alpha / (\alpha + \beta)}
and the \eqn{Variance = (\alpha*\beta) / ((\alpha+\beta)^2*(\alpha+\beta+1))}.
Then, the standard deviation is given by \eqn{STD = sqrt(Variance)}
i.e.
\deqn{cf(t) = cfX_Beta(t,\alpha,\beta) = 1F1(\alpha ,\alpha + \beta , i*t),}
where \eqn{1F1(.;.;.)} is the Confluent hypergeometric function.
}
\examples{
## EXAMPLE 1
# CF of the Beta distribution with alpha = 1/2, beta = 3/2
alpha <- 1 / 2
beta <- 3 / 2
t <- seq(-50, 50, length.out = 501)
plotGraf(function(t)
        cfX_Beta(t, alpha, beta),
        t,
        title = "CF of the Beta distribution with alpha = 1/2, beta = 3/2")

## EXAMPLE 2
# PDF/CDF of the Beta distribution with alpha = 1/2, beta = 3/2
alpha <- 1 / 2
beta <- 3 / 2
cf <- function(t)
        cfX_Beta(t, alpha, beta)
x <- seq(0, 1, length.out = 101)
xRange <- 1
option <- list()
option$dx <- 2 * pi / xRange
result <- cf2DistGP(cf, x, option = option)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Beta_distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_Gamma}}, \code{\link{cf_InverseGamma}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}
}
