% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_ChiSquare.R
\name{cfX_ChiSquare}
\alias{cfX_ChiSquare}
\title{Characteristic function of the CHI-SQUARE distribution}
\usage{
cfX_ChiSquare(t, df, ncp, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{df}{the degrees of freedom parameter \code{df > 0}. If empty, default value is \code{df = 1}.}

\item{npc}{the non-centrality parameter \code{ncp > 0}. If empty, default value is \code{ncp = 0}.}
}
\value{
Characteristic function \eqn{cf(t)} of the CHI-SUQARE distribution.
}
\description{
\code{cfX_ChiSquare(t, df, ncp, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
of the CHI-SQUARE distribution with \code{df > 0} degrees of freedom and the non-cetrality parameter \code{ncp > 0}.

\code{cfX_ChiSquare} is an ALIAS NAME of the more general function
\code{cf_ChiSquare}, used to evaluate the characteristic function of a linear combination
of independent CHI-SQUARE distributed random variables.

The characteristic function of the CHI-SQUARE distribution is defined
by \deqn{cf(t) = (1 - 2*i*t )^(df/2) * exp((i*t*ncp)/(1-2*i*t)).}
}
\examples{
## EXAMPLE 1
# CF of the ChiSquared distribution with df = 1
df <- 1
t <- seq(from = -50,to = 50,length.out = 501)
plotGraf(function(t)
  cfX_ChiSquare(t,df), t, title = "CF of the Chi-square distribution with df = 1")

## EXAMPLE 2
# PDF/CDF of the ChiSquare distribution with df = 3
df <- 3
x <- seq(from = 0,to = 15,length.out = 101)
prob <- c(0.9, 0.95, 0.99)
cf <- function(t) cfX_ChiSquare(t,df)
options <- list()
options$xMin <- 0
options$N <- 2^14
result <- cf2DistGP(cf,x,prob,options)

## EXAMPLE 3
# PDF/CDF of the compound Binomial-ChiSquared distribution
n <- 25
p <- 0.3
df <- 3
cfX <- function(t) cfX_ChiSquare(t,df)
cf <- function(t) cfN_Binomial(t,n,p,cfX)
x <- seq(from = 0,to = 80,length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$isCompound <- TRUE
result <- cf2DistGP(cf,x,prob,options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Chi-squared_distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_Gamma}}, \code{\link{cf_InverseGamma}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}
}
