% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_FisherSnedecor.R
\name{cfX_FisherSnedecor}
\alias{cfX_FisherSnedecor}
\title{Characteristic function of the central FISHER-SNEDECOR F-distribution}
\usage{
cfX_FisherSnedecor(t, df1, df2, coef, niid, tol)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{df1}{vector of the  degrees of freedom \code{df1 > 0}. If empty, default value is \code{df1 = 1}.}

\item{df2}{vector of the  degrees of freedom \code{df2 > 0}. If empty, default value is \code{df2 = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the log-transformed random variables.
If coef is scalar, it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}

\item{tol}{tolerance factor for selecting the Poisson weights, i.e. such that \eqn{PoissProb > tol}.
If empty, default value is \code{tol = 1e-12}.}
}
\value{
Characteristic function \eqn{cf(t)} of the central FISHER-SNEDECOR F-distribution.
}
\description{
\code{cfX_FisherSnedecor(t, df1, df2, coef, niid, tol)} evaluates characteristic function
of the central FISHER-SNEDECOR F-distribution with \eqn{df1 > 0} and \eqn{df2 > 0} degrees of freedom.

\code{cfX_FisherSnedecor} is an ALIAS NAME of the more general function
\code{cf_FisherSnedecor}, used to evaluate the characteristic function
of a linear combination of independent FISHER-SNEDECOR F-distributed random variables.

The characteristic function of \eqn{X ~ F(df1,df2)} is defined
by \eqn{cf(t) = U(df1/2, 1-df2/2, -1i*(df2/df1)*t)},
where \eqn{U(a,b,z)} denotes the confluent hypergeometric function of the second kind.
}
\note{
Ver.: 06-Oct-2018 17:40:03 (consistent with Matlab CharFunTool v1.3.0, 24-Jun-2017 10:07:43).
}
\examples{
## EXAMPLE 1
# CF of the F-distribution with df1 = 3, df2 = 5
df1 <- 3
df2 <- 5
t <- seq(-30, 30, length.out = 2^10+1)
plotReIm(function(t) cfX_FisherSnedecor(t, df1, df2), t,
         title = 'Characteristic function of the F-distribution')

## EXAMPLE 2
# PDF/CDF of the F-distribution with df1 = 3, df2 = 5
df1 <- 3
df2 <- 5
x <- seq(0, 25, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
cf <- function(t) cfX_FisherSnedecor(t, df1, df2)
options <- list()
options$xMin <- 0
options$xMax <- 500
options$N <- 2^15
result <- cf2DistGP(cf, x, prob, options)

## EXAMPLE 3
# PDF/CDF of the compound Binomial-Fisher-Snedecor distribution
n <- 25
p <- 0.3
df1 <- 3
df2 <- 5
cfX <- function(t) cfX_FisherSnedecor(t, df1, df2)
cf <- function(t) cfN_Binomial(t, n, p, cfX)
x <- seq(0, 80, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$isCompound <- TRUE
result <- cf2DistGP(cf, x, prob, options)
str(result)
}
\references{
[1] PHILLIPS, P.C.B. The true characteristic function of the F distribution. Biometrika (1982), 261-264.

[2] WITKOVSKY, V.: On the exact computation of the density and of the quantiles of linear combinations
of t and F random variables. Journal of Statistical Planning and Inference 94 (2001), 1-13.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/F-distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_LogRV_MeansRatioW}},
  \code{\link{cf_LogRV_MeansRatio}},
  \code{\link{cf_LogRV_WilksLambdaNC}},
  \code{\link{cf_LogRV_WilksLambda}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
