% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Gamma.R
\name{cfX_Gamma}
\alias{cfX_Gamma}
\title{Characteristic function of the GAMMA distribution}
\usage{
cfX_Gamma(t, alpha, beta, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{alpha}{the shape parameter \code{alpha > 0}. If empty, default value is \code{alpha = 1}.}

\item{beta}{the rate (\eqn{1/scale}) parameter \code{beta > 0}. If empty, default value is \code{beta = 1}.}

\item{coef}{vector of coefficients of the linear combination of Gamma distributed random variables.
If coef is scalar, it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient.}
}
\value{
Characteristic function \eqn{cf(t)} of the GAMMA distribution.
}
\description{
\code{cfX_Gamma(t, alpha, beta, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
of the GAMMA distribution with the shape parameter \code{alpha > 0} and the rate parameter \code{beta > 0}.

\code{cfX_Gamma} is an ALIAS NAME of the more general function \code{cf_Gamma},
used to evaluate the characteristic function of a linear combination
of independent GAMMA distributed random variables.

The characteristic function of the GAMMA distribution is defined by
\deqn{cf(t) = (1 - i*t/\beta)^(-\alpha).}
}
\note{
Ver.: 16-Sep-2018 19:22:15 (consistent with Matlab CharFunTool v1.3.0, 24-Jun-2017 10:07:43).
}
\examples{
## EXAMPLE 1
# CF of the Gamma distribution with alpha = 2, beta = 2
alpha <- 2
beta <- 2
t <- seq(-20, 20, length.out = 501)
plotReIm(function(t)
        cfX_Gamma(t, alpha, beta),
        t,
        title = "CF of the Gamma distribution with alpha = 2, beta = 2")

## EXAMPLE 2
# PDF/CDF of the Gamma distribution with alpha = 2, beta = 2
alpha <- 2
beta <- 2
x <- seq(from = 0,
         to = 5,
         length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$xMin <- 0
options$N <- 2 ^ 14
cf <- function(t)
        cfX_Gamma(t, alpha, beta)
result <- cf2DistGP(cf, x, prob, options)

## EXAMPLE 3
# PDF/CDF of the compound Binomial-Gamma distribution
n = 25
p = 0.3
alpha <- 2
beta <- 2
cfX <- function(t)
        cfX_Gamma(t, alpha, beta)
cf <- function(t)
        cfN_Binomial(t, n, p, cfX)
x <- seq(from = 0,
         to = 25,
         length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$isCompound <- TRUE
result <- cf2DistGP(cf, x, prob, options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Gamma_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}},
  \code{\link{cfX_FisherSnedecor}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_LogRV_MeansRatioW}},
  \code{\link{cf_LogRV_MeansRatio}},
  \code{\link{cf_LogRV_WilksLambdaNC}},
  \code{\link{cf_LogRV_WilksLambda}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
