% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_LogNormal.R
\name{cfX_LogNormal}
\alias{cfX_LogNormal}
\title{Characteristic function of Lognormal distribution}
\usage{
cfX_LogNormal(t, mu = 0, sigma = 1, tol = 1e-06)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{mu}{real, default value \code{mu = 0}.}

\item{sigma}{> 0, default value \code{sigma = 1}.}

\item{tol}{tolerance, default value \code{tol = 1e-6}.}
}
\value{
Characteristic function \eqn{cf(t)} of the Lognormal distribution.
}
\description{
\code{cfX_LogNormal(t, mu, sigma, tol)} computes the characteristic function \eqn{cf(t)}
of the Lognormal distribution with parameters \code{mu} (real) and \code{sigma > 0}, computed for real (vector) argument t, i.e.
\deqn{cf(t) = cfX_LogNormal(t,mu,sigma).}
}
\details{
In probability theory, a log-normal (or lognormal) distribution is a
continuous probability distribution of a random variable whose logarithm
is normally distributed. The lognormal distribution is defined for \eqn{x} in
\eqn{(0,+inf)} by its PDF/CDF/CF, as follows
 \eqn{pdf(x) = 1/(x*\sigma*sqrt(2*pi))*exp(-(ln(x)-\mu)^2/(2*\sigma^2))
 cdf(x) = 1/2+1/2*erf((ln(x)-\mu)/(sqrt(2)*\sigma))
 cf(t)  = sum_0^infinity{(it)^n/n!*exp(n*\mu + (n*\sigma)^2/2)}.}
As noted, this representation is asymptotically divergent but sufficient
for numerical purposes.

\code{cfX_LogNormal} is based on the standard integral representation of the
characteristic function of the lognormal distribution, i.e.
 \eqn{cf(t) = Integral_0^inf exp(i*t*x) * PDF(x) dx}.
By using the half-space Fourier integral transformation we get
 \eqn{cf(t) = Integral_0^inf (i/t) * exp(-x) * PDF(i*x/t) dx}.
If we define the integrand as \eqn{funCF(t,x) = (i/t) * exp(-x) * PDF(i*x/t)},
then by using a stabilizing transformation from \eqn{[0,inf]} to \eqn{[0,1]}, we can
evaluate the CF by the following (well behaved) integral:
 \eqn{cf(t) = Integral_0^1 2x/(1-x)^3 * funCF(t,(x/(1-x))^2) dx}.

\code{cfX_LogNormal} evaluates this integral by using the R built in
function \code{integrate()}, with precission specified by tolerance \code{tol} (default
value is \eqn{tol = 1e-6}).
}
\examples{
## EXAMPLE1
# CF of the Lognormal distribution with mu = 0,sigma = 1
mu <- 0
sigma <- 1
t <- seq(-20, 20, length.out = 2 ^ 10 + 1)
plotGraf(function(t)
        cfX_LogNormal(t, mu, sigma),
        t,
        title = "Characteristic function of the Lognormal distribution")

## EXAMPLE2
# CDF/PDF of the Lognormal distribution with mu = 0,sigma = 1
mu <- 0
sigma <- 1
x <- seq(0, 15, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
cf <- function(t)
        cfX_LogNormal(t, mu, sigma)
options <- list()
options$xMin <- 0
options$N <- 2 ^ 13
options$SixSigmaRule <- 8
result <- cf2DistGP(cf, x, prob, options)

## EXAMPLE3
# PDF/CDF of the compound Poisson-Lognormal distribution
mu <- 0
sigma <- 1
lambda <- 10
x <- seq(0, 70, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
cfX <- function(t)
        cfX_LogNormal(t, mu, sigma)
cf <- function(t)
        cfN_Poisson(t, lambda, cfX)
options <- list()
options$isCompound <- TRUE
result <- cf2DistGP(cf, x, prob, options)
}
\references{
[1] WITKOVSKY, V.: On the exact computation of the density and
of the quantiles of linear combinations of t and F random variables.
Journal of Statistical Planning and Inference 94 (2001), 1-13.

[2] WITKOVSKY V. (2016). Numerical inversion of a characteristic function:
An alternative tool to form the probability distribution of output quantity
in linear measurement models. Acta IMEKO, 5(3), 32-44.

[3] WITKOVSKY V., WIMMER G., DUBY T. (2016). Computing the aggregate loss distribution
based on numerical inversion of the compound empirical characteristic function
of frequency and severity. Working Paper. Insurance: Mathematics and Economics.

[4] DUBY T., WIMMER G., WITKOVSKY V.(2016). MATLAB toolbox CRM for computing distributions
of collective risk models.  Working Paper. Journal of Statistical Software.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Log-normal_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
