% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Rectangular.R
\name{cfX_Rectangular}
\alias{cfX_Rectangular}
\title{Characteristic function of Rectangular distribution}
\usage{
cfX_Rectangular(t, a = -1, b = 1)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{a}{number, default value \eqn{a = -1}.}

\item{b}{number, default value \eqn{b = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of the Rectangular distribution.
}
\description{
\code{cfX_Rectangular(t, a, b)} evaluates the characteristic function \eqn{cf(t)} of
the Rectangular distribution on the interval \eqn{(}\code{a}, \code{b}\eqn{)}
(Rectangular distribution with \eqn{mean = (a + b)/2} and \eqn{variance = 1/12(b - a)^2)}
\eqn{cfX_Rectangular(t, a, b) = (exp(ibt) - exp(iat))/(i(b - a)t)}.
}
\examples{
## EXAMPLE1 (CF of the Rectangular distribution on (-2,1))
t <- seq(-50, 50, length.out = 501)
plotReIm(function(t)
        cfX_Rectangular(t, a = -2, b = 1),
        t,
        title = "CF of the Rectangular distribution on (-2,1)")

## EXAMPLE2 (PDF/CDF of the Rectangular distribution on (-2,1))
cf <- function(t)
        cfX_Rectangular(t, a = -2, b = 1)
x <- seq(-2, 1, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
xRange <- 3
options <- list()
options$N <- 2 ^ 10
options$dx <- 2 / pi / xRange
result <- cf2DistGP(cf, x, prob, options)

## EXAMPLE3 (PDF/CDF of the Rectangular distribution on (-2,1))
cf <- function(t)
        cfX_Rectangular(t, a = -2, b = 1)
x <- seq(-2, 1, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$xMin <- -2
options$xMax <- 1
options$N <- 2
result <- cf2DistGP(cf = cf, options = options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Normal_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PDF}}, \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
