% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Triangular.R
\name{cfX_Triangular}
\alias{cfX_Triangular}
\title{Characteristic function of Triangular distribution}
\usage{
cfX_Triangular(t, a = -1, b = 1, c = 0)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{a}{number, default value \eqn{a = -1}.}

\item{b}{number, default value \eqn{b = 1}.}

\item{c}{number, (\eqn{a \le c \le b}), default value \eqn{c = 0}}
}
\value{
Characteristic function \eqn{cf(t)} of the Triangular distribution.
}
\description{
\code{cfX_Triangular(t, a, b, c)} (\code{a} \eqn{\le} \code{c} \eqn{\le} \code{b})
evaluates the characteristic function \eqn{cf(t)} of the Triangular distribution
on the interval \eqn{(}\code{a}, \code{b}\eqn{)} with mode \code{c}
(Triangular distribution with \eqn{mean = a + b + c)/3} and \eqn{variance = 1/18(a^2 + b^2 + c^2 + - ab - ac - bc)}).
\deqn{cfX_Triangula(t, a, b, c) = -2((b-c)exp(iat) - (b-a)exp(ict) + (c-a)exp(ibt))/((b-a)(c-a)(b-c)t^2).}
}
\examples{
## EXAMPLE1 (CF of the Triangular distribution on (-1 , 2) mode = 0)
t <- seq(-10, 10, length.out = 501)
plotReIm(function(t)
        cfX_Triangular(t, a = -1, b = 2, c = 0), t,
        title = "CF of the Triangular distribution on (-1 , 2)")

## EXAMPLE2 (PDF/CDF of the Triangular distribution on (-1 , 2) mode = 0)
cf <- function(t)
        cfX_Triangular(t, a = -1, b = 2, c = 0)
x <- seq(-1, 2, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
xRange <- 3
options <- list()
options$N <- 2 ^ 10
options$dx <- 2 / pi / xRange
result <- cf2DistGP(cf = cf, x = x, options = options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Triangular_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PDF}}, \code{\link{cfX_Rectangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}
}
