% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChebPoly.R
\name{ChebPoly}
\alias{ChebPoly}
\title{Evaluates the Chebyshev nth order polynomial of the first kind}
\usage{
ChebPoly(n, x)
}
\arguments{
\item{n}{order of polynomial.}

\item{x}{points, in which we want to compute the value of polynomial.}
}
\value{
Function returns values of nth order Chebysev polynomial of the first kind evaluated in points \code{x}.
}
\description{
\code{ChebPoly(n, x)} evaluates the Chebyshev nth order polynomial of the first kind
\eqn{T_n(x)}, where \eqn{T_n(x) = cos(n*acos(x))}.
}
\note{
Ver.: 01-Oct-2018 12:33:49 (consistent with Matlab CharFunTool v1.3.0, 07-Sep-2017 16:20:17).
}
\examples{
## EXAMPLE 1
n <- 0:5
x <- ChebPoints(32, c(-1, 1))
pval <- ChebPoly(n, x)
matplot(x, t(pval), main = 'Chebyshev polynomials T_n(x) evaluated at Chebyshev points', type = "b")
}
\seealso{
For more details see also \url{https://en.wikipedia.org/wiki/Chebyshev_polynomials}.

Other Utility Function: \code{\link{ChebPoints}},
  \code{\link{GammaLog}}, \code{\link{GammaMultiLog}},
  \code{\link{GammaMulti}}, \code{\link{GammaZX}},
  \code{\link{Hypergeom1F1MatApprox}},
  \code{\link{Hypergeom1F1Mat}},
  \code{\link{Hypergeom2F1Mat}},
  \code{\link{Hypergeom2F1}},
  \code{\link{HypergeompFqMat}},
  \code{\link{hypergeom1F1}},
  \code{\link{interpBarycentric}}
}
