% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaMulti.R
\name{GammaMulti}
\alias{GammaMulti}
\title{Evaluates the multivariate gamma function of order p}
\usage{
GammaMulti(z, p)
}
\arguments{
\item{z}{complex argument, of any size (vector, matrix).}

\item{p}{order of the multivariate gamma, if empty, default value is \code{p = 1}.}
}
\value{
Function returns values of the multivariate gamma function of order \code{p} evaluated in points \code{z}.
}
\description{
\code{GammaMulti(z, p)} evaluates the multivariate gamma function of order \code{p},
the argument \code{z} may be complex and of any size.
}
\note{
Ver.: 01-Oct-2018 13:41:36 (consistent with Matlab CharFunTool v1.3.0, 17-Aug-2018 19:45:37).
}
\examples{
## EXAMPLE 1
t <- seq(-10, 10, length.out = 101)
z <- 1i * t
p <- 10
f <- GammaMulti(z, p)
}
\seealso{
Other Utility Function: \code{\link{ChebPoints}},
  \code{\link{ChebPoly}}, \code{\link{GammaLog}},
  \code{\link{GammaMultiLog}}, \code{\link{GammaZX}},
  \code{\link{Hypergeom1F1MatApprox}},
  \code{\link{Hypergeom1F1Mat}},
  \code{\link{Hypergeom2F1Mat}},
  \code{\link{Hypergeom2F1}},
  \code{\link{HypergeompFqMat}},
  \code{\link{hypergeom1F1}},
  \code{\link{interpBarycentric}}
}
