% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaMultiLog.R
\name{GammaMultiLog}
\alias{GammaMultiLog}
\title{Evaluates logarithm of the multivariate gamma function of order p}
\usage{
GammaMultiLog(z, p, funmode)
}
\arguments{
\item{z}{complex argument, of any size (vector, matrix).}

\item{p}{order of the multivariate gamma, if empty, default value is \code{p = 1}.}

\item{funmode}{function mode: If \code{funmode = 0} \eqn{f = log[gamma_p(z))}. If \code{funmode = 1} \eqn{f = gamma_p(z)}.}
}
\value{
Function returns values of the logarithm of multivariate gamma function of order \code{p} evaluated in points \code{z}.
}
\description{
\code{GammaMultiLog(z, p, funmode)} evaluates logarithm of the multivariate gamma function
of order \code{p}, the argument \code{z} may be complex and of any size.

Multivariate gamma function of order \eqn{p}, say \eqn{gamma_p(z)},
is defined as \eqn{gamma_p(z) = pi^(p*(p-1)/2) * prod_{j=1}^p gamma(z-(j-1)/2)}.
Hence, the logarithm of multivariate gamma
is given by \eqn{log(gamma_p(z))=(p*(p-1)/2)*log(pi) + sum_{j=1}^p log(gamma(z-(j-1)/2))}.
}
\note{
Ver.: 18-Sep-2019 17:14:16 (consistent with Matlab CharFunTool v1.3.0, 17-Aug-2018 19:45:37).
}
\examples{
## EXAMPLE 1
t <- seq(-10, 10, length.out = 101)
z <- 1i * t
p <- 10
f <- GammaMultiLog(z, p)

## EXAMPLE 2
t <- seq(-10, 10, length.out = 101)
z <- 1i * t
p <- 10
funmode <- 1
f <- GammaMultiLog(z, p, funmode)
}
\seealso{
Other Utility Function: 
\code{\link{ChebCoefficients}()},
\code{\link{ChebPoints}()},
\code{\link{ChebPolyValues}()},
\code{\link{ChebPoly}()},
\code{\link{ChebValues}()},
\code{\link{GammaLog}()},
\code{\link{GammaMulti}()},
\code{\link{GammaZX}()},
\code{\link{Hypergeom1F1MatApprox}()},
\code{\link{Hypergeom1F1Mat}()},
\code{\link{Hypergeom2F1Mat}()},
\code{\link{Hypergeom2F1}()},
\code{\link{HypergeompFqMat}()},
\code{\link{InterpChebValues}()},
\code{\link{hypergeom1F1}()},
\code{\link{interpBarycentric}()}
}
\concept{Utility Function}
