% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeom1F1.R
\name{hypergeom1F1}
\alias{hypergeom1F1}
\title{Confluent hypergeometric function}
\usage{
hypergeom1F1(a, b, z, n)
}
\arguments{
\item{a}{parameter \code{a} (real scalar).}

\item{b}{parameter \code{b} (real scalar).}

\item{z}{complex argument (scalar, vector or array).}

\item{n}{number of the Gauss-Laguerre nodes and weights on \eqn{(0,Inf)}. If empty, default value is \code{n = 64}.}
}
\value{
List including the following items:

\item{f}{calculated \eqn{1F1(a,b,z)} of the same dimension as \eqn{z},}
\item{method}{indicator of the used method: \cr
             -1 is the method used for special cases,
              0 is the method used for z == 0, \cr
              1 is the method based on the series expansion, \cr
              2 is the method based on the steepest descent integration, \cr
              3 is the method based on the asymptotic expansion, }
\item{loops}{indicator of the used number of recursive loops.}
}
\description{
\code{hypergeom1F1(a, b, z)} computes the confluent hypergeometric function \eqn{1F1(a, b, z)}
also known as the Kummer's function \eqn{M(a, b, z)}, for the parameters \code{a} and \code{b}
(here assumed to be real scalars), and the complex argument \code{z} (scalar, vector or array).
}
\details{
The present algorithm was adapted from the MATLAB version of the FORTRAN program
suggested by S.Zhang and J.Jin (1996). For more details see also
\url{http://iris-lee3.ece.uiuc.edu/~jjin/routines/routines.html}.

Computation of \eqn{1F1(a, b, z)} for large arguments \eqn{z} (with \eqn{|Im(z)| >= |Re(z)|}) and for \eqn{b > a > 0})
is based on using the steepest descent integration method as suggested by G. Navas Palencia and A.A. Arratia Quesada (2016).
}
\examples{
## EXAMPLE 1
a <- 10
b <- 15
z <- c(0, 1, 10i, 50i, 10 + 50i, 100 + 50i)
n <- 64
result <- hypergeom1F1(a, b, z, n)

## EXAMPLE 2
# CF of Beta(1/2,1/2) distribution
a  = 1 / 2
b  = 1 / 2
t  = seq(-100, 100, length.out = 1001)
plotGraf(
        function(t)
                hypergeom1F1(a, a + b, 1i * t)$f,
        t,
        title = "Characteristic function of Beta(1/2,1/2) distribution",
        xlab = "t",
        ylab = "CF"
)
}
\references{
[1] Jin, J. M., and Zhang Shan Jjie. Computation of Special Functions. Wiley, 1996.

[2] Navas Palencia, G. and Arratia Quesada, A.A., 2016. On the computation of confluent
hypergeometric functions for large imaginary part of parameters b and z.
In Mathematical Software - ICMS 2016: 5th International Conference, Berlin, Germany, July 11-14, 2016:
proceedings, pp. 241-248. Springer.
}
\seealso{
For more details on confluent hypergeometric function \eqn{1F1(a, b, z)}
or the Kummer's (confluent hypergeometric) function \eqn{M(a, b, z)} see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Confluent_hypergeometric_function}.

Other Special Function: \code{\link{ChebPoints}},
  \code{\link{GammaLog}}, \code{\link{interpBarycentric}}
}
