% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypergeom1F1Mat.R
\name{Hypergeom1F1Mat}
\alias{Hypergeom1F1Mat}
\title{The confluent hypergeometric function of matrix argument}
\usage{
Hypergeom1F1Mat(a, b, X, MAX)
}
\arguments{
\item{a}{complex vector of parameters of the hypergeometric function \eqn{1F1^alpha(a;b;X)}.}

\item{b}{complex vector of parameters  of the hypergeometric function \eqn{1F1^alpha(a;b;X)}.}

\item{X}{real symmetric \eqn{(p x p)}-matrix argument (alternatively can be specified
as a \eqn{(p x p)}-diagonal matrix or a \eqn{p}-vector of the eigenvalues of \eqn{X}).}

\item{MAX}{maximum number of partitions, \eqn{|\kappa| <= MAX}, default value is \code{MAX = 20}.}
}
\value{
Hypergeometric sum, \eqn{1F1(a;b;X)}.
}
\description{
\code{Hypergeom1F1Mat(a, b, X, MAX)} computes the confluent hypergeometric function \eqn{1F1(a;b;X)}
of a \eqn{(p x p)}-matrix argument \eqn{X}. \code{Hypergeom1F1Mat} is defined for the complex parameters
\code{a} and \code{b}, with \eqn{Re(a) > (p-1)/2} and  \eqn{Re(b-a) > (p-1)/2}, and a REAL symmetric matrix argument \code{X}.

For more details and definition of the hypergeometric functions with matrix argument see, e.g.,
Koev and Edelman (2006) or Muirhead (2009).
}
\note{
Ver.: 18-Oct-2018 13:16:53 (consistent with Matlab CharFunTool v1.3.0, 25-Oct-2017 14:56:37).
}
\examples{
## EXAMPLE
a <- 3
b <- 5
X <- c(1, 2, 3)
MAX <- 10
f <- Hypergeom1F1Mat(a, b, X, MAX)

}
\references{
[1] Koev, P. and Edelman, A., 2006. The efficient evaluation of the hypergeometric function of a matrix argument.
\emph{Mathematics of Computation}, 75(254), 833-846.

[2] Muirhead RJ. Aspects of multivariate statistical theory. John Wiley & Sons; 2009 Sep 25.

[3] Butler RW, Wood AT. Laplace approximations for hypergeometric functions with matrix argument.
\emph{The Annals of Statistics}. 2002;30(4):1155-77.
}
\seealso{
Other Utility Function: 
\code{\link{ChebCoefficients}()},
\code{\link{ChebPoints}()},
\code{\link{ChebPolyValues}()},
\code{\link{ChebPoly}()},
\code{\link{ChebValues}()},
\code{\link{GammaLog}()},
\code{\link{GammaMultiLog}()},
\code{\link{GammaMulti}()},
\code{\link{GammaZX}()},
\code{\link{Hypergeom1F1MatApprox}()},
\code{\link{Hypergeom2F1Mat}()},
\code{\link{Hypergeom2F1}()},
\code{\link{HypergeompFqMat}()},
\code{\link{InterpChebValues}()},
\code{\link{hypergeom1F1}()},
\code{\link{interpBarycentric}()}
}
\concept{Utility Function}
