% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypergeom2F1.R
\name{Hypergeom2F1}
\alias{Hypergeom2F1}
\title{Gauss hypergeometric function 2F1(a,b,c,z)}
\usage{
Hypergeom2F1(a, b, c, z)
}
\arguments{
\item{a}{parameter - real scalar.}

\item{b}{parameter - real scalar.}

\item{c}{parameter - real scalar.}

\item{z}{complex argument - scalar, vector or array.}
}
\value{
The Gauss hypergeometric function \eqn{2F1(a,b,c,z)}.
}
\description{
\code{Hypergeom2F1(a, b, c, z)} computes the Gauss hypergeometric function \eqn{2F1(a,b,c,z)},
for the real parameters \code{a}, \code{b} and \code{c} (here assumed to be scalars),
and the complex argument \code{z} (could be scalar, vector or array).
}
\details{
The functions based on \code{Hypergeom2F1} is badly conditioned for when \eqn{c} is negative integer.
In such situations, approximate the function value by using the noninteger parameter \eqn{c},
say \eqn{c = c + eps}, for some small \eqn{eps}.
}
\note{
Ver.: 06-Oct-2018 18:36:05 (consistent with Matlab CharFunTool v1.3.0, 18-Aug-2018 18:32:27).
}
\examples{
## EXAMPLE 1
a <- 3
b <- 2.5
c <- 1.5
z <- 1i * seq(0, 1, by = 0.05)
f <- Hypergeom2F1(a, b, c, z)

# EXAMPLE 2
t <- 1i * seq(-5, 5, length.out = 11)
a <- 3 * t
b <- 2.5 * t
c <- 1.5 * t
z <- 0.75
f <- Hypergeom2F1(a,b,c,z)
}
\references{
The algorithm is based on a Fortran program in

S. Zhang & J. Jin "Computation of Special Functions" (Wiley, 1996).
Converted by Ben Barrowes (barrowes@alum.mit.edu).
}
\seealso{
Other Utility Function: 
\code{\link{ChebCoefficients}()},
\code{\link{ChebPoints}()},
\code{\link{ChebPolyValues}()},
\code{\link{ChebPoly}()},
\code{\link{ChebValues}()},
\code{\link{GammaLog}()},
\code{\link{GammaMultiLog}()},
\code{\link{GammaMulti}()},
\code{\link{GammaZX}()},
\code{\link{Hypergeom1F1MatApprox}()},
\code{\link{Hypergeom1F1Mat}()},
\code{\link{Hypergeom2F1Mat}()},
\code{\link{HypergeompFqMat}()},
\code{\link{InterpChebValues}()},
\code{\link{hypergeom1F1}()},
\code{\link{interpBarycentric}()}
}
\concept{Utility Function}
