% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypergeom2F1Mat.R
\name{Hypergeom2F1Mat}
\alias{Hypergeom2F1Mat}
\title{The Gauss hypergeometric function of matrix argument X}
\usage{
Hypergeom2F1Mat(a, b, c, X, MAX)
}
\arguments{
\item{a}{complex vector of parameters of the hypergeometric function \eqn{2F1(a,b;c;X)}.}

\item{b}{complex vector of parameters of the hypergeometric function \eqn{2F1(a,b;c;X)}.}

\item{c}{complex vector of parameters of the hypergeometric function \eqn{2F1(a,b;c;X)}.}

\item{X}{real symmetric \eqn{(p x p)}-matrix argument (alternatively can be specified
as a \eqn{(p x p)}-diagonal matrix or a \eqn{p}-vector of the eigenvalues of \eqn{X}).}

\item{MAX}{maximum number of partitions, \eqn{|\kappa| <= MAX}, default value is \code{MAX = 20}.}
}
\value{
Hypergeometric sum, \eqn{2F1(a,b;c;X)}.
}
\description{
\code{Hypergeom2F1Mat(a, b, c, X, MAX)} computes the The Gauss hypergeometric function \eqn{2F1(a,b;c;X)}
of a \eqn{(p x p)}-matrix argument \eqn{X}. \code{Hypergeom2F1Mat} is defined for the complex
parameters \code{a}, \code{b}, and \code{c} with \eqn{Re(a) > (p-1)/2} and  \eqn{Re(c-a) > (p-1)/2},
and a real symmetric matrix argument \code{X}, with \eqn{Re(X) < I}.

For more details and definition of the hypergeometric functions with matrix argument see, e.g.,
Koev and Edelman (2006) or Muirhead (2009).
}
\note{
Ver.: 18-Oct-2018 13:52:32 (consistent with Matlab CharFunTool v1.3.0, 22-Aug-2018 12:23:08).
}
\examples{
## EXAMPLE
a <- 3
b <- 2.5
c <- 1.5
# X <- c(1, 2, 3) / 5
X <- t(c(1, 2, 3) / 5)
MAX <- 50
f <- Hypergeom2F1Mat(a, b, c, X, MAX)

# a <- 3
# b <- c(1,2,3,4,5)
# c <- c(5,4,3,2,1)
# X <-  t(c(1, 2, 3)) / 5
# MAX <- 10
# f <- Hypergeom2F1Mat(a, b, c, X, MAX)

}
\references{
[1] Koev, P. and Edelman, A., 2006. The efficient evaluation of the hypergeometric function of a matrix argument.
\emph{Mathematics of Computation}, 75(254), 833-846.

[2] Muirhead RJ. Aspects of multivariate statistical theory. John Wiley & Sons; 2009 Sep 25.

[3] Butler RW, Wood AT. Laplace approximations for hypergeometric functions with matrix argument.
\emph{The Annals of Statistics}. 2002;30(4):1155-77.
}
\seealso{
Other Utility Function: 
\code{\link{ChebCoefficients}()},
\code{\link{ChebPoints}()},
\code{\link{ChebPolyValues}()},
\code{\link{ChebPoly}()},
\code{\link{ChebValues}()},
\code{\link{GammaLog}()},
\code{\link{GammaMultiLog}()},
\code{\link{GammaMulti}()},
\code{\link{GammaZX}()},
\code{\link{Hypergeom1F1MatApprox}()},
\code{\link{Hypergeom1F1Mat}()},
\code{\link{Hypergeom2F1}()},
\code{\link{HypergeompFqMat}()},
\code{\link{InterpChebValues}()},
\code{\link{hypergeom1F1}()},
\code{\link{interpBarycentric}()}
}
\concept{Utility Function}
