% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPolar.R
\name{plotPolar}
\alias{plotPolar}
\title{Plot a line in polar coordinates}
\usage{
plotPolar(angle, radius, title = "", type = "l", color = "blue")
}
\arguments{
\item{angle}{angle.}

\item{radius}{amplitude.}

\item{title}{an overall title for the plot.}

\item{type}{type of lines, dots.}

\item{color}{color.}
}
\description{
\code{plotPolar(angle, radius, title, type, color)} plots a line in polar coordinates,
 with \code{angle} indicating the angle in radians and \code{radius} indicating the radius value for each point.
}
\seealso{
Other Graphic Function: \code{\link{plotGraf}},
  \code{\link{plotReIm}}
}
