% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReIm2.R
\name{plotReIm2}
\alias{plotReIm2}
\title{Plot real and imaginary part of f(x)}
\usage{
plotReIm2(f, x, title = "", xmin, xmax, dx = 0.1, xlab = "t",
  ylab = "f(t)", px = 0, py = 0, labelx = TRUE, labely = TRUE)
}
\arguments{
\item{f}{list of functions.}

\item{x}{list of numerical values (numerical vectors).}

\item{title}{an overall title for the plot.}

\item{xmin}{vector of numbers, default values \code{xmin = -1}.}

\item{xmax}{vector of numbers, default values \code{xmax = 1}.}

\item{dx}{positive number, default value \code{dx = 0.1}.}

\item{xlab}{a title for the \eqn{x} axis: see \link[graphics]{title}.}

\item{ylab}{a title for the \eqn{y} axis: see \link[graphics]{title}.}

\item{px}{numerical values (number or vector) in which are auxiliary lines on axis \eqn{x} rendered, default value \code{px = 0}.}

\item{py}{numerical values (number or vector) in which are auxiliary lines on axis \eqn{y} rendered, default value \code{py = 0}.}

\item{labelx}{rendering auxiliary line, default value \code{TRUE}.}

\item{labely}{rendering auxiliary line, default value \code{TRUE}.}
}
\value{
Graf of function \code{f} in argument \code{x} or in range \code{[xmin, xmax]} with step \code{dx}.
}
\description{
\code{plotReIm2(f, x)} plots values of complex functions \code{f} in real arguments \code{x} or in ranges \code{[xmin, xmax]} with steps \code{dx}.
}
\details{
Function with optional parameters must be pass as lambda function \code{f = function(x) f(x, optional parameters)}.
}
\note{
Ver.: 22-Sep-2018 20:41:29.
}
\seealso{
Other Graphic Function: \code{\link{plotPolar}},
  \code{\link{plotReIm}}
}
