% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Severity.R
\docType{data}
\name{Severity}
\alias{Severity}
\title{The Danish Insurance Data}
\format{
A data frame with one column containing fire losses over the period 1980 to 1990.
}
\source{
\href{http://www.macs.hw.ac.uk/~mcneil/data.html}{SeverityData}
}
\usage{
data(Severity)
}
\description{
These data were collected at Copenhagen Reinsurance and comprise 2167 fire losses over the period 1980 to 1990.
}
\references{
Resnick, S. I. (1997). Discussions of the Danish Data on Large Fire Insurance Losses.
ASTIN Bulletin: The Journal of the IAA, 27(1), 139-151.

Cooray, K., & Ananda, M. M. A. (2005). Modeling actuarial data with a composite lognormal-Pareto model.
Scandinavian Actuarial Journal, 2005(5), 321-334.

Dell'Aquila, R., & Embrechts, P. (2006). Extremes and Robustness:
A Contradiction? Financial Markets and Portfolio Management, 20(1), 103-118.

Diethelm Wuertz, Tobias Setz and Yohan Chalabi (2017). fExtremes: Rmetrics - Modelling Extreme
Events in Finance. R package version 3042.82. \url{https://CRAN.R-project.org/package=fExtremes}
}
\keyword{datasets}
