% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_LogRV_FisherSnedecor.R
\name{cf_LogRV_FisherSnedecor}
\alias{cf_LogRV_FisherSnedecor}
\title{Characteristic function of a linear combination
of independent  LOG-TRANSFORMED FISHER-SNEDECOR F random variables}
\usage{
cf_LogRV_FisherSnedecor(t, df1, df2, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{df1}{vector of the  degrees of freedom \code{df1 > 0}. If empty, default value is \code{df1 = 1}.}

\item{df2}{vector of the  degrees of freedom \code{df2 > 0}. If empty, default value is \code{df2 = 1}.}

\item{coef}{vector of the coefficients of the linear combination of the Beta distributed random variables.
If coef is scalar, it is assumed that all coefficients are equal. If empty, default value is \code{coef = 1}.}

\item{niid}{scalar convolution coeficient \code{niid}, such that \eqn{Z = Y + ... + Y}
is sum of \eqn{niid} iid random variables \eqn{Y}, where each \eqn{Y = sum_{i=1}^N coef(i) * log(X_i)}
is independently and identically distributed random variable. If empty, default value is \code{niid = 1}.}

\item{tol}{tolerance factor for selecting the Poisson weights, i.e. such that \eqn{PoissProb > tol}.
If empty, default value is \code{tol = 1e-12}.}
}
\value{
Characteristic function \eqn{cf(t)} of a linear combination
of independent  LOG-TRANSFORMED FISHER-SNEDECOR F random variables.
}
\description{
\code{cf_LogRV_FisherSnedecor(t, df1, df2, coef, niid)} evaluates characteristic function
of a linear combination (resp. convolution) of independent  LOG-TRANSFORMED FISHER-SNEDECOR F
random variables (RVs) \eqn{log(X)}, where \eqn{X ~ F(df1,df2)} has the FISHER-SNEDECOR F distribution
with \eqn{df1 > 0} and \eqn{df2 > 0} degrees of freedom.

That is, \code{cf_LogRV_FisherSnedecor} evaluates the characteristic function
\eqn{cf(t) of  Y = coef_1*log(X_1) +...+ coef_N*log(X_N)}, where X_i ~ F(df1_i,df2_i),
with degrees of freedom \eqn{df1_i} and \eqn{df2_i}, for \eqn{i = 1,...,N}.

The characteristic function of \eqn{Y = log(X)}, with \eqn{X ~ F(df1,df2,\lambda)},
where \eqn{\lambda} is the non-centrality parameter, is defined by
\eqn{cf_Y(t) = E(exp(1i*t*Y)) = E(exp(1i*t*log(X))) = E(X^(1i*t))}.
That is, the characteristic function can be derived from expression
for the r-th moment of \eqn{X}, \eqn{E(X^r)} by using \eqn{(1i*t)} instead of \eqn{r}.
In particular, the characteristic function of \eqn{Y = log(X)} is defined by
\deqn{cf_Y(t) = (df2/df1)^(1i*t) * gamma(df1/2 + 1i*t) / gamma(df1/2) * gamma(df2/2 - 1i*t) / gamma(df2/2) * 1F1(-1i*t;df1/2;-\lambda/2),}
where \eqn{1F1(a;b;z)} is the hypergeometric function confluent hypergeometric function,
also known as the Kummer function \eqn{M(a,b,z)}.

Hence,the characteristic function of \eqn{Y  = coef(1)*Y1 + ... + coef(N)*YN}
is  \deqn{cf_Y(t) =  cf_Y1(coef(1)*t) * ... * cf_YN(coef(N)*t),} where \eqn{cf_Yi(t)}
is evaluated with the parameters \eqn{df1(i)}, \eqn{df2(i)}, and \eqn{\lambda(i)}.
}
\examples{
## EXAMPLE 1
# CF of a weighted linear combination of independent log-F RVs
coef <- c(1, 2, 3, 4, 5)
weight <- coef / sum(coef)
df1 <- 5
df2 <- 3
t <- seq(from = -10, to = 10, length.out = 201)
plotGraf(function(t)
  cf_LogRV_FisherSnedecor(t, df1, df2, weight), t, title = "Characteristic function of a linear combination of log-F RVs")

## EXAMPLE 2
# PDF/CDF from the CF by cf2DistGP
coef <- c(1, 2, 3, 4, 5)
weight <- coef / sum(coef)
df1 <- 5
df2 <- 3
cf <- function(t) cf_LogRV_FisherSnedecor(t, df1, df2, weight)
options <- list()
options$N <- 2^12
prob <- c(0.9, 0.95, 0.99)
result <- cf2DistGP(cf = cf, prob = prob, options = options)
str(result)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/F-distribution}.

Other Continuous Probability Distribution: \code{\link{cf_BetaNC}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}}
}
