% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfS_Gaussian.R
\name{cfS_Gaussian}
\alias{cfS_Gaussian}
\title{Characteristic function of a GAUSSIAN (standard normal) distribution}
\usage{
cfS_Gaussian(t, mu, sigma, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}
}
\value{
Characteristic function \eqn{cf(t)} of a GAUSSIAN (standard normal) distribution.
}
\description{
\code{cfS_Gaussian(t, mu, sigma, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
of a GAUSSIAN (standard normal) distribution.

cfS_Gaussian is an ALIAS of the more general function cf_Normal, used
to evaluate the characteristic function of a linear combination
of independent normally distributed random variables.

The characteristic function of the standard normally distributed random variable, \eqn{X ~ N(0,1)},
is defined by \deqn{cf(t) = cfS_Gaussian(t) = exp(-t^2/2).}
}
\examples{
## EXAMPLE1
# CF of the Gaussian distribution N(0,1)
t <- seq(-5, 5, length.out = 501)
plotGraf(function(t)
        cfS_Gaussian(t), t, title = "CF of the Gaussian distribution N(0,1)")

## EXAMPLE2
# PDF/CDF of the Gaussian distribution N(0,1)
cf <- function(t)
        cfS_Gaussian(t)
x <- seq(-4, 4, length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$N <- 2 ^ 5
options$SixSigmaRule <- 8
result <- cf2DistGP(cf, x, prob, options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Normal_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Student}}, \code{\link{cfS_Trapezoidal}},
  \code{\link{cfS_Triangular}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}

Other Symmetric Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Rectangular}},
  \code{\link{cfS_Student}}, \code{\link{cfS_Trapezoidal}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_TrapezoidalSymmetric}}
}
