% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfS_Rectangular.R
\name{cfS_Rectangular}
\alias{cfS_Rectangular}
\title{Characteristic function of the zero-mean symmetric RECTANGULAR distribution}
\usage{
cfS_Rectangular(t, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}
}
\value{
Characteristic function \eqn{cf(t)} of the zero-mean symmetric RECTANGULAR distribution.
}
\description{
\code{cfS_Rectangular(t, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
of the zero-mean symmetric RECTANGULAR distribution defined on the interval \eqn{(-1,1)}.

\code{cfS_Rectangular} is an ALIAS of the more general function
\code{cf_RectangularSymmetric}, used to evaluate the characteristic function
of a linear combination of independent RECTANGULAR distributed random variables.

The characteristic function of \eqn{X ~ RectangularSymmetric} is defined by
\deqn{cf(t) = sinc(t) = sin(t)/t.}
}
\examples{
## EXAMPLE1
# CF of the Rectangular distribution on (-1,1)
t <- seq(-50, 50, length.out = 501)
plotGraf(function(t)
  cfS_Rectangular(t), t, title = "CF of the Rectangular distribution on (-1,1)")

## EXAMPLE2
# PDF/CDF of the Rectangular distribution on (-1,1)
cf <- function(t)
  cfS_Rectangular(t)
x <- seq(-2, 2, length.out = 101)
xRange <- 2
options <- list()
options$N <- 2 ^ 5
options$dx <- 2 / pi / xRange
result <- cf2DistGP(cf = cf, x = x, options = options)
}
\references{
WITKOVSKY V. (2016). Numerical inversion of a characteristic
function: An alternative tool to form the probability distribution
of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Uniform_distribution_(continuous)}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Student}}, \code{\link{cfS_Trapezoidal}},
  \code{\link{cfX_Beta}}, \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_Gamma}}, \code{\link{cf_InverseGamma}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}
}
