% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfS_Trapezoidal.R
\name{cfS_Trapezoidal}
\alias{cfS_Trapezoidal}
\title{Characteristic function of the zero-mean symmetric TRAPEZOIDAL distribution}
\usage{
cfS_Trapezoidal(t, lambda, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{lambda}{parameter of the offset, \eqn{0 \le} \code{lambda} \eqn{\le 1}.
If empty, default value is \code{lambda = 0}.}
}
\value{
Characteristic function \eqn{cf(t)} of the zero-mean symmetric TRAPEZOIDAL distribution.
}
\description{
\code{cfS_Trapezoidal(t, lambda, coef, niid)} evaluates
the characteristic function of the zero-mean symmetric TRAPEZOIDAL distribution defined on the interval \eqn{(-1,1)}.

\code{cfS_Trapezoidal} is an ALIAS of the more  general function
\code{cf_TrapezoidalSymmetric}, used to evaluate the characteristic function
of a linear combination of independent TRAPEZOIDAL distributed random variables.

The characteristic function of \eqn{X ~ TrapezoidalSymmetric(\lambda)}, where
\eqn{0\le \lambda \le 1} is the offset parameter is defined by
\deqn{cf(t) = (sin(w*t)/(w*t))*(sin((1-w)*t)/((1-w)*t))},
where \eqn{w  = (1+\lambda)/2}.
}
\examples{
## EXAMPLE1
# CF of the symmetric Trapezoidal distribution, lambda = 0.5
lambda <- 0.5
t <- seq(-50, 50, length.out = 501)
plotReIm(function(t)
        cfS_Trapezoidal(t, lambda), t,
        title = "CF of the symmetric Trapezoidal distribution with lambda = 0.5")

## EXAMPLE2
# PDF/CDF of the symmetric Trapezoidal distribution, lambda = 0.5
lambda <- 0.5
cf <- function(t)
        cfS_Trapezoidal(t, lambda)
x <- seq(-1, 1, length.out = 100)
xRange <- 2
options <- list()
options$N <- 2 ^ 10
options$dx <- 2 / pi / xRange
result <- cf2DistGP(cf = cf, x = x, options = options)
}
\references{
WITKOVSKY V. (2016). Numerical inversion of a characteristic function:
An alternative tool to form the probability distribution
of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Trapezoidal_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Triangular}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}

Other Symmetric Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_TrapezoidalSymmetric}}
}
