% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfS_Triangular.R
\name{cfS_Triangular}
\alias{cfS_Triangular}
\title{Characteristic function of the zero-mean symmetric TRIANGULAR distribution}
\usage{
cfS_Triangular(t, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}
}
\value{
Characteristic function \eqn{cf(t)} of the zero-mean symmetric TRIANGULAR distribution.
}
\description{
\code{cfS_Triangular(t, coef, niid)} evaluates the characteristic function \eqn{cf(t)}
of the zero-mean symmetric TRIANGULAR distribution defined on the interval \eqn{(-1,1)}.

\code{cfS_Triangular} is an ALIAS of the more general function \code{cf_TriangularSymmetric},
used to evaluate the characteristic function of a linear combination
of independent TRIANGULAR distributed random variables.

The characteristic function of \eqn{X ~ TriangularSymmetric} is defined by
\deqn{cf(t) = (2-2*cos(t))/t^2.}
}
\examples{
## EXAMPLE 1
# CF of the symmetric Triangular distribution on (-1,1)
t <- seq(from = -50,
         to = 50,
         length.out = 501)
plotReIm(function(t)
        cfS_Triangular(t), t, title = "CF of the the symmetric Triangular distribution on (-1,1)")

## EXAMPLE 2
# PDF/CDF of the the symmetric Triangular distribution on (-1,1)
cf <- function(t)
        cfS_Triangular(t)
x <- seq(from = -1,
         to = 1,
         length.out = 101)
xRange <- 2
options <- list()
options$dt <- 2 * pi / xRange
result <- cf2DistGP(cf = cf, x = x, options = options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Triangular_distribution}.

Other Continuous Probability Distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}},
  \code{\link{cfX_ChiSquare}},
  \code{\link{cfX_Exponential}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_PDF}}, \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaNC}}, \code{\link{cf_BetaSymmetric}},
  \code{\link{cf_Beta}}, \code{\link{cf_ChiSquare}},
  \code{\link{cf_Exponential}},
  \code{\link{cf_FisherSnedecorNC}},
  \code{\link{cf_FisherSnedecor}}, \code{\link{cf_Gamma}},
  \code{\link{cf_InverseGamma}},
  \code{\link{cf_LogRV_BetaNC}},
  \code{\link{cf_LogRV_Beta}},
  \code{\link{cf_LogRV_ChiSquareNC}},
  \code{\link{cf_LogRV_ChiSquare}},
  \code{\link{cf_LogRV_FisherSnedecorNC}},
  \code{\link{cf_LogRV_FisherSnedecor}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}},
  \code{\link{cf_vonMises}}

Other Symmetric Probability Distribution

WITKOVSKY V. (2016). Numerical inversion of a characteristic function:
An alternative tool to form the probability distribution
of output quantity in linear measurement models. Acta IMEKO, 5(3), 32-44.: \code{\link{cf_TriangularSymmetric}}
}
