% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_Exponential.R
\name{cfX_Exponential}
\alias{cfX_Exponential}
\title{Characteristic function of the EXPONENTIAL distribution}
\usage{
cfX_Exponential(t, lambda, coef, niid)
}
\arguments{
\item{t}{vector or array of real values, where the CF is evaluated.}

\item{lambda}{vector of the 'rate' parameters \code{lambda > 0}. If empty, default value is \code{lambda = 1}.}
}
\value{
Characteristic function \eqn{cf(t)} of the EXPONENTIAL distribution.
}
\description{
\code{cfX_Exponential(t, lambda)} evaluates characteristic function of the EXPONENTIAL distribution
with the rate parameter \code{lambda > 0}.

\code{cfX_Exponential} is an ALIAS NAME of the more general function \code{cf_Exponential},
used to evaluate the characteristic function of a linear combination

of independent EXPONENTIAL distributed random variables.
The characteristic function of the EXPONENTIAL distribution is
\deqn{cf(t) = \lambda / (\lambda - 1i*t).}
}
\examples{
## EXAMPLE 1
# CF of the Exponential distribution with lambda = 5
lambda <- 5
t <- seq(from = -50,
         to = 50,
         length.out = 501)
plotGraf(function(t)
        cfX_Exponential(t, lambda), t, title = "CF of the Exponential distribution with lambda = 5")

## EXAMPLE 2
# PDF/CDF of the Exponential distribution with lambda = 5
lambda <- 5
cf <- function(t)
        cfX_Exponential(t, lambda)
x  <- seq(from = 0,
          to = 1.5,
          length.out = 101)
options <- list()
options$xMin <- 0
options$SixSigmaRule <- 8
result <- cf2DistGP(cf = cf, x = x, options = options)

## EXAMPLE 3
# PDF/CDF of the compound Binomial-Exponential distribution
n <- 25
p <- 0.3
lambda <- 5
cfX <- function(t)
        cfX_Exponential(t, lambda)
cf <- function(t)
        cfN_Binomial(t, n, p, cfX)
x <- seq(from = 0,
         to = 5,
         length.out = 101)
prob <- c(0.9, 0.95, 0.99)
options <- list()
options$isCompound <- TRUE
result <- cf2DistGP(cf, x, prob, options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Exponential_distribution}.

Other Continuous Probability distribution: \code{\link{cfS_Arcsine}},
  \code{\link{cfS_Beta}}, \code{\link{cfS_Gaussian}},
  \code{\link{cfS_Rectangular}}, \code{\link{cfS_Student}},
  \code{\link{cfS_Trapezoidal}}, \code{\link{cfX_Beta}},
  \code{\link{cfX_ChiSquare}}, \code{\link{cfX_Gamma}},
  \code{\link{cfX_InverseGamma}},
  \code{\link{cfX_LogNormal}}, \code{\link{cfX_Normal}},
  \code{\link{cfX_Rectangular}},
  \code{\link{cfX_Triangular}},
  \code{\link{cf_ArcsineSymmetric}},
  \code{\link{cf_BetaSymmetric}}, \code{\link{cf_Beta}},
  \code{\link{cf_ChiSquare}}, \code{\link{cf_Exponential}},
  \code{\link{cf_Gamma}}, \code{\link{cf_InverseGamma}},
  \code{\link{cf_Normal}},
  \code{\link{cf_RectangularSymmetric}},
  \code{\link{cf_Student}},
  \code{\link{cf_TrapezoidalSymmetric}},
  \code{\link{cf_TriangularSymmetric}}
}
