% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfX_PearsonV.R
\name{cfX_PearsonV}
\alias{cfX_PearsonV}
\title{Characteristic function of Pearson type V distribution}
\usage{
cfX_PearsonV(t, alpha = 1, beta = 1)
}
\arguments{
\item{t}{vector of real values where the CF is evaluated, i.e. \eqn{CF(t)}.}

\item{alpha}{scalar shape parameter of the Pearson VI distribution, \code{alpha > 0}.}

\item{beta}{scalar scale parameter of the Pearson VI distribution, \code{beta > 0}.}
}
\value{
Characteristic function \eqn{cf(t)} of the Pearson type V distribution.
}
\description{
\code{cfX_PearsonV(t, alpha, beta)} evaluates the characteristic function
of the Pearson type V distribution with the parameters \eqn{alpha} (shape, \code{alpha > 0}) and
\eqn{beta} (scale, \code{beta > 0}), computed for real vector argument \code{t}, i.e.
\deqn{cfX_PearsonV(t, \alpha, \beta) = (2/gamma(\alpha)) * (-1i*t/\beta)^(\alpha/2) * besselk(\alpha,2*sqrt(-1i*t/\beta)),}
where \eqn{besselk(a,z)} denotes the modified Bessel function of the second order.
}
\note{
Ver.: 16-Sep-2018 19:28:37 (consistent with Matlab CharFunTool v1.3.0, 15-Nov-2016 13:36:26).
}
\examples{
## EXAMPLE 1
# CF of the PearsonV distribution
alpha <- 3 / 2
beta <- 2 / 3
t <- seq(-10, 10, length.out = 1001)
plotReIm(function(t)
        cfX_PearsonV(t, alpha, beta), t,
        title = "CF of the PearsonV distribution with alpha = 3/2, beta = 2/3")

## EXAMPLE 2
# PDF/CDF of the Beta distribution with alpha = 3/2, beta = 2/3
alpha <- 3 / 2
beta <- 2 / 3
prob <- c(0.9, 0.95, 0.99)
x <- seq(0, 40, length.out = 101)
cf <- function(t)
        cfX_PearsonV(t, alpha, beta)
options <- list()
options$xMin <- 0
options$N <- 2 ^ 10
options$SixSigmaRule <- 10
result <- cf2DistGP(cf, x, prob, options)

## EXAMPLE 3
# PDF/CDF of the compound Binomial-PearsonV distribution
n <- 25
p <- 0.3
alpha <- 3 / 2
beta <- 2 / 3
prob <- c(0.9, 0.95, 0.99)
x <- seq(0, 200, length.out = 101)
cfX <- function(t)
        cfX_PearsonV(t, alpha, beta)
cf <- function(t)
        cfN_Binomial(t, n, p, cfX)
options <- list()
options$isCompound <- TRUE
options$N <- 2 ^ 10
result <- cf2DistGP(cf, x, prob, options)
}
\seealso{
For more details see WIKIPEDIA:
\url{https://en.wikipedia.org/wiki/Pearson_distribution}.
}
\concept{Continuous Probability Distribution

 @references
[1] WITKOVSKY, V.: On the exact computation of the density and
of the quantiles of linear combinations of t and F random variables.
Journal of Statistical Planning and Inference 94 (2001), 1-13.

[2] WITKOVSKY V. (2016). Numerical inversion of a characteristic function:
An alternative tool to form the probability distribution of output quantity
in linear measurement models. Acta IMEKO, 5(3), 32-44.

[3] WITKOVSKY V., WIMMER G., DUBY T. (2016). Computing the aggregate loss distribution
based on numerical inversion of the compound empirical characteristic function
of frequency and severity. Working Paper. Insurance: Mathematics and Economics.

[4] DUBY T., WIMMER G., WITKOVSKY V.(2016). MATLAB toolbox CRM for computing distributions
of collective risk models.  Working Paper. Journal of Statistical Software.}
