% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaLog.R
\name{GammaLog}
\alias{GammaLog}
\title{Natural Log of the Gamma function}
\usage{
GammaLog(z)
}
\arguments{
\item{z}{may be complex and of any size.}
}
\value{
Natural Log of the Gamma function.
}
\description{
\code{GammaLog(z)} computes  natural Log of the Gamma function valid in the entire complex
plane. This routine uses an excellent Lanczos series
approximation for the complex \eqn{ln(Gamma)} function.
}
\references{
C. Lanczos, SIAM JNA  1, 1964. pp. 86-96. \cr
Y. Luke, "The Special ... approximations", 1969 pp. 29-31. \cr
Y. Luke, "Algorithms ... functions", 1977. \cr
J. Spouge,  SIAM JNA 31, 1994. pp. 931. \cr
W. Press,  "Numerical Recipes". \cr
S. Chang, "Computation of special functions", 1996.
}
\seealso{
Other Special Function: \code{\link{ChebPoints}},
  \code{\link{hypergeom1F1}},
  \code{\link{interpBarycentric}}
}
