% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReIm.R
\name{plotReIm}
\alias{plotReIm}
\title{Plot real and imaginary part of f(x)}
\usage{
plotReIm(
  f,
  x,
  title,
  xmin,
  xmax,
  dx = 0.1,
  xlab,
  ylab,
  px = 0,
  py = 0,
  labelx = TRUE,
  labely = TRUE
)
}
\arguments{
\item{f}{function.}

\item{x}{numerical values (number or vector).}

\item{title}{an overall title for the plot.}

\item{xmin}{number, default value \code{xmin = -1}.}

\item{xmax}{number, default value \code{xmax = 1}.}

\item{dx}{positive number, default value \code{dx = 0.1}.}

\item{xlab}{a title for the \eqn{x} axis: see \link[graphics]{title}.}

\item{ylab}{a title for the \eqn{y} axis: see \link[graphics]{title}.}

\item{px}{numerical values (number or vector) in which are auxiliary lines on axis \eqn{x} rendered, default value \code{px = 0}.}

\item{py}{numerical values (number or vector) in which are auxiliary lines on axis \eqn{y} rendered, default value \code{py = 0}.}

\item{labelx}{rendering auxiliary line, default value \code{TRUE}.}

\item{labely}{rendering auxiliary line, default value \code{TRUE}.}
}
\value{
Graf of function \code{f} in argument \code{x} or in range \code{[xmin, xmax]} with step \code{dx}.
}
\description{
\code{plotReIm(f, x, xmin, xmax)} plots values of complex function \code{f} in real arguments \code{x} or in range \code{[xmin, xmax]} with step \code{dx}.
}
\details{
Function with optional parameters must be pass as lambda function \code{f = function(x) f(x, optional parameters)}.
}
\note{
Ver.: 22-Sep-2018 20:32:20.
}
\seealso{
Other Graphic Function: 
\code{\link{plotPolar}()},
\code{\link{plotReIm2}()}
}
\concept{Graphic Function}
